% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-data.R
\name{add_cols}
\alias{add_cols}
\title{Add columns to a worksheet}
\usage{
add_cols(ws, n)
}
\arguments{
\item{ws}{worksheet object}

\item{n}{\code{numeric} the number of columns to add}
}
\description{
Add columns to a worksheet
}

