% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permissions.R
\name{add_perm}
\alias{add_perm}
\title{Add a permission to a spreadsheet}
\usage{
add_perm(ss, name = NULL, type = NULL, role = NULL, with_link = TRUE,
  verbose = TRUE)
}
\arguments{
\item{name}{The email address or domain name for the entity.}

\item{type}{The account type. Allowed values "user", "group", "domain", or
"anyone".}

\item{role}{The primary role for this user. Allowed values are: "owner",
"writer", or "reader".}

\item{with_link}{boolean, whether the link is required for this permission}

\item{verbose}{logical, do you want informative messages?}
}
\value{
Information about newly added permission
}
\description{
An email will be sent automatically to the subject to notify them of the
permission.
}
\examples{
\dontrun{
foo <- new_ss("foo")

# Add someone as a writer/reader:
add_perm(foo, value = "someone@gmail.com", type = "user", role = "writer")
add_perm(foo, value = "someone@gmail.com", type = "user", role = "reader")

# Add anyone as a writer/reader:
add_perm(foo, value = NULL, type = "anyone", role = "writer")
add_perm(foo, value = NULL, type = "anyone", role = "reader")

# Add a domain as reader:
add_perm(foo, value = "hotmail.com", type = "domain", role = "reader")

# Add a group as reader/writer:
add_perm(foo, value = "some_cool_group@googlegroups.com", type = "group", role = "reader")
add_perm(foo, value = "some_cool_group@googlegroups.com", type = "group", role = "writer")

}
}

