% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{add_rows}
\alias{add_rows}
\title{Add rows to a worksheet}
\usage{
add_rows(ws, n)
}
\arguments{
\item{ws}{worksheet object}

\item{n}{\code{numeric} the number of rows to add}
}
\description{
Add rows to a worksheet
}

