% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_worksheet}
\alias{add_worksheet}
\title{Add a new (empty) worksheet to spreadsheet}
\usage{
add_worksheet(ss, title, nrow, ncol)
}
\arguments{
\item{ss}{spreadsheet object}

\item{title}{character string for title of new worksheet}

\item{nrow}{number of rows}

\item{ncol}{number of columns}
}
\description{
Add a new (empty) worksheet to spreadsheet, specify title, number of rows
and columns. The title of the new worksheet should not be the same as any of
the existing worksheets
}

