% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{add_ws}
\alias{add_ws}
\title{Add a new (empty) worksheet to spreadsheet}
\usage{
add_ws(ss, ws_title = "Sheet1", nrow = 1000, ncol = 26, verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google sheet}

\item{ws_title}{character string for title of new worksheet}

\item{nrow}{number of rows (default is 1000)}

\item{ncol}{number of columns (default is 26)}

\item{verbose}{logical; do you want informative message?}
}
\value{
a spreadsheet object, resulting from re-registering the host
  spreadsheet after adding the new worksheet
}
\description{
Add a new (empty) worksheet to spreadsheet: specify title and worksheet
extent (number of rows and columns). The title of the new worksheet can not
be the same as any existing worksheets in the sheet.
}

