% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cell-specification.R
\name{as.cell_limits.list}
\alias{as.cell_limits.list}
\title{Convert a list into a cell_limits object}
\usage{
\method{as.cell_limits}{list}(x)
}
\arguments{
\item{x}{list of length two}
}
\value{
a \code{cell_limits} object, which is a list with components named
  \code{rows} and \code{cols}, each holding a minimum and maximum in an
  integer vector of length two
}
\description{
Convert a list into a cell_limits object
}
\examples{
\dontrun{
as.cell_limits(list(rows = c(1, 5), cols = c(1, 4)))
as.cell_limits(list(c(1, 5), c(1, 4)))
}
}
\keyword{internal}

