% Generated by roxygen2 (4.0.2): do not edit by hand
\name{authorize}
\alias{authorize}
\title{Authorize client using Oauth2.0 Credentials}
\usage{
authorize()
}
\value{
Object of class client which stores the Oauth2.0 token required for subsequent requests.

This method is following the demo described at:
\url{https://github.com/hadley/httr/blob/master/demo/oauth2-google.r}

Authorization token will be stored in client as auth.
}
\description{
Authorize user using email and password.
}

