% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/http-requests.R
\name{build_query}
\alias{build_query}
\title{Build query string for GET URL}
\usage{
build_query(min_row, max_row, min_col, max_col)
}
\arguments{
\item{min_row,max_row,min_col,max_col}{query parameters}
}
\description{
Form the query string to append to GET URL.
}

