% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{build_range}
\alias{build_range}
\title{Determine the range occupied by a vector/data frame}
\usage{
build_range(limits, input)
}
\arguments{
\item{limits}{from convert_range_limit_list}

\item{input}{either a vector of a data frame}
}
\value{
character string for the range taken up by a vector or data frame
}
\description{
Determine the range occupied by a vector/data frame
}

