% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_missing}
\alias{check_missing}
\title{Check for missing cells in row or column}
\usage{
check_missing(x, by_col = TRUE)
}
\arguments{
\item{x}{data frame returned from \code{\link{create_lookup_tbl}}}
}
\value{
A vector of values contained in row with NAs for missing values.
}
\description{
Loop through cell indices and substitute NA for missing cell values.
}

