% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/auth.R
\name{check_token}
\alias{check_token}
\title{Check Google token for validity}
\usage{
check_token(token)
}
\arguments{
\item{token}{Google authorization token}
}
\description{
Make sure Google token is good to use upon retrieval so error is found
before making requests.
}

