% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_url_from_key}
\alias{construct_url_from_key}
\title{Construct a browser URL from a key}
\usage{
construct_url_from_key(key)
}
\arguments{
\item{key}{character, unique key for a spreadsheet}
}
\description{
Construct a browser URL from a key
}
\keyword{internal}
