% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_ws_feed_from_key}
\alias{construct_ws_feed_from_key}
\title{Construct a worksheets feed from a key}
\usage{
construct_ws_feed_from_key(key, visibility = "private")
}
\arguments{
\item{key}{character, unique key for a spreadsheet}

\item{visibility}{character, either "private" (default) or "public",
  indicating whether further requests will be made with or without
  authentication, respectively}
}
\description{
Construct a worksheets feed from a key
}
\keyword{internal}

