% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_limit_list_to_range}
\alias{convert_limit_list_to_range}
\title{Convert a limits list to a cell range}
\usage{
convert_limit_list_to_range(limits, pn = c("R1C1", "A1"))
}
\arguments{
\item{limits}{limits list}

\item{pn}{positioning notation}
}
\description{
Convert a limits list to a cell range
}

