% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_range_to_limit_list}
\alias{convert_range_to_limit_list}
\title{Convert a cell range into a limits list}
\usage{
convert_range_to_limit_list(range)
}
\arguments{
\item{range}{character vector, length one, such as "A1:D7"}
}
\description{
Convert a cell range into a limits list
}
\keyword{internal}

