% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{coord_to_label}
\alias{coord_to_label}
\title{Convert R1C1 positioning notation to A1 notation}
\usage{
coord_to_label(x)
}
\arguments{
\item{x}{cell position in R1C1 notation}
}
\description{
Convert R1C1 positioning notation to A1 notation
}
\keyword{internal}

