% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{copy_sheet}
\alias{copy_sheet}
\title{Make a copy of an existing spreadsheet}
\usage{
copy_sheet(from, key = NULL, to = NULL, verbose = TRUE)
}
\arguments{
\item{from}{character string giving identifying information for the
sheet: title, key, URL}

\item{key}{character string guaranteed to provide unique key of the
sheet; overrides \code{from}}

\item{to}{character string giving the new title of the sheet; if
\code{NULL}, then the copied sheet will be titled "Copy of ..."}

\item{verbose}{logical; do you want informative message?}
}
\description{
You can copy a spreadsheet that you own or a sheet owned by a third party
that has been made accessible via the sharing dialog options. If the sheet
you want to copy is visible in the listing provided by
\code{\link{list_sheets}}, you can specify it by title. Otherwise you can
extract the key from the browser URL via \code{\link{extract_key_from_url}}
and explicitly specify the sheet by key.
}
\note{
if two sheets with the same name exist in your Google drive then
  sheet with the most recent "last updated" timestamp will be copied.
}

