% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit.R
\name{copy_spreadsheet}
\alias{copy_spreadsheet}
\title{Copy an existing spreadsheet}
\usage{
copy_spreadsheet(title_or_key, new_title = NULL)
}
\arguments{
\item{title_or_key}{either the title of a spreadsheet or the key of a
spreadsheet}

\item{new_title}{if \code{new_title} is NULL then the copied spreadsheet will
be given the default name: "Copy of ..."}
}
\description{
Make a copy of an existing spreadsheet into your Google Drive. You can either
make a copy of your own spreadsheet or another user's spreadsheet. If the
spreadsheet you want to make a copy of already exists in your Google Drive,
pass in the title of the spreadsheet. To get a copy of another user's
spreadsheet, enter the key of that spreadsheet. Make sure that the target
spreadsheet is made 'accessible' in the link sharing options or else it wont
be found.
}

