% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{copy_ss}
\alias{copy_ss}
\title{Make a copy of an existing spreadsheet}
\usage{
copy_ss(x, new_title = NULL)
}
\arguments{
\item{x}{the title or key or url of a spreadsheet}

\item{new_title}{a character string for the new title of the spreadsheet,
if \code{new_title} is NULL then the copied spreadsheet will
be given the default name: "Copy of ..."}
}
\description{
You can either make a copy of your own spreadsheet or another user's
spreadsheet. If the spreadsheet you want to make a copy of already exists in
your Google Drive, pass in the title of the spreadsheet. To get a copy of
another user's spreadsheet, enter the key of that spreadsheet. Make sure
that the target spreadsheet is made 'accessible' in the sharing dialog
options or else it wont be found.
}
\note{
if two spreadsheets with the same name exist in your Google drive then
spreadsheet with the most recent "last updated" timestamp will be copied.
}

