% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_lookup_tbl}
\alias{create_lookup_tbl}
\title{#Find number of rows and columns of worksheet
#
#Get the rows and columns of worksheet by making a request for cellfeed
#
#dparam ws Worksheet object
#param auth Google token
#param visibility set to \code{public} for public sheets
#
#The is faster but with a trade-off. For listfeeds, data will stop before an
#entire blank row. For example if row 2 is entirely blank, no data will be
#returned. Blank columns are also ignored. The number of columns will be the
#maximum number of cells in a row that contain a value, regardless
#of spacing.
#
#importFrom XML getNodeSet
#importFrom XML xmlChildren
Create lookup table for cell indices and its values}
\usage{
create_lookup_tbl(cellsfeed)
}
\arguments{
\item{cellsfeed}{Cell based feed parsed with \code{\link{gsheets_parse}}}
}
\value{
A data frame with cell row and col number and corresponding input value.
}
\description{
Convert cell row and col attributes stored in entry nodes, along with
input values into a single data frame and serve as a lookup table for
further processing.
}

