% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_update_feed}
\alias{create_update_feed}
\title{Generate a cells feed to update cell values}
\usage{
create_update_feed(feed, new_values)
}
\arguments{
\item{feed}{cell feed returned and parsed from GET request}

\item{new_values}{vector of new values to update cells}
}
\description{
Create an update feed for the new values.
}

