% Generated by roxygen2 (4.0.2): do not edit by hand
\name{del_spreadsheet}
\alias{del_spreadsheet}
\title{Move spreadsheet to trash}
\usage{
del_spreadsheet(title)
}
\arguments{
\item{title}{a character string for the title of spreadsheet}
}
\description{
Move a spreadsheet to trash in Google Drive.
}

