% Generated by roxygen2 (4.0.2): do not edit by hand
\name{del_worksheet}
\alias{del_worksheet}
\title{Delete worksheet from spreadsheet}
\usage{
del_worksheet(ss, ws)
}
\arguments{
\item{ss}{Spreadsheet object}

\item{ws}{Worksheet object}
}
\description{
The worksheet and all of its data will be removed from spreadsheet.
}

