% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permissions.R
\name{delete_perm}
\alias{delete_perm}
\title{Delete a permission from a spreadsheet}
\usage{
delete_perm(ss, value = NULL, perm_id = NULL, verbose = TRUE)
}
\arguments{
\item{value}{The email address or domain name for the entity.}

\item{perm_id}{The permission ID}

\item{verbose}{logical; do you want informative messages?}
}
\description{
Delete a permission from a spreadsheet
}

