% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{delete_sheet}
\alias{delete_sheet}
\title{Move a spreadsheet to trash}
\usage{
delete_sheet(x, verbose = TRUE)
}
\arguments{
\item{x}{character string giving identifying information for the sheet:
title, key, URL}

\item{verbose}{logical; do you want informative message?}
}
\description{
Move a spreadsheet to trash in Google Drive.
}
\note{
Use the key when there are multiple sheets with the same name, since
  the default will just send the most recent sheet to the trash.

Shared sheets can not be removed from your Google Drive with this
  function. You must remove it manually in the web browser.
}

