% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{delete_ss}
\alias{delete_ss}
\title{Move a spreadsheet to trash in Google Drive.}
\usage{
delete_ss(x)
}
\arguments{
\item{x}{the title or key of a spreadsheet}
}
\description{
You must be the owner of the spreadsheet in order to move it to the trash.
If you try to delete a spreadsheet that doesnt belong to you, a 403 Forbidden
HTTP status code will be returned; such shared spreadsheets can only be moved
to the trash manually in the web browser. If you trash a spreadsheet that is
shared with others, it will no longer appear in any of their Google Drives.
}
\note{
Using the key is useful when there are spreadsheets with the same name
since the default is to send the most recent spreadsheet to the trash.
}

