% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/download-spreadsheets.R
\name{download_ss}
\alias{download_ss}
\title{Download a Google Sheet}
\usage{
download_ss(from, key = NULL, ws = NULL, to = "my_sheet.xlsx",
  overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{from}{sheet-identifying information, either a spreadsheet object or a
  character vector of length one, giving a URL, sheet title, key or
  worksheets feed}

\item{key}{character string guaranteed to provide unique key of the sheet;
  overrides \code{from}}

\item{ws}{positive integer or character string specifying index or title,
  respectively, of the worksheet to export ; if \code{NULL} then the entire
  spreadsheet will be exported}

\item{to}{path to write file, if it does not contain the absolute path,
then the file is relative to the current working directory}

\item{overwrite}{will only overwrite existing path if TRUE}

\item{verbose}{logical; do you want informative message?}
}
\description{
Export a Google Spreadsheet as a .csv, .pdf, or .xlsx file. You can download
a spreadsheet that you own or a sheet owned by a third party that has been
made accessible via the sharing dialog options. You can download an entire
spreadsheet or a single worksheet from a spreadsheet if you provide worksheet
identifying information. If the chosen format is csv, the first worksheet
will be exported, unless another worksheet is specified. If pdf format is
chosen, all sheets will be appended into 1 pdf document.
}

