% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_drive_user.R
\name{drive_user}
\alias{drive_user}
\title{Retrieve Google Drive user data}
\usage{
drive_user()
}
\value{
an object of S3 class `drive_user`, which is just a list
}
\description{
Unexported workhorse function.
}
\keyword{internal}

