% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_key_from_url}
\alias{extract_key_from_url}
\title{Extract sheet key from its browser URL}
\usage{
extract_key_from_url(url)
}
\arguments{
\item{url}{URL seen in the browser when visiting the sheet}
}
\description{
Extract sheet key from its browser URL
}

