% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fetch_ws}
\alias{fetch_ws}
\title{Retrieve worksheet object from worksheets feed}
\usage{
fetch_ws(node, sheet_id)
}
\description{
Store worksheet info (spreadsheet id, worksheet id, worksheet title,
listfeed and cellfeed uris) from entry nodes
}

