% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fill_missing_row}
\alias{fill_missing_row}
\title{Fill in missing rows in column}
\usage{
fill_missing_row(x)
}
\arguments{
\item{x}{data frame returned by \code{\link{create_look_tbl}}

This function operates on the lookup table grouped by column.}
}
\description{
The lookup table returned by create_lookup_tbl may contain missing tuples
for empty cells. This function fills in the table so that there is a tuple
for every row down to the bottom-most row of the column.
}

