% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{find_all}
\alias{find_all}
\title{Find all cells with string value}
\usage{
find_all(ws, x)
}
\arguments{
\item{ws}{worksheet object}

\item{x}{a character string}
}
\value{
a data frame listing the cell locations in label and coordinate
format
}
\description{
Get all the cell locations for a string value
}

