% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{find_cell}
\alias{find_cell}
\title{Find first cell matching string value}
\usage{
find_cell(ws, x)
}
\arguments{
\item{ws}{worksheet object}

\item{x}{a character string (case sensitive)}
}
\description{
Get the cell location of the first occurence of a cell value.
}

