% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_token-print.R
\name{gd_token}
\alias{gd_token}
\alias{gs_token}
\title{Retrieve and report on the current token}
\usage{
gd_token(verbose = TRUE)

gs_token(verbose = TRUE)
}
\arguments{
\item{verbose}{logical; do you want informative messages?}
}
\value{
an OAuth token object, specifically a
  \code{\link[=Token-class]{Token2.0}}, invisibly
}
\description{
Prints information about the Google token that is in force and returns the
token invisibly.
}
\examples{
\dontrun{
## load/refresh existing credentials, if available
## otherwise, go to browser for authentication and authorization
gs_auth()

gd_token()
}
}
