% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_requests.R
\name{gdrive_POST}
\alias{gdrive_POST}
\title{Make POST request to Google Drive API}
\usage{
gdrive_POST(url, ...)
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{...}{optional; further named parameters, such as \code{query},
\code{path}, etc, passed on to \code{\link[httr]{modify_url}}. Unnamed
parameters will be combined with \code{\link[httr]{config}}.}
}
\description{
Used in gs_new(), gs_delete(), gs_copy()
}
\keyword{internal}

