% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http_requests.R
\name{gdrive_PUT}
\alias{gdrive_PUT}
\title{Make PUT request to Google Drive API}
\usage{
gdrive_PUT(url, ...)
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{...}{optional; further named parameters, such as \code{query},
  \code{path}, etc, passed on to \code{\link[httr]{modify_url}}. Unnamed
  parameters will be combined with \code{\link[httr]{config}}.}
}
\description{
Used in gs_upload()
}
\keyword{internal}

