% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{get_cell}
\alias{get_cell}
\title{Get the value of a cell}
\usage{
get_cell(ws, cell)
}
\arguments{
\item{ws}{worksheet object}

\item{cell}{character string specifying cell position,
either in label or coordinate notation}
}
\value{
The value of the cell as a character string.
}
\description{
Get the value of a cell using label (A1) notation or coordinate (R1C1)
notation.
}

