% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/consume-data.R
\name{get_cells}
\alias{get_cells}
\title{Get data from a cell or range of cells}
\usage{
get_cells(ss, ws = 1, range)
}
\arguments{
\item{ss}{a registered Google spreadsheet}

\item{ws}{positive integer or character string specifying index or title,
  respectively, of the worksheet to consume}

\item{range}{single character string specifying which cell or range of cells
  to retrieve; positioning notation can be either "A1" or "R1C1"; a single
  cell can be requested, e.g. "B4" or "R4C2" or a rectangular range can be
  requested, e.g. "B2:D4" or "R2C2:R4C4"}
}
\description{
Get data via the cell feed for a rectangular range of cells
}
\examples{
\dontrun{
gap_key <- "1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA"
gap_ss <- copy_ss(key = gap_key, to = "gap_copy")
get_cells(gap_ss, "Europe", range = "B3:D7")
simplify_cf(get_cells(gap_ss, "Europe", range = "A1:F1"))
}
}
\seealso{
\code{\link{reshape_cf}} to reshape the retrieved data into a more
  usable data.frame

Other data.consumption.functions: \code{\link{get_col}};
  \code{\link{get_row}}; \code{\link{get_via_cf}};
  \code{\link{get_via_csv}}; \code{\link{get_via_lf}};
  \code{\link{reshape_cf}}; \code{\link{simplify_cf}}
}

