% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/consume-data.R
\name{get_col}
\alias{get_col}
\title{Get all values in a column.}
\usage{
get_col(ws, col)
}
\arguments{
\item{ws}{worksheet object}

\item{col}{column number or letter (case insensitive)}
}
\value{
A data frame.
}
\description{
Get all values in a column.
}
\seealso{
\code{\link{get_cols}}, \code{\link{get_row}},
\code{\link{get_rows}}, \code{\link{read_all}}, \code{\link{read_region}},
\code{\link{read_range}}
}

