% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_dataframe}
\alias{get_dataframe}
\title{Get worksheet object as a dataframe}
\usage{
get_dataframe(ws, client = NULL)
}
\arguments{
\item{ws}{Worksheet object}

\item{client}{Client object}
}
\value{
A dataframe.
}
\description{
Use worksheet object and turn it into a dataframe.
}

