% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_lookup_tbl}
\alias{get_lookup_tbl}
\title{Get lookup table}
\usage{
get_lookup_tbl(feed, include_sheet_title = FALSE)
}
\arguments{
\item{feed}{cellfeed from GET request}

\item{include_sheet_title}{include column with worksheet title in lookup
table}
}
\description{
Create lookup table from cellfeed.
}

