% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permissions.R
\name{get_perm}
\alias{get_perm}
\title{Retrieve a permission from a spreadsheet}
\usage{
get_perm(ss, info, verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google Spreadsheet}

\item{verbose}{logical; do you want informative messages?}

\item{id}{identifying info, should be either email or permission ID}
}
\description{
Retrieve a permission from a spreadsheet
}
\keyword{internal}

