% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_region}
\alias{get_region}
\title{Get a region of a worksheet}
\usage{
get_region(ws, from_row, to_row, from_col, to_col, vis = "private")
}
\arguments{
\item{ws}{Worksheet object}

\item{from_row,to_row}{range of rows to extract}

\item{from_col,to_col}{range of cols to extract}

\item{vis}{Either \code{private} or \code{public}}
}
\value{
A dataframe.
}
\description{
Extract cells of a worksheet according to specified range.
}

