% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_rows}
\alias{get_rows}
\title{Get rows of worksheet}
\usage{
get_rows(ws, from, to, vis = "private")
}
\arguments{
\item{ws}{worksheet object returned by \code{\link{open_worksheet}}}

\item{from,to}{start and end row indexes}
}
\value{
Dataframe of requested rows.
}
\description{
Specify range of rows to get from worksheet.
}

