% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{get_user_info}
\alias{get_user_info}
\title{Retrieve User data}
\usage{
get_user_info()
}
\description{
Store user name, email address and datetime of request in environment
variable .state$user
}
\keyword{internal}

