% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/consume-data.R
\name{get_via_cf}
\alias{get_via_cf}
\title{Create a data.frame of the non-empty cells in a rectangular region of a
worksheet}
\usage{
get_via_cf(ss, ws = 1, min_row = NULL, max_row = NULL, min_col = NULL,
  max_col = NULL)
}
\arguments{
\item{ss}{a registered Google spreadsheet}

\item{ws}{positive integer or character specifying index or title,
respectively, of the worksheet to consume; defaults to 1, i.e. the first
worksheet}

\item{min_row}{positive integer, optional}

\item{max_row}{positive integer, optional}

\item{min_col}{positive integer, optional}

\item{max_col}{positive integer, optional}
}
\description{
No attempt to shape the returned data! Data.frame will have one row per cell.
}
\details{
Use the limits, e.g. min_row or max_col, to delineate the rectangular region
of interest. You can specify any subset of the limits or none at all. If
limits are provided, validity will be checked as well as internal consistency
and compliance with known extent of the worksheet. If no limits are provided,
all cells will be returned but user should realize that access via the list
feed is potentially a much faster to consume data from a rectangular
worksheet.

Empty cells, even if "embedded" in a rectangular region of populated cells,
are not returned by the API and will not appear in the returned data.frame.
}

