% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_worksheet}
\alias{get_worksheet}
\title{Open worksheet given worksheet title}
\usage{
get_worksheet(client = NULL, ss, title)
}
\arguments{
\item{client}{Client object}

\item{ss}{Spreadsheet object containing worksheet}

\item{title}{title of worksheet to retrieve}
}
\value{
An object of class worksheet and number of rows and cols attribute.
}
\description{
Use title of worksheet to retrieve object of class worksheet.
}

