% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ws}
\alias{get_ws}
\title{Retrieve a worksheet from a spreadsheet}
\usage{
get_ws(ss, ws)
}
\arguments{
\item{ss}{a registered spreadsheet}

\item{ws}{a positive integer or character string specifying which worksheet}
}
\description{
Retrieve a worksheet from a spreadsheet based on either a positive integer
index or worksheet title.
}

