% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/register.R
\name{get_ws_feed}
\alias{get_ws_feed}
\title{Obtain the worksheets feed for a spreadsheet}
\usage{
get_ws_feed(x, visibility = "private")
}
\arguments{
\item{x}{character vector of length one, with spreadsheet-identifying
information}

\item{visibility}{either "public" or "private"}
}
\value{
The worksheets feed for the specified spreadsheet
}
\description{
Given a Google spreadsheet's URL, unique key, title, or worksheets feed,
return its worksheets feed. The worksheets feed is simply a URL -- different
from the one you see when visiting a spreadsheet in the browser! -- and is
the very best way to specify a spreadsheet for API access. There's no simple
way to capture a spreadsheet's worksheets feed, so this function helps you
convert readily available information (spreadsheet title or URL) into the
worksheets feed.
}
\details{
Simple regexes are used to detect if the input is a worksheets feed or the
URL one would see when visiting a spreadsheet in the browser. If it's a URL,
we attempt to extract the spreadsheet's unique key, assuming the URL follows
the pattern characteristic of "new style" Google spreadsheets.

Otherwise the input is assumed to be the spreadsheet's title or unique key.
When we say title, we mean the name of the spreadsheet in, say, Google Drive
or in the \code{sheet_title} variable of the data.frame returned by
\code{\link{list_spreadsheets}}. Spreadsheet title or key will be sought in
the listing of spreadsheets visible to the authenticated user and, if a match
is found, the associated worksheets feed is returned.
}

