% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets.R
\docType{package}
\name{googlesheets}
\alias{googlesheets}
\alias{googlesheets-package}
\title{\code{googlesheets} package}
\description{
Google spreadsheets R API
}
\details{
See the README on
\href{https://cran.r-project.org/web/packages/googlesheets/README.html}{CRAN}
or \href{https://github.com/jennybc/googlesheets}{GitHub}
}
