% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{gs_auth}
\alias{gs_auth}
\title{Authorize \code{googlesheets} to access user data from Google}
\usage{
gs_auth(new_user = FALSE)
}
\arguments{
\item{new_user}{logical, defaults to \code{FALSE}. Set to \code{TRUE} if you
  want to wipe the slate clean and re-authenticate with the same or different
  Google account.}
}
\description{
The user will be directed to a web browser, asked to sign in to their Google
account, and to grant \code{googlesheets} access to user data for Google
Spreadsheets and Google Drive. User credentials will be cached in a file
named \code{.httr-oauth} in the current working directory.
}
\details{
Based on
\href{https://github.com/hadley/httr/blob/master/demo/oauth2-google.r}{this
demo} from \code{\link[httr]{httr}}.
}

