% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{gs_auth_revoke}
\alias{gs_auth_revoke}
\title{Revoke authentication}
\usage{
gs_auth_revoke(rm_httr_oauth = FALSE, verbose = TRUE)
}
\description{
This unexported function exists so we can revoke all authentication for
testing purposes.
}
\keyword{internal}

