% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{gs_auth_suspend}
\alias{gs_auth_suspend}
\title{Suspend authorization}
\usage{
gs_auth_suspend(disable_httr_oauth = TRUE, verbose = TRUE)
}
\description{
This unexported function exists so we can suspend authorization for
testing purposes.
}
\keyword{internal}

