% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_autofill.R
\name{gs_autofill}
\alias{gs_autofill}
\title{Autofill Cells}
\usage{
gs_autofill(ss, ws = 1, range = "A1", alternate_series = FALSE,
  verbose = FALSE)
}
\arguments{
\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet}

\item{range}{a cell range, as described in \code{\link{cell-specification}}}

\item{alternate_series}{a logical indicating whether to fill the data using 
the format of an alternate series (i.e. alternate shading)}

\item{verbose}{logical; do you want informative messages?}
}
\description{
A function to fill in more data based on the source of existing data.
}
\examples{
\dontrun{

# if rows 1 thru 4 are blank, then autofill up based on row 5
# if rows 2 thru 5 are blank, then autofill down based on row 1
gs_autofill(gap_ss, ws = 1, range = cell_rows(1:5))

# autofill columns 6 & 7 based on column 5
gs_autofill(gap_ss, ws = 1, range = cell_cols(5:7))

}
}

