% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_browse.R
\name{gs_browse}
\alias{gs_browse}
\title{Visit a Google Sheet in the browser}
\usage{
gs_browse(ss, ws = 1)
}
\arguments{
\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet}
}
\value{
the \code{\link{googlesheet}} object given as input, invisibly
}
\description{
Visit a Google Sheet in the browser
}
\examples{
\dontrun{
gap_ss <- gs_gap()
gs_browse(gap_ss)
gs_browse(gap_ss, ws = 3)
gs_browse(gap_ss, ws = "Europe")

## assign and browse at once
gap_ss <- gs_gap() \%>\% gs_browse()
}
}
