% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_clear_cells.R
\name{gs_clear_cells}
\alias{gs_clear_cells}
\title{Clear Cells}
\usage{
gs_clear_cells(ss, ws = NULL, range = NULL, clear_type = c("values",
  "formats", "all"), verbose = FALSE)
}
\arguments{
\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet}

\item{range}{a cell range, as described in \code{\link{cell-specification}}}

\item{clear_type}{a character string that indicates what type of data to clear:
"values", "formats", or "all" data associated with the cell.}

\item{verbose}{logical; do you want informative messages?}
}
\description{
A function to clear the values of all cells in a range. The function allows 
for clearing worksheets or everything in the entire spreedsheet. Use with caution!
There is an option to clear formatting as well, but the default is to only 
clear the values of the cells.
}
\examples{
\dontrun{

gs_clear_cells(gap_ss, ws = 1, range = "A1:B2")

# clear cell values in A1 on every worksheet
gs_clear_cells(gap_ss, range = "A1")

# clear cell formatting in the first 2 rows
gs_clear_cells(gap_ss, ws = 1, range = cell_rows(1:2), clear_type = 'formats')

# you can clear all cells on a worksheet
gs_clear_cells(gap_ss, ws = 1)

# or even clear all cell values in the entire spreadsheet
gs_clear_cells(gap_ss, clear_type = 'values')

# or the nuclear option to clear values, formats, all cell data across all worksheets
gs_clear_cells(gap_ss, clear_type = 'all')

}
}

