% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{gs_copy}
\alias{gs_copy}
\title{Make a copy of an existing spreadsheet}
\usage{
gs_copy(from, to = NULL, verbose = TRUE)
}
\arguments{
\item{from}{a \code{\link{googlesheet}} object, i.e. a registered Google
sheet}

\item{to}{character string giving the new title of the sheet; if \code{NULL},
then the copy will be titled "Copy of ..."}

\item{verbose}{logical; do you want informative message?}
}
\description{
You can copy a spreadsheet that you own or a sheet owned by a third party
that has been made accessible via the sharing dialog options. If the sheet
you want to copy is visible in the listing provided by \code{\link{gs_ls}},
you can specify it by title (or any of the other spreadsheet-identifying
methods). Otherwise, you'll have to explicitly specify it by key.
}
\note{
If two sheets with the same name exist in your Google drive, then sheet
  with the most recent "last updated" timestamp will be copied. If you don't
  like that, then identify the target by something other than title.
}
\examples{
\dontrun{
gap_key <- "1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA"
gap_ss <- gs_copy(gs_key(gap_key), to = "Gapminder_copy")
gap_ss
}
}

