% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{gs_deauth}
\alias{gs_deauth}
\title{Suspend authorization}
\usage{
gs_deauth(clear_cache = TRUE, verbose = TRUE)
}
\arguments{
\item{clear_cache}{logical indicating whether to disable the
\code{.httr-oauth} file in working directory, if such exists, by renaming
to \code{.httr-oauth-SUSPENDED}}

\item{verbose}{logical; do you want informative messages?}
}
\description{
Suspend \code{\link{googlesheets}}' authorization to place requests to the
Drive and Sheets APIs on behalf of the authenticated user.
}
\examples{
\dontrun{
gs_deauth()
}
}
\seealso{
Other auth functions: \code{\link{gd_user}},
  \code{\link{gs_auth}}
}
\concept{auth functions}
