% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_dimensions.R
\name{gs_delete_range}
\alias{gs_delete_cells}
\alias{gs_delete_columns}
\alias{gs_delete_range}
\alias{gs_delete_rows}
\title{Delete Rows, Columns, or Cells}
\usage{
gs_delete_rows(ss, ws = 1, range = "A1", verbose = FALSE)

gs_delete_columns(ss, ws = 1, range = "A1", verbose = FALSE)

gs_delete_cells(ss, ws = 1, range = "A1", shift_direction = c("left",
  "up"), verbose = FALSE)
}
\arguments{
\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet}

\item{range}{a cell range, as described in \code{\link{cell-specification}}}

\item{verbose}{logical; do you want informative messages?}

\item{shift_direction}{character, of either "left" or "up" specifying
which direction to shift the existing cells when the range is deleted}
}
\description{
Delete rows or columns occupying a range. Also, delete cells and shift 
the remaining cells left or up to fill in the gap.
}
\examples{
\dontrun{

gs_delete_rows(gap_ss, ws = "Africa", range = "A2:F4")
gs_delete_rows(gap_ss, ws = "Africa", range = cell_rows(1:3))

gs_delete_columns(gap_ss, ws = "Africa", range = "A2:C4")
gs_delete_columns(gap_ss, ws = "Africa", range = cell_cols(1:2))

gs_delete_cells(gap_ss, ws = "Africa", range = "C3:E5")
gs_delete_cells(gap_ss, ws = "Africa", range = "B3:B5", shift_direction = 'up')

}
}

