% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alpha-inspect.R
\name{gs_inspect}
\alias{gs_inspect}
\title{Visual overview of populated cells}
\usage{
gs_inspect(x)
}
\arguments{
\item{x}{data.frame or tbl_df}
}
\value{
a ggplot object
}
\description{
\emph{This function is still experimental. Expect it to change!} This
function plots a data frame and is much like a visual representation of
\code{str} a data frame. You can inspect your data for populated and empty
cells. Empty cells (ie. \code{NA}'s) are represented by no colour fill. The
purpose is to get oriented to sheets that contain more than one data
rectangle.
}
\examples{
\dontrun{
gs_inspect(iris)

# data recorded from a game of ultimate frisbee
ulti_key <- "1223dpf3vnjZUYUnCM8rBSig3JlGrAu1Qu6VmPvdEn4M"
ulti_ss <- ulti_key \%>\% gs_key()
ulti_csv <- ulti_ss \%>\% get_col(ws = 2, col = 1:6)  \%>\% reshape_cf()
gs_inspect(ulti_csv)

}
}

