% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gs_ls.R
\name{gs_ls}
\alias{gs_ls}
\title{List available spreadsheets}
\usage{
gs_ls()
}
\value{
a \code{googlesheet_ls} object, which is a
  \code{\link[dplyr]{tbl_df}} with one row per sheet
}
\description{
Lists spreadsheets that the user would see in the Google Sheets home screen:
\url{https://docs.google.com/spreadsheets/}. This function returns the
information available from the
\href{https://developers.google.com/google-apps/spreadsheets/#retrieving_a_list_of_spreadsheets}{spreadsheets
feed} of the Google Sheets API. Since this is non-public user data, use of
\code{gs_ls} will require authentication.
}
\details{
This listing gives a \emph{partial} view of the sheets available for access
(why just partial? see below). For these sheets, get sheet title, sheet key,
owner, user's permission, date-time of last update, version (old vs new
sheet?), various links, and an alternate key (only relevant to old sheets).

The resulting table provides a map between readily available information,
such as sheet title, and more obscure information you might use in scripts,
such as the sheet key. This sort of "table lookup" is implemented in the
helper function \code{\link{identify_ss}}.

Which sheets show up here? Certainly those owned by the user. But also a
subset of the sheets owned by others but visible to the user. We have yet to
find explicit Google documentation on this matter. Anecdotally, sheets owned
by a third party but for which the user has read access seem to appear in
this listing if the user has visited them in the browser. This is an
important point for usability because a sheet can be summoned by title
instead of key \emph{only} if it appears in this listing. For shared sheets
that may not appear in this listing, a more robust workflow is to extract the
key from the browser URL via \code{\link{extract_key_from_url}} and
explicitly specify the sheet in \code{googlesheets} functions by key.
}
\examples{
\dontrun{
gs_ls()
}
}

