% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_merge.R
\name{gs_merge}
\alias{gs_get_merged_cells}
\alias{gs_merge}
\alias{gs_merge_cells}
\alias{gs_unmerge_cells}
\title{Merge/Unmerge Cells or Get Identify Ranges of Merged Cells}
\usage{
gs_merge_cells(ss, ws = 1, range = "A1", merge_type = c("MERGE_ALL",
  "MERGE_COLUMNS", "MERGE_ROWS"), verbose = FALSE)

gs_unmerge_cells(ss, ws = NULL, range = NULL, verbose = FALSE)

gs_get_merged_cells(ss, verbose = FALSE)
}
\arguments{
\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet}

\item{range}{a cell range, as described in \code{\link{cell-specification}}}

\item{merge_type}{string. How the cells should be merged. merge_type must 
take one of the following values: MERGE_ALL, MERGE_COLUMNS, MERGE_ROWS
See the details section for the definition of each of these values.}

\item{verbose}{logical; do you want informative messages?}

\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{verbose}{logical; do you want informative messages?}
}
\description{
These functions will accept a range to merge. The unmerge function will accept 
a range, but the range and ws arguments can be left NULL to unmerge all cells 
in a worksheet or the entire spreadsheet. There is a function to determine the 
ranges where merges occur which helps to formulate the unmerge request which 
must exactly match the range of cells that have been merged.
}
\details{
merge_type takes one of the following values:
\itemize{
 \item{MERGE_ALL - Create a single merge from the range}
 \item{MERGE_COLUMNS - Create a merge for each column in the range}
 \item{MERGE_ROWS - Create a merge for each row in the range}
}
}
\examples{
\dontrun{

gs_merge_cells(gap_ss, ws = 1, range = "A1:B2")
gs_merge_cells(gap_ss, ws = "Africa", range = "A5:E10", merge_type="MERGE_ROWS")
gs_merge_cells(gap_ss, ws = "Africa", range = "A11:C15", merge_type="MERGE_COLUMNS")

gs_unmerge_cells(gap_ss, ws = "Africa", range = "A11:C15")
gs_unmerge_cells(gap_ss, ws = "Africa", range = "A5:E10")
gs_unmerge_cells(gap_ss, ws = 1, range = "A1:B2")

# you can unmerge all cells on a worksheet
gs_unmerge_cells(gap_ss, ws = 1)

# or even unmerge all cells in the entire spreadsheet
gs_unmerge_cells(gap_ss)

}
}

