% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_named_range.R
\name{gs_named_range}
\alias{gs_add_named_range}
\alias{gs_delete_named_range}
\alias{gs_get_named_range}
\alias{gs_named_range}
\alias{gs_update_named_range}
\title{Get, Add, Update, or Delete Named Ranges}
\usage{
gs_add_named_range(ss, name, range = "A1", verbose = FALSE)

gs_update_named_range(ss, named_range, name = NULL, range = NULL,
  verbose = FALSE)

gs_delete_named_range(ss, named_range, verbose = FALSE)

gs_get_named_range(ss, verbose = FALSE)
}
\arguments{
\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{name}{character; a string to name this range when created}

\item{range}{a cell range, as described in \code{\link{cell-specification}}}

\item{verbose}{logical; do you want informative messages?}

\item{named_range}{string; id or name of a named range in the sheet. Use 
\code{\link{gs_get_named_range}} to get details on all named ranges in a sheet.}
}
\description{
These functions provide basic CRUD (Create-Read-Update-Delete) functionality 
with respect to named ranges in Google Sheets.
}
\examples{
\dontrun{

gs_add_named_range(gap_ss, name = "RangeAtAfricaA2", range = "Africa!A2")
gs_get_named_range(gap_ss)

gs_update_named_range(gap_ss, named_range = "RangeAtAfricaA2", 
                      name = "RangeAtAfricaA3", range = "Africa!A3")
gs_update_named_range(gap_ss, named_range = "RangeAtAfricaA3", name = "Range1")
gs_update_named_range(gap_ss, named_range = "Range1", range = "Africa!A4")
gs_get_named_range(gap_ss)

gs_delete_named_range(gap_ss, named_range = "Range1")
gs_get_named_range(gap_ss)
}
}

