% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_note.R
\name{gs_note}
\alias{gs_clear_note}
\alias{gs_insert_note}
\alias{gs_note}
\title{Insert or Clear Notes}
\usage{
gs_insert_note(ss, ws = 1, range = "A1", note = "", byrow = FALSE,
  verbose = FALSE)

gs_clear_note(ss, ws = 1, range = NULL, verbose = FALSE)
}
\arguments{
\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet}

\item{range}{a cell range, as described in \code{\link{cell-specification}}}

\item{note}{character; a single string or vector of strings to populate the note 
of cells in the range. If the range contains more cells than the note vector is 
long, then the elements are recycled.}

\item{byrow}{logical; should we take the note vector and apply to the range
going across a row (byrow = TRUE) or down a column (byrow = FALSE, default)}

\item{verbose}{logical; do you want informative messages?}
}
\description{
Insert or clear cell notes on a particular cell or range of cells. These 
functions allow you to manipulate notes, which are strings as part of the 
cell's metadata. Notes are different than comments, which do not appear to be 
supported by the Sheets V4 API at this time.
}
\details{
The default value for the range argument is NULL, which implies that 
notes should be cleared from the entire sheet. Make sure to specify a range
if you do not want this behavior.
}
\examples{
\dontrun{

gs_insert_note(gap_ss, ws = 1, range = "A1", note = "Test Note")
gs_insert_note(gap_ss, ws = 1, range = "E1:E2", note = c("Note- E1", "Note - E2"))
gs_insert_note(gap_ss, ws = 1, range = "B1:D4", note = "Hello, This is a Note Test!")

gs_clear_note(gap_ss, ws = 1, range = "A1:B2")
gs_clear_note(gap_ss, ws = 1, range = cell_rows(3))
gs_clear_note(gap_ss, ws = 1)

}
}

