% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_protected_range.R
\name{gs_protected_range}
\alias{gs_add_protected_range}
\alias{gs_delete_protected_range}
\alias{gs_get_protected_range}
\alias{gs_protected_range}
\alias{gs_update_protected_range}
\title{Get, Add, Update, or Delete Protected Ranges}
\usage{
gs_add_protected_range(ss, description, range = "A1", named_range = NULL,
  editors = NULL, domain_users_can_edit = FALSE, warning_only = FALSE,
  verbose = FALSE)

gs_update_protected_range(ss, protected_range, description = NULL,
  range = NULL, named_range = NULL, editors = NULL,
  domain_users_can_edit = NULL, warning_only = NULL, verbose = FALSE)

gs_delete_protected_range(ss, protected_range, verbose = FALSE)

gs_get_protected_range(ss, verbose = FALSE)
}
\arguments{
\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{description}{character; The description of this protected range.}

\item{range}{a cell range, as described in \code{\link{cell-specification}}}

\item{named_range}{string; id or name of a named range in the sheet. Use 
\code{\link{gs_get_named_range}} to get details on all named ranges in a sheet.}

\item{editors}{vector, list, or 1 column data.frame with the email addresses 
of people to have edit access to the protected range. If this argument is left 
NULL then only the creating user will have edit access to the protected range.}

\item{domain_users_can_edit}{logical; True if anyone in the document's domain 
has edit access to the protected range. Domain protection is only supported 
on documents within a domain.}

\item{warning_only}{logical; True if this protected range will show a warning 
when editing. Warning-based protection means that every user can edit data 
in the protected range, except editing will prompt a warning asking the user 
to confirm the edit. When writing: if this field is true, then editors is ignored. 
Additionally, if this field is changed from true to false and the editors field is 
not set (nor included in the field mask), then the editors will be set to all 
the editors in the document.}

\item{verbose}{logical; do you want informative messages?}

\item{protected_range}{string; id or name of a Protected Range in the sheet. Use 
\code{\link{gs_get_protected_range}} to get details on all Protected Ranges in a sheet.}

\item{description}{character; The description of this protected range.}

\item{range}{a cell range, as described in \code{\link{cell-specification}}}

\item{named_range}{string; id or name of a named range in the sheet. Use 
\code{\link{gs_get_named_range}} to get details on all named ranges in a sheet.}

\item{editors}{vector, list, or 1 column data.frame with the email addresses 
of people to have edit access to the protected range. If this argument is left 
NULL then only the creating user will have edit access to the protected range.}

\item{domain_users_can_edit}{logical; True if anyone in the document's domain 
has edit access to the protected range. Domain protection is only supported 
on documents within a domain.}

\item{warning_only}{logical; True if this protected range will show a warning 
when editing. Warning-based protection means that every user can edit data 
in the protected range, except editing will prompt a warning asking the user 
to confirm the edit. When writing: if this field is true, then editors is ignored. 
Additionally, if this field is changed from true to false and the editors field is 
not set (nor included in the field mask), then the editors will be set to all 
the editors in the document.}

\item{protected_range}{string; id or name of a Protected Range in the sheet. Use 
\code{\link{gs_get_protected_range}} to get details on all Protected Ranges in a sheet.}
}
\description{
These functions provide basic CRUD (Create-Read-Update-Delete) functionality 
with respect to protected ranges in Google Sheets.
}
\examples{
\dontrun{

gs_add_protected_range(gap_ss, range = "Africa!A1", description = "PR - Test")
gs_add_protected_range(gap_ss, range = cell_limits(sheet="Africa")) # protect whole sheet!
gs_get_protected_range(gap_ss)

gs_update_protected_range(gap_ss, protected_range = "PR - Test", range = "Africa!A2")
gs_update_protected_range(gap_ss, protected_range = "PR - Test", description = "PR - Test2")
gs_update_protected_range(gap_ss, protected_range = "PR - Test2", named_range = 'Range1')

gs_delete_protected_range(gap_ss, protected_range = "PR - Test2")
pr <- gs_get_protected_range(gap_ss)
gs_delete_protected_range(gap_ss, protected_range = pr$protectedRangeId[1])
gs_get_protected_range(gap_ss)
}
}

