% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_register_old.R
\name{gs_register_old}
\alias{gs_register_old}
\alias{gs_key_old}
\alias{gs_url_old}
\title{Register an old Google Sheet}
\usage{
gs_key_old(x, lookup = NULL, visibility = NULL, verbose = TRUE)

gs_url_old(x, lookup = NULL, visibility = NULL, verbose = TRUE)
}
\arguments{
\item{x}{sheet-identifying information; a character vector of length one
holding old sheet key or browser URL.}

\item{lookup}{logical, optional. Controls whether \code{googlesheets} will
place authorized API requests during registration. If unspecified, will
be set to \code{TRUE} if authorization has previously been used in this R
session, if working directory contains a file named \code{.httr-oauth}, or
if \code{x} is a worksheets feed or \code{googlesheet} object that
specifies "public" visibility.}

\item{visibility}{character, either "public" or "private". Consulted during
explicit construction of a worksheets feed from a key, which happens only
when \code{lookup = FALSE} and \code{googlesheets} is prevented from
looking up information in the spreadsheets feed. If unspecified, will be
set to "public" if \code{lookup = FALSE} and "private" if \code{lookup =
TRUE}. Consult the API docs for more info about
\href{https://developers.google.com/google-apps/spreadsheets/worksheets#sheets_api_urls_visibilities_and_projections}{visibility}}

\item{verbose}{logical; do you want informative messages?}
}
\value{
a \code{googlesheet} object
}
\description{
Register an old Google Sheet
}
\examples{
# The following spreadsheet from Gapminder collection would not get identified using gs_key()
gs_key_old("phAwcNAVuyj0TAlJeCEzcGQ", lookup=FALSE)

# this spreadsheet from Gapminder collection would not get identified using gs_key()
gs_url_old("http://docs.google.com/spreadsheet/pub?key=phAwcNAVuyj0TAlJeCEzcGQ", lookup=FALSE)
}
