% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny-with-gs.R
\name{gs_shiny_form_url}
\alias{gs_shiny_form_url}
\title{Form the URL to use to authenticate a user}
\usage{
gs_shiny_form_url(redirect_uri)
}
\arguments{
\item{redirect_uri}{redirect uri (shiny app url or localhost for local testing)}

\item{client_id}{application client id}

\item{client}{secret application client secret}
}
\description{
Once a user authenticates, the response is sent to the \code{redirect_uri}
and contains an authorization code which \code{shiny_get_token} uses to
exchange for an access token.
}

