% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gs_shiny.R
\name{gs_shiny_get_token}
\alias{gs_shiny_get_token}
\title{Exchange authorization code for an access token}
\usage{
gs_shiny_get_token(auth_code, redirect_uri)
}
\arguments{
\item{auth_code}{authorization code returned by Google that appears in URL}

\item{redirect_uri}{redirect uri (shiny app url or localhost for local testing)}

\item{client_id}{application client id}

\item{client}{secret application client secret}
}
\value{
A list containing access_token, access_type, expires_in.
}
\description{
Use the authorization code in the return URL to exchange for an access_token.
}

