% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{gs_user}
\alias{gs_user}
\title{Print information about authorized user}
\usage{
gs_user()
}
\description{
Display information about a user that has been authorized via
\code{authorize}: the name of the user, the Google account of the user, the
date-time of authorization for the current session, date-time of access
token expiry, and the validity of the current access token.
}
\examples{
\dontrun{
gs_user()
}
}

