% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_ws.R
\name{gs_ws}
\alias{gs_ws}
\title{Retrieve a worksheet-describing list from a \code{googlesheet}}
\usage{
gs_ws(ss, ws, verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet}

\item{verbose}{logical; do you want informative messages?}
}
\description{
From a \code{\link{googlesheet}}, retrieve a list (actually a row of a
data.frame) giving everything we know about a specific worksheet.
}
\keyword{internal}
