% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gs_ws.R
\name{gs_ws_ls}
\alias{gs_ws_ls}
\title{List the worksheets in a Google Sheet}
\usage{
gs_ws_ls(ss)
}
\arguments{
\item{ss}{a registered Google spreadsheet}
}
\description{
Retrieve the titles of all the worksheets in a \code{\link{googlesheet}}.
}
\examples{
\dontrun{
gap_key <- "1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA"
gap_ss <- gs_key(gap_key)
gs_ws_ls(gap_ss)
}
}

