% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_ws.R
\name{gs_ws_modify}
\alias{gs_ws_modify}
\title{Modify a worksheet's title or size}
\usage{
gs_ws_modify(ss, from = NULL, to = NULL, new_dim = NULL,
  verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{from}{positive integer or character string specifying index or title,
respectively, of the worksheet}

\item{to}{character string for new title of worksheet}

\item{new_dim}{vector of length 2 specifying the row and column extent of the
worksheet}

\item{verbose}{logical; do you want informative messages?}
}
\value{
a \code{\link{googlesheet}} object
}
\description{
Modify a worksheet's title or size
}
\keyword{internal}
