% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auth.R
\name{gsheets_auth}
\alias{gsheets_auth}
\title{Format token for making request}
\usage{
gsheets_auth(token)
}
\arguments{
\item{token}{Google token}
}
\description{
Check if token is obtained from Google login or oauth2.0 and format it for
making request. Format token as a header (login) or in configuations
(oauth2.0).
}

