% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/auth.R
\name{gsheets_check}
\alias{gsheets_check}
\title{Check status of http response}
\usage{
gsheets_check(req)
}
\arguments{
\item{req}{response from \code{\link{gsheets_GET}} request}
}
\description{
Google returns status 200 (success) or 403 (failure), show error msg if 403.
}

