% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsheets_DELETE}
\alias{gsheets_DELETE}
\title{Create DELETE request}
\usage{
gsheets_DELETE(url, token = get_google_token())
}
\arguments{
\item{url}{URL for DELETE request}

\item{token}{Google auth token obtained from \code{\link{login}}
or \code{\link{authorize}}}
}
\description{
Make DELETE request to Google Sheets API.
}

