% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsheets_GET}
\alias{gsheets_GET}
\title{Create GET request}
\usage{
gsheets_GET(feed, client = NULL, key = NULL, ws_id = NULL,
  visibility = "private", projection = "full", min_row = NULL,
  max_row = NULL, min_col = NULL, max_col = NULL)
}
\arguments{
\item{client}{Client object or NULL if accessing public worksheets feed}

\item{key}{Spreadsheet key}

\item{visibility}{Either private or public}

\item{projection}{Either full or basic}

\item{feed_type}{One of the following: spreadsheets, worksheets, list, cells}

\item{worksheet_id}{id of worksheet contained in spreadsheet}
}
\description{
Construct GET requests to Google Sheets API. This function handles the
construction of urls and making GET requests so urls arent lying around.
}
\examples{
gsheets_GET("spreadsheets", NULL,
"1WNUDoBbGsPccRkXlLqeUK9JUQNnqq2yvc9r7-cmEaZU", visibility = "public")
}

