% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/helpers.R
\name{gsheets_parse}
\alias{gsheets_parse}
\title{Wrapper around xmlInternalTreeParse}
\usage{
gsheets_parse(req)
}
\arguments{
\item{req}{response from \code{\link{gsheets_GET}} request}
}
\description{
Simply for code neatness.
}

