% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http_requests.R
\name{gsheets_POST}
\alias{gsheets_POST}
\title{Create POST request}
\usage{
gsheets_POST(url, the_body)
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{the_body}{body of POST request}
}
\description{
Make POST request to Google Sheets API.
}
\keyword{internal}

