% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_requests.R
\name{gsheets_PUT}
\alias{gsheets_PUT}
\title{Create PUT request}
\usage{
gsheets_PUT(url, the_body)
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{the_body}{body of PUT request}
}
\description{
Make PUT request to Google Sheets API.
}
\keyword{internal}

