% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gspreadr.R
\docType{package}
\name{gspreadr}
\alias{gspreadr}
\alias{gspreadr-package}
\title{gspreadr}
\description{
Google spreadsheets R API
}
\details{
See the \href{https://github.com/jennybc/gspreadr}{README} on GitHub
}

