% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_AddBandingRequest}
\alias{gsv4_AddBandingRequest}
\title{Google Sheets Object Makers}
\usage{
gsv4_AddBandingRequest(bandedRange=NULL)
}
\arguments{
\item{bandedRange}{\code{\link{gsv4_BandedRange}} object. A banded (alternating colors) range in a sheet.}
}
\value{
AddBandingRequest
}
\description{
A set of functions to make API objects from R data types
}
\details{
gsv4_AddBandingRequest

Adds a new banded range to the spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#AddBandingRequest}{Google's Documentation for AddBandingRequest}
}

