% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_AddChartRequest}
\alias{gsv4_AddChartRequest}
\title{gsv4_AddChartRequest}
\usage{
gsv4_AddChartRequest(chart = NULL)
}
\arguments{
\item{chart}{\code{\link{gsv4_EmbeddedChart}} object. A chart embedded in a sheet.}
}
\value{
AddChartRequest
}
\description{
Adds a chart to a sheet in the spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#AddChartRequest}{Google's Documentation for AddChartRequest}
}

