% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_AddConditionalFormatRuleRequest}
\alias{gsv4_AddConditionalFormatRuleRequest}
\title{gsv4_AddConditionalFormatRuleRequest}
\usage{
gsv4_AddConditionalFormatRuleRequest(index=NULL, rule=NULL)
}
\arguments{
\item{index}{integer. The zero-based index where the rule should be inserted.}

\item{rule}{\code{\link{gsv4_ConditionalFormatRule}} object. A rule describing a conditional format.}
}
\value{
AddConditionalFormatRuleRequest
}
\description{
Adds a new conditional format rule at the given index.
All subsequent rules' indexes are incremented.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#AddConditionalFormatRuleRequest}{Google's Documentation for AddConditionalFormatRuleRequest}
}

