% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_AddFilterViewRequest}
\alias{gsv4_AddFilterViewRequest}
\title{gsv4_AddFilterViewRequest}
\usage{
gsv4_AddFilterViewRequest(filter = NULL)
}
\arguments{
\item{filter}{\code{\link{gsv4_FilterView}} object. A filter view.}
}
\value{
AddFilterViewRequest
}
\description{
Adds a filter view.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#AddFilterViewRequest}{Google's Documentation for AddFilterViewRequest}
}

