% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_AddNamedRangeRequest}
\alias{gsv4_AddNamedRangeRequest}
\title{gsv4_AddNamedRangeRequest}
\usage{
gsv4_AddNamedRangeRequest(namedRange = NULL)
}
\arguments{
\item{namedRange}{\code{\link{gsv4_NamedRange}} object. A named range.}
}
\value{
AddNamedRangeRequest
}
\description{
Adds a named range to the spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#AddNamedRangeRequest}{Google's Documentation for AddNamedRangeRequest}
}

