% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_AddProtectedRangeRequest}
\alias{gsv4_AddProtectedRangeRequest}
\title{gsv4_AddProtectedRangeRequest}
\usage{
gsv4_AddProtectedRangeRequest(protectedRange = NULL)
}
\arguments{
\item{protectedRange}{\code{\link{gsv4_ProtectedRange}} object. A protected range.}
}
\value{
AddProtectedRangeRequest
}
\description{
Adds a new protected range.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#AddProtectedRangeRequest}{Google's Documentation for AddProtectedRangeRequest}
}

