% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_AddSheetRequest}
\alias{gsv4_AddSheetRequest}
\title{gsv4_AddSheetRequest}
\usage{
gsv4_AddSheetRequest(properties=NULL)
}
\arguments{
\item{properties}{\code{\link{gsv4_SheetProperties}} object. Properties of a sheet.}
}
\value{
AddSheetRequest
}
\description{
Adds a new sheet.
When a sheet is added at a given index,
all subsequent sheets' indexes are incremented.
To add an object sheet, use AddChartRequest instead and specify
EmbeddedObjectPosition.sheetId or
EmbeddedObjectPosition.newSheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#AddSheetRequest}{Google's Documentation for AddSheetRequest}
}

