% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gsv4_anchor_to_grid_coordinate}
\alias{gsv4_anchor_to_grid_coordinate}
\title{gsv4_anchor_to_grid_coordinate}
\usage{
gsv4_anchor_to_grid_coordinate(anchor, ss, ws = 1)
}
\arguments{
\item{anchor}{single character string specifying the upper left cell of the
cell range to edit; positioning notation can be either "A1" or "R1C1"}

\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet}
}
\description{
Converting an anchor reference to a gridCoordinate object as specified in the Sheets V4 API.
}
\examples{
\dontrun{
gsv4_anchor_to_grid_coordinate(gap_ss, anchor="A1")
gsv4_anchor_to_grid_coordinate(gap_ss, ws=1, anchor="A1")
gsv4_anchor_to_grid_coordinate(gap_ss, ws="Americas", anchor="A1")
gsv4_anchor_to_grid_coordinate(gap_ss, anchor="Americas!A1")
}

}

