% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_AppendCellsRequest}
\alias{gsv4_AppendCellsRequest}
\title{gsv4_AppendCellsRequest}
\usage{
gsv4_AppendCellsRequest(sheetId = NULL, rows = NULL, fields = NULL)
}
\arguments{
\item{sheetId}{integer. The sheet ID to append the data to.}

\item{rows}{list of \code{\link{gsv4_RowData}} objects. The data to append.}

\item{fields}{string. The fields of CellData that should be updated.
At least one field must be specified.
The root is the CellData; 'row.values.' should not be specified.
A single `"*"` can be used as short-hand for listing every field.}
}
\value{
AppendCellsRequest
}
\description{
Adds new cells after the last row with data in a sheet,
inserting new rows into the sheet if necessary.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#AppendCellsRequest}{Google's Documentation for AppendCellsRequest}
}

