% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_AppendDimensionRequest}
\alias{gsv4_AppendDimensionRequest}
\title{gsv4_AppendDimensionRequest}
\usage{
gsv4_AppendDimensionRequest(sheetId=NULL, dimension=NULL, length=NULL)
}
\arguments{
\item{sheetId}{integer. The sheet to append rows or columns to.}

\item{dimension}{string. Whether rows or columns should be appended. dimension must take one of the following values: DIMENSION_UNSPECIFIED, ROWS, COLUMNS
See the details section for the definition of each of these values.}

\item{length}{integer. The number of rows or columns to append.}
}
\value{
AppendDimensionRequest
}
\description{
Appends rows or columns to the end of a sheet.
}
\details{
dimension takes one of the following values:
\itemize{
 \item{DIMENSION_UNSPECIFIED - The default value, do not use.}
 \item{ROWS - Operates on the rows of a sheet.}
 \item{COLUMNS - Operates on the columns of a sheet.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#AppendDimensionRequest}{Google's Documentation for AppendDimensionRequest}
}

