% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_AutoFillRequest}
\alias{gsv4_AutoFillRequest}
\title{gsv4_AutoFillRequest}
\usage{
gsv4_AutoFillRequest(range=NULL, sourceAndDestination=NULL, useAlternateSeries=NULL)
}
\arguments{
\item{range}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}

\item{sourceAndDestination}{\code{\link{gsv4_SourceAndDestination}} object. A combination of a source range and how to extend that source.}

\item{useAlternateSeries}{logical. TRUE if we should generate data with the "alternate" series.
This differs based on the type and amount of source data.}
}
\value{
AutoFillRequest
}
\description{
Fills in more data based on existing data.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#AutoFillRequest}{Google's Documentation for AutoFillRequest}
}

