% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_AutoResizeDimensionsRequest}
\alias{gsv4_AutoResizeDimensionsRequest}
\title{gsv4_AutoResizeDimensionsRequest}
\usage{
gsv4_AutoResizeDimensionsRequest(dimensions = NULL)
}
\arguments{
\item{dimensions}{\code{\link{gsv4_DimensionRange}} object. A range along a single dimension on a sheet.
All indexes are zero-based.
Indexes are half open: the start index is inclusive
and the end index is exclusive.
Missing indexes indicate the range is unbounded on that side.}
}
\value{
AutoResizeDimensionsRequest
}
\description{
Automatically resizes one or more dimensions based on the contents
of the cells in that dimension.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#AutoResizeDimensionsRequest}{Google's Documentation for AutoResizeDimensionsRequest}
}

