% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_BandedRange}
\alias{gsv4_BandedRange}
\title{gsv4_BandedRange}
\usage{
gsv4_BandedRange(range=NULL, bandedRangeId=NULL, columnProperties=NULL, rowProperties=NULL)
}
\arguments{
\item{range}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}

\item{bandedRangeId}{integer. The id of the banded range.}

\item{columnProperties}{\code{\link{gsv4_BandingProperties}} object. Properties referring a single dimension (either row or column). If both
BandedRange.row_properties and BandedRange.column_properties are
set, the fill colors are applied to cells according to the following rules:

* header_color and footer_color take priority over band colors.
* first_band_color takes priority over second_band_color.
* row_properties takes priority over column_properties.

For example, the first row color takes priority over the first column
color, but the first column color takes priority over the second row color.
Similarly, the row header takes priority over the column header in the
top left cell, but the column header takes priority over the first row
color if the row header is not set.}

\item{rowProperties}{\code{\link{gsv4_BandingProperties}} object. Properties referring a single dimension (either row or column). If both
BandedRange.row_properties and BandedRange.column_properties are
set, the fill colors are applied to cells according to the following rules:

* header_color and footer_color take priority over band colors.
* first_band_color takes priority over second_band_color.
* row_properties takes priority over column_properties.

For example, the first row color takes priority over the first column
color, but the first column color takes priority over the second row color.
Similarly, the row header takes priority over the column header in the
top left cell, but the column header takes priority over the first row
color if the row header is not set.}
}
\value{
BandedRange
}
\description{
A banded (alternating colors) range in a sheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#BandedRange}{Google's Documentation for BandedRange}
}

