% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_BasicChartAxis}
\alias{gsv4_BasicChartAxis}
\title{gsv4_BasicChartAxis}
\usage{
gsv4_BasicChartAxis(format = NULL, position = NULL, title = NULL)
}
\arguments{
\item{format}{\code{\link{gsv4_TextFormat}} object. The format of a run of text in a cell.
Absent values indicate that the field isn't specified.}

\item{position}{string. The position of this axis. position must take one of the following values: BASIC_CHART_AXIS_POSITION_UNSPECIFIED, BOTTOM_AXIS, LEFT_AXIS, RIGHT_AXIS
See the details section for the definition of each of these values.}

\item{title}{string. The title of this axis. If set, this overrides any title inferred
from headers of the data.}
}
\value{
BasicChartAxis
}
\description{
An axis of the chart.
A chart may not have more than one axis per
axis position.
}
\details{
position takes one of the following values:
\itemize{
 \item{BASIC_CHART_AXIS_POSITION_UNSPECIFIED - Default value, do not use.}
 \item{BOTTOM_AXIS - The axis rendered at the bottom of a chart.
For most charts, this is the standard major axis.
For bar charts, this is a minor axis.}
 \item{LEFT_AXIS - The axis rendered at the left of a chart.
For most charts, this is a minor axis.
For bar charts, this is the standard major axis.}
 \item{RIGHT_AXIS - The axis rendered at the right of a chart.
For most charts, this is a minor axis.
For bar charts, this is an unusual major axis.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#BasicChartAxis}{Google's Documentation for BasicChartAxis}
}

