% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_BasicChartDomain}
\alias{gsv4_BasicChartDomain}
\title{gsv4_BasicChartDomain}
\usage{
gsv4_BasicChartDomain(domain = NULL)
}
\arguments{
\item{domain}{\code{\link{gsv4_ChartData}} object. The data included in a domain or series.}
}
\value{
BasicChartDomain
}
\description{
The domain of a chart.
For example, if charting stock prices over time, this would be the date.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#BasicChartDomain}{Google's Documentation for BasicChartDomain}
}

