% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_BasicChartSeries}
\alias{gsv4_BasicChartSeries}
\title{gsv4_BasicChartSeries}
\usage{
gsv4_BasicChartSeries(series=NULL, targetAxis=NULL, type=NULL)
}
\arguments{
\item{series}{\code{\link{gsv4_ChartData}} object. The data included in a domain or series.}

\item{targetAxis}{string. The minor axis that will specify the range of values for this series.
For example, if charting stocks over time, the "Volume" series
may want to be pinned to the right with the prices pinned to the left,
because the scale of trading volume is different than the scale of
prices.
It is an error to specify an axis that isn't a valid minor axis
for the chart's type. targetAxis must take one of the following values: BASIC_CHART_AXIS_POSITION_UNSPECIFIED, BOTTOM_AXIS, LEFT_AXIS, RIGHT_AXIS
See the details section for the definition of each of these values.}

\item{type}{string. The type of this series. Valid only if the
chartType is
COMBO.
Different types will change the way the series is visualized.
Only LINE, AREA,
and COLUMN are supported. type must take one of the following values: BASIC_CHART_TYPE_UNSPECIFIED, BAR, LINE, AREA, COLUMN, SCATTER, COMBO
See the details section for the definition of each of these values.}
}
\value{
BasicChartSeries
}
\description{
A single series of data in a chart.
For example, if charting stock prices over time, multiple series may exist,
one for the "Open Price", "High Price", "Low Price" and "Close Price".
}
\details{
targetAxis takes one of the following values:
\itemize{
 \item{BASIC_CHART_AXIS_POSITION_UNSPECIFIED - Default value, do not use.}
 \item{BOTTOM_AXIS - The axis rendered at the bottom of a chart.
For most charts, this is the standard major axis.
For bar charts, this is a minor axis.}
 \item{LEFT_AXIS - The axis rendered at the left of a chart.
For most charts, this is a minor axis.
For bar charts, this is the standard major axis.}
 \item{RIGHT_AXIS - The axis rendered at the right of a chart.
For most charts, this is a minor axis.
For bar charts, this is an unusual major axis.}
}

type takes one of the following values:
\itemize{
 \item{BASIC_CHART_TYPE_UNSPECIFIED - Default value, do not use.}
 \item{BAR - A <a href="/chart/interactive/docs/gallery/barchart">bar chart</a>.}
 \item{LINE - A <a href="/chart/interactive/docs/gallery/linechart">line chart</a>.}
 \item{AREA - An <a href="/chart/interactive/docs/gallery/areachart">area chart</a>.}
 \item{COLUMN - A <a href="/chart/interactive/docs/gallery/columnchart">column chart</a>.}
 \item{SCATTER - A <a href="/chart/interactive/docs/gallery/scatterchart">scatter chart</a>.}
 \item{COMBO - A <a href="/chart/interactive/docs/gallery/combochart">combo chart</a>.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#BasicChartSeries}{Google's Documentation for BasicChartSeries}
}

