% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_BasicChartSpec}
\alias{gsv4_BasicChartSpec}
\title{gsv4_BasicChartSpec}
\usage{
gsv4_BasicChartSpec(axis = NULL, chartType = NULL, domains = NULL,
  headerCount = NULL, legendPosition = NULL, series = NULL)
}
\arguments{
\item{axis}{list of \code{\link{gsv4_BasicChartAxis}} objects. The axis on the chart.}

\item{chartType}{string. The type of the chart. chartType must take one of the following values: BASIC_CHART_TYPE_UNSPECIFIED, BAR, LINE, AREA, COLUMN, SCATTER, COMBO
See the details section for the definition of each of these values.}

\item{domains}{list of \code{\link{gsv4_BasicChartDomain}} objects. The domain of data this is charting.
Only a single domain is supported.}

\item{headerCount}{integer. The number of rows or columns in the data that are "headers".
If not set, Google Sheets will guess how many rows are headers based
on the data.

(Note that BasicChartAxis.title may override the axis title
 inferred from the header values.)}

\item{legendPosition}{string. The position of the chart legend. legendPosition must take one of the following values: BASIC_CHART_LEGEND_POSITION_UNSPECIFIED, BOTTOM_LEGEND, LEFT_LEGEND, RIGHT_LEGEND, TOP_LEGEND, NO_LEGEND
See the details section for the definition of each of these values.}

\item{series}{list of \code{\link{gsv4_BasicChartSeries}} objects. The data this chart is visualizing.}
}
\value{
BasicChartSpec
}
\description{
The specification for a basic chart.  See BasicChartType for the list
of charts this supports.
}
\details{
chartType takes one of the following values:
\itemize{
 \item{BASIC_CHART_TYPE_UNSPECIFIED - Default value, do not use.}
 \item{BAR - A <a href="/chart/interactive/docs/gallery/barchart">bar chart</a>.}
 \item{LINE - A <a href="/chart/interactive/docs/gallery/linechart">line chart</a>.}
 \item{AREA - An <a href="/chart/interactive/docs/gallery/areachart">area chart</a>.}
 \item{COLUMN - A <a href="/chart/interactive/docs/gallery/columnchart">column chart</a>.}
 \item{SCATTER - A <a href="/chart/interactive/docs/gallery/scatterchart">scatter chart</a>.}
 \item{COMBO - A <a href="/chart/interactive/docs/gallery/combochart">combo chart</a>.}
}

legendPosition takes one of the following values:
\itemize{
 \item{BASIC_CHART_LEGEND_POSITION_UNSPECIFIED - Default value, do not use.}
 \item{BOTTOM_LEGEND - The legend is rendered on the bottom of the chart.}
 \item{LEFT_LEGEND - The legend is rendered on the left of the chart.}
 \item{RIGHT_LEGEND - The legend is rendered on the right of the chart.}
 \item{TOP_LEGEND - The legend is rendered on the top of the chart.}
 \item{NO_LEGEND - No legend is rendered.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#BasicChartSpec}{Google's Documentation for BasicChartSpec}
}

