% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_BasicFilter}
\alias{gsv4_BasicFilter}
\title{gsv4_BasicFilter}
\usage{
gsv4_BasicFilter(range = NULL, criteria = NULL, sortSpecs = NULL)
}
\arguments{
\item{range}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}

\item{criteria}{list or data.frame of \code{\link{gsv4_FilterCriteria}} objects. The criteria for showing/hiding values per column.
The map's key is the column index, and the value is the criteria for
that column.}

\item{sortSpecs}{list of \code{\link{gsv4_SortSpec}} objects. The sort order per column. Later specifications are used when values
are equal in the earlier specifications.}
}
\value{
BasicFilter
}
\description{
The default filter associated with a sheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#BasicFilter}{Google's Documentation for BasicFilter}
}

