% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_BatchClearValuesRequest}
\alias{gsv4_BatchClearValuesRequest}
\title{gsv4_BatchClearValuesRequest}
\usage{
gsv4_BatchClearValuesRequest(ranges = NULL)
}
\arguments{
\item{ranges}{list. The ranges to clear, in A1 notation.}
}
\value{
BatchClearValuesRequest
}
\description{
The request for clearing more than one range of values in a spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#BatchClearValuesRequest}{Google's Documentation for BatchClearValuesRequest}
}

