% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheets.R
\name{gsv4_batchUpdate}
\alias{gsv4_batchUpdate}
\title{gsv4_batchUpdate}
\usage{
gsv4_batchUpdate(spreadsheetId, input, standard_params = list(...), ...)
}
\arguments{
\item{spreadsheetId}{string (required). The spreadsheet to apply the updates to.}

\item{input}{\code{\link{gsv4_BatchUpdateSpreadsheetRequest}}. The request for updating any aspect of a spreadsheet.}

\item{standard_params}{a list of parameters for controlling the HTTP request and its response.
Refer to \code{\link{gsv4_standard_parameters}} for details on its arguments}

\item{...}{arguments to be used to form standard_params argument if it is not supplied directly.}
}
\value{
BatchUpdateSpreadsheetResponse
}
\description{
Applies one or more updates to the spreadsheet.
}
\details{
Each request is validated before
being applied. If any request is not valid then the entire request will
fail and nothing will be applied.

Some requests have replies to
give you some information about how
they are applied. The replies will mirror the requests.  For example,
if you applied 4 updates and the 3rd one had a reply, then the
response will have 2 empty replies, the actual reply, and another empty
reply, in that order.

Due to the collaborative nature of spreadsheets, it is not guaranteed that
the spreadsheet will reflect exactly your changes after this completes,
however it is guaranteed that the updates in the request will be
applied together atomically. Your changes may be altered with respect to
collaborator changes. If there are no collaborators, the spreadsheet
should reflect your changes.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets/batchUpdate}{Google's Documentation for batchUpdate}
}

