% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_BatchUpdateSpreadsheetRequest}
\alias{gsv4_BatchUpdateSpreadsheetRequest}
\title{gsv4_BatchUpdateSpreadsheetRequest}
\usage{
gsv4_BatchUpdateSpreadsheetRequest(includeSpreadsheetInResponse=NULL, requests=NULL, responseIncludeGridData=NULL, responseRanges=NULL)
}
\arguments{
\item{includeSpreadsheetInResponse}{logical. Determines if the update response should include the spreadsheet
resource.}

\item{requests}{list of \code{\link{gsv4_Request}} objects. A list of updates to apply to the spreadsheet.
Requests will be applied in the order they are specified.
If any request is not valid, no requests will be applied.}

\item{responseIncludeGridData}{logical. TRUE if grid data should be returned. Meaningful only if
if include_spreadsheet_response is 'TRUE'.
This parameter is ignored if a field mask was set in the request.}

\item{responseRanges}{list. Limits the ranges included in the response spreadsheet.
Meaningful only if include_spreadsheet_response is 'TRUE'.}
}
\value{
BatchUpdateSpreadsheetRequest
}
\description{
The request for updating any aspect of a spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#BatchUpdateSpreadsheetRequest}{Google's Documentation for BatchUpdateSpreadsheetRequest}
}

