% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_BatchUpdateValuesRequest}
\alias{gsv4_BatchUpdateValuesRequest}
\title{gsv4_BatchUpdateValuesRequest}
\usage{
gsv4_BatchUpdateValuesRequest(data = NULL, includeValuesInResponse = NULL,
  responseDateTimeRenderOption = NULL, responseValueRenderOption = NULL,
  valueInputOption = NULL)
}
\arguments{
\item{data}{list of \code{\link{gsv4_ValueRange}} objects. The new values to apply to the spreadsheet.}

\item{includeValuesInResponse}{logical. Determines if the update response should include the values
of the cells that were updated. By default, responses
do not include the updated values. The `updatedData` field within
each of the BatchUpdateValuesResponse.responses will contain
the updated values. If the range to write was larger than than the range
actually written, the response will include all values in the requested
range (excluding trailing empty rows and columns).}

\item{responseDateTimeRenderOption}{string. Determines how dates, times, and durations in the response should be
rendered. This is ignored if response_value_render_option is
FORMATTED_VALUE.
The default dateTime render option is [DateTimeRenderOption.SERIAL_NUMBER]. responseDateTimeRenderOption must take one of the following values: SERIAL_NUMBER, FORMATTED_STRING
See the details section for the definition of each of these values.}

\item{responseValueRenderOption}{string. Determines how values in the response should be rendered.
The default render option is ValueRenderOption.FORMATTED_VALUE. responseValueRenderOption must take one of the following values: FORMATTED_VALUE, UNFORMATTED_VALUE, FORMULA
See the details section for the definition of each of these values.}

\item{valueInputOption}{string. How the input data should be interpreted. valueInputOption must take one of the following values: INPUT_VALUE_OPTION_UNSPECIFIED, RAW, USER_ENTERED
See the details section for the definition of each of these values.}
}
\value{
BatchUpdateValuesRequest
}
\description{
The request for updating more than one range of values in a spreadsheet.
}
\details{
responseDateTimeRenderOption takes one of the following values:
\itemize{
 \item{SERIAL_NUMBER - Instructs date, time, datetime, and duration fields to be output
as doubles in "serial number" format, as popularized by Lotus 1-2-3.
The whole number portion of the value (left of the decimal) counts
the days since December 30th 1899. The fractional portion (right of
the decimal) counts the time as a fraction of the day. For example,
January 1st 1900 at noon would be 2.5, 2 because it's 2 days after
December 30st 1899, and .5 because noon is half a day.  February 1st
1900 at 3pm would be 33.625. This correctly treats the year 1900 as
not a leap year.}
 \item{FORMATTED_STRING - Instructs date, time, datetime, and duration fields to be output
as strings in their given number format (which is dependent
on the spreadsheet locale).}
}

responseValueRenderOption takes one of the following values:
\itemize{
 \item{FORMATTED_VALUE - Values will be calculated & formatted in the reply according to the
cell's formatting.  Formatting is based on the spreadsheet's locale,
not the requesting user's locale.
For example, if `A1` is `1.23` and `A2` is `=A1` and formatted as currency,
then `A2` would return `"$1.23"`.}
 \item{UNFORMATTED_VALUE - Values will be calculated, but not formatted in the reply.
For example, if `A1` is `1.23` and `A2` is `=A1` and formatted as currency,
then `A2` would return the number `1.23`.}
 \item{FORMULA - Values will not be calculated.  The reply will include the formulas.
For example, if `A1` is `1.23` and `A2` is `=A1` and formatted as currency,
then A2 would return `"=A1"`.}
}

valueInputOption takes one of the following values:
\itemize{
 \item{INPUT_VALUE_OPTION_UNSPECIFIED - Default input value. This value must not be used.}
 \item{RAW - The values the user has entered will not be parsed and will be stored
as-is.}
 \item{USER_ENTERED - The values will be parsed as if the user typed them into the UI.
Numbers will stay as numbers, but strings may be converted to numbers,
dates, etc. following the same rules that are applied when entering
text into a cell via the Google Sheets UI.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#BatchUpdateValuesRequest}{Google's Documentation for BatchUpdateValuesRequest}
}

