% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_BooleanCondition}
\alias{gsv4_BooleanCondition}
\title{gsv4_BooleanCondition}
\usage{
gsv4_BooleanCondition(type = NULL, values = NULL)
}
\arguments{
\item{type}{string. The type of condition. type must take one of the following values: CONDITION_TYPE_UNSPECIFIED, NUMBER_GREATER, NUMBER_GREATER_THAN_EQ, NUMBER_LESS, NUMBER_LESS_THAN_EQ, NUMBER_EQ, NUMBER_NOT_EQ, NUMBER_BETWEEN, NUMBER_NOT_BETWEEN, TEXT_CONTAINS, TEXT_NOT_CONTAINS, TEXT_STARTS_WITH, TEXT_ENDS_WITH, TEXT_EQ, TEXT_IS_EMAIL, TEXT_IS_URL, DATE_EQ, DATE_BEFORE, DATE_AFTER, DATE_ON_OR_BEFORE, DATE_ON_OR_AFTER, DATE_BETWEEN, DATE_NOT_BETWEEN, DATE_IS_VALID, ONE_OF_RANGE, ONE_OF_LIST, BLANK, NOT_BLANK, CUSTOM_FORMULA
See the details section for the definition of each of these values.}

\item{values}{list of \code{\link{gsv4_ConditionValue}} objects. The values of the condition. The number of supported values depends
on the condition type.  Some support zero values,
others one or two values,
and ConditionType.ONE_OF_LIST supports an arbitrary number of values.}
}
\value{
BooleanCondition
}
\description{
A condition that can evaluate to true or false.
BooleanConditions are used by conditional formatting,
data validation, and the criteria in filters.
}
\details{
type takes one of the following values:
\itemize{
 \item{CONDITION_TYPE_UNSPECIFIED - The default value, do not use.}
 \item{NUMBER_GREATER - The cell's value must be greater than the condition's value.
Supported by data validation, conditional formatting and filters.
Requires a single ConditionValue.}
 \item{NUMBER_GREATER_THAN_EQ - The cell's value must be greater than or equal to the condition's value.
Supported by data validation, conditional formatting and filters.
Requires a single ConditionValue.}
 \item{NUMBER_LESS - The cell's value must be less than the condition's value.
Supported by data validation, conditional formatting and filters.
Requires a single ConditionValue.}
 \item{NUMBER_LESS_THAN_EQ - The cell's value must be less than or equal to the condition's value.
Supported by data validation, conditional formatting and filters.
Requires a single ConditionValue.}
 \item{NUMBER_EQ - The cell's value must be equal to the condition's value.
Supported by data validation, conditional formatting and filters.
Requires a single ConditionValue.}
 \item{NUMBER_NOT_EQ - The cell's value must be not equal to the condition's value.
Supported by data validation, conditional formatting and filters.
Requires a single ConditionValue.}
 \item{NUMBER_BETWEEN - The cell's value must be between the two condition values.
Supported by data validation, conditional formatting and filters.
Requires exactly two ConditionValues.}
 \item{NUMBER_NOT_BETWEEN - The cell's value must not be between the two condition values.
Supported by data validation, conditional formatting and filters.
Requires exactly two ConditionValues.}
 \item{TEXT_CONTAINS - The cell's value must contain the condition's value.
Supported by data validation, conditional formatting and filters.
Requires a single ConditionValue.}
 \item{TEXT_NOT_CONTAINS - The cell's value must not contain the condition's value.
Supported by data validation, conditional formatting and filters.
Requires a single ConditionValue.}
 \item{TEXT_STARTS_WITH - The cell's value must start with the condition's value.
Supported by conditional formatting and filters.
Requires a single ConditionValue.}
 \item{TEXT_ENDS_WITH - The cell's value must end with the condition's value.
Supported by conditional formatting and filters.
Requires a single ConditionValue.}
 \item{TEXT_EQ - The cell's value must be exactly the condition's value.
Supported by data validation, conditional formatting and filters.
Requires a single ConditionValue.}
 \item{TEXT_IS_EMAIL - The cell's value must be a valid email address.
Supported by data validation.
Requires no ConditionValues.}
 \item{TEXT_IS_URL - The cell's value must be a valid URL.
Supported by data validation.
Requires no ConditionValues.}
 \item{DATE_EQ - The cell's value must be the same date as the condition's value.
Supported by data validation, conditional formatting and filters.
Requires a single ConditionValue.}
 \item{DATE_BEFORE - The cell's value must be before the date of the condition's value.
Supported by data validation, conditional formatting and filters.
Requires a single ConditionValue
that may be a relative date.}
 \item{DATE_AFTER - The cell's value must be after the date of the condition's value.
Supported by data validation, conditional formatting and filters.
Requires a single ConditionValue
that may be a relative date.}
 \item{DATE_ON_OR_BEFORE - The cell's value must be on or before the date of the condition's value.
Supported by data validation.
Requires a single ConditionValue
that may be a relative date.}
 \item{DATE_ON_OR_AFTER - The cell's value must be on or after the date of the condition's value.
Supported by data validation.
Requires a single ConditionValue
that may be a relative date.}
 \item{DATE_BETWEEN - The cell's value must be between the dates of the two condition values.
Supported by data validation.
Requires exactly two ConditionValues.}
 \item{DATE_NOT_BETWEEN - The cell's value must be outside the dates of the two condition values.
Supported by data validation.
Requires exactly two ConditionValues.}
 \item{DATE_IS_VALID - The cell's value must be a date.
Supported by data validation.
Requires no ConditionValues.}
 \item{ONE_OF_RANGE - The cell's value must be listed in the grid in condition value's range.
Supported by data validation.
Requires a single ConditionValue,
and the value must be a valid range in A1 notation.}
 \item{ONE_OF_LIST - The cell's value must in the list of condition values.
Supported by data validation.
Supports any number of condition values,
one per item in the list.
Formulas are not supported in the values.}
 \item{BLANK - The cell's value must be empty.
Supported by conditional formatting and filters.
Requires no ConditionValues.}
 \item{NOT_BLANK - The cell's value must not be empty.
Supported by conditional formatting and filters.
Requires no ConditionValues.}
 \item{CUSTOM_FORMULA - The condition's formula must evaluate to TRUE.
Supported by data validation, conditional formatting and filters.
Requires a single ConditionValue.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#BooleanCondition}{Google's Documentation for BooleanCondition}
}

