% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_BooleanRule}
\alias{gsv4_BooleanRule}
\title{gsv4_BooleanRule}
\usage{
gsv4_BooleanRule(condition = NULL, format = NULL)
}
\arguments{
\item{condition}{\code{\link{gsv4_BooleanCondition}} object. A condition that can evaluate to TRUE or FALSE.
BooleanConditions are used by conditional formatting,
data validation, and the criteria in filters.}

\item{format}{\code{\link{gsv4_CellFormat}} object. The format of a cell.}
}
\value{
BooleanRule
}
\description{
A rule that may or may not match, depending on the condition.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#BooleanRule}{Google's Documentation for BooleanRule}
}

