% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_Borders}
\alias{gsv4_Borders}
\title{gsv4_Borders}
\usage{
gsv4_Borders(bottom=NULL, left=NULL, right=NULL, top=NULL)
}
\arguments{
\item{bottom}{\code{\link{gsv4_Border}} object. A border along a cell.}

\item{left}{\code{\link{gsv4_Border}} object. A border along a cell.}

\item{right}{\code{\link{gsv4_Border}} object. A border along a cell.}

\item{top}{\code{\link{gsv4_Border}} object. A border along a cell.}
}
\value{
Borders
}
\description{
The borders of the cell.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#Borders}{Google's Documentation for Borders}
}

