% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_CellData}
\alias{gsv4_CellData}
\title{gsv4_CellData}
\usage{
gsv4_CellData(dataValidation=NULL, effectiveFormat=NULL, effectiveValue=NULL, formattedValue=NULL, hyperlink=NULL, note=NULL, pivotTable=NULL, textFormatRuns=NULL, userEnteredFormat=NULL, userEnteredValue=NULL)
}
\arguments{
\item{dataValidation}{\code{\link{gsv4_DataValidationRule}} object. A data validation rule.}

\item{effectiveFormat}{\code{\link{gsv4_CellFormat}} object. The format of a cell.}

\item{effectiveValue}{\code{\link{gsv4_ExtendedValue}} object. The kinds of value that a cell in a spreadsheet can have.}

\item{formattedValue}{string. The formatted value of the cell.
This is the value as it's shown to the user.
This field is read-only.}

\item{hyperlink}{string. A hyperlink this cell points to, if any.
This field is read-only.  (To set it, use a `=HYPERLINK` formula.)}

\item{note}{string. Any note on the cell.}

\item{pivotTable}{\code{\link{gsv4_PivotTable}} object. A pivot table.}

\item{textFormatRuns}{list of \code{\link{gsv4_TextFormatRun}} objects. Runs of rich text applied to subsections of the cell.  Runs are only valid
on user entered strings, not formulas, bools, or numbers.
Runs start at specific indexes in the text and continue until the next
run. Properties of a run will continue unless explicitly changed
in a subsequent run (and properties of the first run will continue
the properties of the cell unless explicitly changed).

When writing, the new runs will overwrite any prior runs.  When writing a
new user_entered_value, previous runs will be erased.}

\item{userEnteredFormat}{\code{\link{gsv4_CellFormat}} object. The format of a cell.}

\item{userEnteredValue}{\code{\link{gsv4_ExtendedValue}} object. The kinds of value that a cell in a spreadsheet can have.}
}
\value{
CellData
}
\description{
Data about a specific cell.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#CellData}{Google's Documentation for CellData}
}

