% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_ChartData}
\alias{gsv4_ChartData}
\title{gsv4_ChartData}
\usage{
gsv4_ChartData(sourceRange = NULL)
}
\arguments{
\item{sourceRange}{\code{\link{gsv4_ChartSourceRange}} object. Source ranges for a chart.}
}
\value{
ChartData
}
\description{
The data included in a domain or series.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#ChartData}{Google's Documentation for ChartData}
}

