% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_ChartSourceRange}
\alias{gsv4_ChartSourceRange}
\title{gsv4_ChartSourceRange}
\usage{
gsv4_ChartSourceRange(sources=NULL)
}
\arguments{
\item{sources}{list of \code{\link{gsv4_GridRange}} objects. The ranges of data for a series or domain.
Exactly one dimension must have a length of 1,
and all sources in the list must have the same dimension
with length 1.
The domain (if it exists) & all series must have the same number
of source ranges. If using more than one source range, then the source
range at a given offset must be contiguous across the domain and series.

For example, these are valid configurations:

    domain sources: A1:A5
    series1 sources: B1:B5
    series2 sources: D6:D10

    domain sources: A1:A5, C10:C12
    series1 sources: B1:B5, D10:D12
    series2 sources: C1:C5, E10:E12}
}
\value{
ChartSourceRange
}
\description{
Source ranges for a chart.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#ChartSourceRange}{Google's Documentation for ChartSourceRange}
}

