% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_ChartSpec}
\alias{gsv4_ChartSpec}
\title{gsv4_ChartSpec}
\usage{
gsv4_ChartSpec(basicChart = NULL, hiddenDimensionStrategy = NULL,
  pieChart = NULL, title = NULL)
}
\arguments{
\item{basicChart}{\code{\link{gsv4_BasicChartSpec}} object. The specification for a basic chart.  See BasicChartType for the list
of charts this supports.}

\item{hiddenDimensionStrategy}{string. Determines how the charts will use hidden rows or columns. hiddenDimensionStrategy must take one of the following values: CHART_HIDDEN_DIMENSION_STRATEGY_UNSPECIFIED, SKIP_HIDDEN_ROWS_AND_COLUMNS, SKIP_HIDDEN_ROWS, SKIP_HIDDEN_COLUMNS, SHOW_ALL
See the details section for the definition of each of these values.}

\item{pieChart}{\code{\link{gsv4_PieChartSpec}} object. A <a href="/chart/interactive/docs/gallery/piechart">pie chart</a>.}

\item{title}{string. The title of the chart.}
}
\value{
ChartSpec
}
\description{
The specifications of a chart.
}
\details{
hiddenDimensionStrategy takes one of the following values:
\itemize{
 \item{CHART_HIDDEN_DIMENSION_STRATEGY_UNSPECIFIED - Default value, do not use.}
 \item{SKIP_HIDDEN_ROWS_AND_COLUMNS - Charts will skip hidden rows and columns.}
 \item{SKIP_HIDDEN_ROWS - Charts will skip hidden rows only.}
 \item{SKIP_HIDDEN_COLUMNS - Charts will skip hidden columns only.}
 \item{SHOW_ALL - Charts will not skip any hidden rows or columns.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#ChartSpec}{Google's Documentation for ChartSpec}
}

