% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_ClearBasicFilterRequest}
\alias{gsv4_ClearBasicFilterRequest}
\title{gsv4_ClearBasicFilterRequest}
\usage{
gsv4_ClearBasicFilterRequest(sheetId=NULL)
}
\arguments{
\item{sheetId}{integer. The sheet ID on which the basic filter should be cleared.}
}
\value{
ClearBasicFilterRequest
}
\description{
Clears the basic filter, if any exists on the sheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#ClearBasicFilterRequest}{Google's Documentation for ClearBasicFilterRequest}
}

