% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_ClearValuesRequest}
\alias{gsv4_ClearValuesRequest}
\title{gsv4_ClearValuesRequest}
\usage{
gsv4_ClearValuesRequest()
}
\value{
ClearValuesRequest
}
\description{
The request for clearing a range of values in a spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#ClearValuesRequest}{Google's Documentation for ClearValuesRequest}
}

