% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_ConditionalFormatRule}
\alias{gsv4_ConditionalFormatRule}
\title{gsv4_ConditionalFormatRule}
\usage{
gsv4_ConditionalFormatRule(booleanRule = NULL, gradientRule = NULL,
  ranges = NULL)
}
\arguments{
\item{booleanRule}{\code{\link{gsv4_BooleanRule}} object. A rule that may or may not match, depending on the condition.}

\item{gradientRule}{\code{\link{gsv4_GradientRule}} object. A rule that applies a gradient color scale format, based on
the interpolation points listed. The format of a cell will vary
based on its contents as compared to the values of the interpolation
points.}

\item{ranges}{list of \code{\link{gsv4_GridRange}} objects. The ranges that will be formatted if the condition is TRUE.
All the ranges must be on the same grid.}
}
\value{
ConditionalFormatRule
}
\description{
A rule describing a conditional format.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#ConditionalFormatRule}{Google's Documentation for ConditionalFormatRule}
}

