% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_ConditionValue}
\alias{gsv4_ConditionValue}
\title{gsv4_ConditionValue}
\usage{
gsv4_ConditionValue(relativeDate = NULL, userEnteredValue = NULL)
}
\arguments{
\item{relativeDate}{string. A relative date (based on the current date).
Valid only if the type is
DATE_BEFORE,
DATE_AFTER,
DATE_ON_OR_BEFORE or
DATE_ON_OR_AFTER.

Relative dates are not supported in data validation.
They are supported only in conditional formatting and
conditional filters. relativeDate must take one of the following values: RELATIVE_DATE_UNSPECIFIED, PAST_YEAR, PAST_MONTH, PAST_WEEK, YESTERDAY, TODAY, TOMORROW
See the details section for the definition of each of these values.}

\item{userEnteredValue}{string. A value the condition is based on.
The value will be parsed as if the user typed into a cell.
Formulas are supported (and must begin with an `=`).}
}
\value{
ConditionValue
}
\description{
The value of the condition.
}
\details{
relativeDate takes one of the following values:
\itemize{
 \item{RELATIVE_DATE_UNSPECIFIED - Default value, do not use.}
 \item{PAST_YEAR - The value is one year before today.}
 \item{PAST_MONTH - The value is one month before today.}
 \item{PAST_WEEK - The value is one week before today.}
 \item{YESTERDAY - The value is yesterday.}
 \item{TODAY - The value is today.}
 \item{TOMORROW - The value is tomorrow.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#ConditionValue}{Google's Documentation for ConditionValue}
}

