% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_CopyPasteRequest}
\alias{gsv4_CopyPasteRequest}
\title{gsv4_CopyPasteRequest}
\usage{
gsv4_CopyPasteRequest(destination = NULL, pasteOrientation = NULL,
  pasteType = NULL, source = NULL)
}
\arguments{
\item{destination}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}

\item{pasteOrientation}{string. How that data should be oriented when pasting. pasteOrientation must take one of the following values: NORMAL, TRANSPOSE
See the details section for the definition of each of these values.}

\item{pasteType}{string. What kind of data to paste. pasteType must take one of the following values: PASTE_NORMAL, PASTE_VALUES, PASTE_FORMAT, PASTE_NO_BORDERS, PASTE_FORMULA, PASTE_DATA_VALIDATION, PASTE_CONDITIONAL_FORMATTING
See the details section for the definition of each of these values.}

\item{source}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}
}
\value{
CopyPasteRequest
}
\description{
Copies data from the source to the destination.
}
\details{
pasteOrientation takes one of the following values:
\itemize{
 \item{NORMAL - Paste normally.}
 \item{TRANSPOSE - Paste transposed, where all rows become columns and vice versa.}
}

pasteType takes one of the following values:
\itemize{
 \item{PASTE_NORMAL - Paste values, formulas, formats, and merges.}
 \item{PASTE_VALUES - Paste the values ONLY without formats, formulas, or merges.}
 \item{PASTE_FORMAT - Paste the format and data validation only.}
 \item{PASTE_NO_BORDERS - Like PASTE_NORMAL but without borders.}
 \item{PASTE_FORMULA - Paste the formulas only.}
 \item{PASTE_DATA_VALIDATION - Paste the data validation only.}
 \item{PASTE_CONDITIONAL_FORMATTING - Paste the conditional formatting rules only.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#CopyPasteRequest}{Google's Documentation for CopyPasteRequest}
}

