% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_CopySheetToAnotherSpreadsheetRequest}
\alias{gsv4_CopySheetToAnotherSpreadsheetRequest}
\title{gsv4_CopySheetToAnotherSpreadsheetRequest}
\usage{
gsv4_CopySheetToAnotherSpreadsheetRequest(destinationSpreadsheetId=NULL)
}
\arguments{
\item{destinationSpreadsheetId}{string. The ID of the spreadsheet to copy the sheet to.}
}
\value{
CopySheetToAnotherSpreadsheetRequest
}
\description{
The request to copy a sheet across spreadsheets.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#CopySheetToAnotherSpreadsheetRequest}{Google's Documentation for CopySheetToAnotherSpreadsheetRequest}
}

