% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheets.R
\name{gsv4_create}
\alias{gsv4_create}
\title{gsv4_create}
\usage{
gsv4_create(input, standard_params = list(...), ...)
}
\arguments{
\item{input}{\code{\link{gsv4_Spreadsheet}}. Resource that represents a spreadsheet.}

\item{standard_params}{a list of parameters for controlling the HTTP request and its response.
Refer to \code{\link{gsv4_standard_parameters}} for details on its arguments}

\item{...}{arguments to be used to form standard_params argument if it is not supplied directly.}
}
\value{
Spreadsheet
}
\description{
Creates a spreadsheet, returning the newly created spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets/create}{Google's Documentation for create}
}

