% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DataValidationRule}
\alias{gsv4_DataValidationRule}
\title{gsv4_DataValidationRule}
\usage{
gsv4_DataValidationRule(condition = NULL, inputMessage = NULL,
  showCustomUi = NULL, strict = NULL)
}
\arguments{
\item{condition}{\code{\link{gsv4_BooleanCondition}} object. A condition that can evaluate to TRUE or FALSE.
BooleanConditions are used by conditional formatting,
data validation, and the criteria in filters.}

\item{inputMessage}{string. A message to show the user when adding data to the cell.}

\item{showCustomUi}{logical. TRUE if the UI should be customized based on the kind of condition.
If TRUE, "List" conditions will show a dropdown.}

\item{strict}{logical. TRUE if invalid data should be rejected.}
}
\value{
DataValidationRule
}
\description{
A data validation rule.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DataValidationRule}{Google's Documentation for DataValidationRule}
}

