% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DeleteBandingRequest}
\alias{gsv4_DeleteBandingRequest}
\title{gsv4_DeleteBandingRequest}
\usage{
gsv4_DeleteBandingRequest(bandedRangeId=NULL)
}
\arguments{
\item{bandedRangeId}{integer. The ID of the banded range to delete.}
}
\value{
DeleteBandingRequest
}
\description{
Removes the banded range with the given ID from the spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DeleteBandingRequest}{Google's Documentation for DeleteBandingRequest}
}

