% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DeleteConditionalFormatRuleRequest}
\alias{gsv4_DeleteConditionalFormatRuleRequest}
\title{gsv4_DeleteConditionalFormatRuleRequest}
\usage{
gsv4_DeleteConditionalFormatRuleRequest(sheetId = NULL, index = NULL)
}
\arguments{
\item{sheetId}{integer. The sheet the rule is being deleted from.}

\item{index}{integer. The zero-based index of the rule to be deleted.}
}
\value{
DeleteConditionalFormatRuleRequest
}
\description{
Deletes a conditional format rule at the given index.
All subsequent rules' indexes are decremented.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DeleteConditionalFormatRuleRequest}{Google's Documentation for DeleteConditionalFormatRuleRequest}
}

