% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DeleteDimensionRequest}
\alias{gsv4_DeleteDimensionRequest}
\title{gsv4_DeleteDimensionRequest}
\usage{
gsv4_DeleteDimensionRequest(range=NULL)
}
\arguments{
\item{range}{\code{\link{gsv4_DimensionRange}} object. A range along a single dimension on a sheet.
All indexes are zero-based.
Indexes are half open: the start index is inclusive
and the end index is exclusive.
Missing indexes indicate the range is unbounded on that side.}
}
\value{
DeleteDimensionRequest
}
\description{
Deletes the dimensions from the sheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DeleteDimensionRequest}{Google's Documentation for DeleteDimensionRequest}
}

