% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DeleteEmbeddedObjectRequest}
\alias{gsv4_DeleteEmbeddedObjectRequest}
\title{gsv4_DeleteEmbeddedObjectRequest}
\usage{
gsv4_DeleteEmbeddedObjectRequest(objectId=NULL)
}
\arguments{
\item{objectId}{integer. The ID of the embedded object to delete.}
}
\value{
DeleteEmbeddedObjectRequest
}
\description{
Deletes the embedded object with the given ID.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DeleteEmbeddedObjectRequest}{Google's Documentation for DeleteEmbeddedObjectRequest}
}

