% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DeleteFilterViewRequest}
\alias{gsv4_DeleteFilterViewRequest}
\title{gsv4_DeleteFilterViewRequest}
\usage{
gsv4_DeleteFilterViewRequest(filterId=NULL)
}
\arguments{
\item{filterId}{integer. The ID of the filter to delete.}
}
\value{
DeleteFilterViewRequest
}
\description{
Deletes a particular filter view.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DeleteFilterViewRequest}{Google's Documentation for DeleteFilterViewRequest}
}

