% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DeleteNamedRangeRequest}
\alias{gsv4_DeleteNamedRangeRequest}
\title{gsv4_DeleteNamedRangeRequest}
\usage{
gsv4_DeleteNamedRangeRequest(namedRangeId=NULL)
}
\arguments{
\item{namedRangeId}{string. The ID of the named range to delete.}
}
\value{
DeleteNamedRangeRequest
}
\description{
Removes the named range with the given ID from the spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DeleteNamedRangeRequest}{Google's Documentation for DeleteNamedRangeRequest}
}

