% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DeleteProtectedRangeRequest}
\alias{gsv4_DeleteProtectedRangeRequest}
\title{gsv4_DeleteProtectedRangeRequest}
\usage{
gsv4_DeleteProtectedRangeRequest(protectedRangeId = NULL)
}
\arguments{
\item{protectedRangeId}{integer. The ID of the protected range to delete.}
}
\value{
DeleteProtectedRangeRequest
}
\description{
Deletes the protected range with the given ID.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DeleteProtectedRangeRequest}{Google's Documentation for DeleteProtectedRangeRequest}
}

