% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DeleteRangeRequest}
\alias{gsv4_DeleteRangeRequest}
\title{gsv4_DeleteRangeRequest}
\usage{
gsv4_DeleteRangeRequest(range=NULL, shiftDimension=NULL)
}
\arguments{
\item{range}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}

\item{shiftDimension}{string. The dimension from which deleted cells will be replaced with.
If ROWS, existing cells will be shifted upward to
replace the deleted cells. If COLUMNS, existing cells
will be shifted left to replace the deleted cells. shiftDimension must take one of the following values: DIMENSION_UNSPECIFIED, ROWS, COLUMNS
See the details section for the definition of each of these values.}
}
\value{
DeleteRangeRequest
}
\description{
Deletes a range of cells, shifting other cells into the deleted area.
}
\details{
shiftDimension takes one of the following values:
\itemize{
 \item{DIMENSION_UNSPECIFIED - The default value, do not use.}
 \item{ROWS - Operates on the rows of a sheet.}
 \item{COLUMNS - Operates on the columns of a sheet.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DeleteRangeRequest}{Google's Documentation for DeleteRangeRequest}
}

