% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DeleteSheetRequest}
\alias{gsv4_DeleteSheetRequest}
\title{gsv4_DeleteSheetRequest}
\usage{
gsv4_DeleteSheetRequest(sheetId=NULL)
}
\arguments{
\item{sheetId}{integer. The ID of the sheet to delete.}
}
\value{
DeleteSheetRequest
}
\description{
Deletes the requested sheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DeleteSheetRequest}{Google's Documentation for DeleteSheetRequest}
}

