% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DimensionProperties}
\alias{gsv4_DimensionProperties}
\title{gsv4_DimensionProperties}
\usage{
gsv4_DimensionProperties(hiddenByFilter=NULL, hiddenByUser=NULL, pixelSize=NULL)
}
\arguments{
\item{hiddenByFilter}{logical. TRUE if this dimension is being filtered.
This field is read-only.}

\item{hiddenByUser}{logical. TRUE if this dimension is explicitly hidden.}

\item{pixelSize}{integer. The height (if a row) or width (if a column) of the dimension in pixels.}
}
\value{
DimensionProperties
}
\description{
Properties about a dimension.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DimensionProperties}{Google's Documentation for DimensionProperties}
}

