% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DimensionRange}
\alias{gsv4_DimensionRange}
\title{gsv4_DimensionRange}
\usage{
gsv4_DimensionRange(sheetId = NULL, dimension = NULL, endIndex = NULL,
  startIndex = NULL)
}
\arguments{
\item{sheetId}{integer. The sheet this span is on.}

\item{dimension}{string. The dimension of the span. dimension must take one of the following values: DIMENSION_UNSPECIFIED, ROWS, COLUMNS
See the details section for the definition of each of these values.}

\item{endIndex}{integer. The end (exclusive) of the span, or not set if unbounded.}

\item{startIndex}{integer. The start (inclusive) of the span, or not set if unbounded.}
}
\value{
DimensionRange
}
\description{
A range along a single dimension on a sheet.
All indexes are zero-based.
Indexes are half open: the start index is inclusive
and the end index is exclusive.
Missing indexes indicate the range is unbounded on that side.
}
\details{
dimension takes one of the following values:
\itemize{
 \item{DIMENSION_UNSPECIFIED - The default value, do not use.}
 \item{ROWS - Operates on the rows of a sheet.}
 \item{COLUMNS - Operates on the columns of a sheet.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DimensionRange}{Google's Documentation for DimensionRange}
}

