% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DuplicateFilterViewRequest}
\alias{gsv4_DuplicateFilterViewRequest}
\title{gsv4_DuplicateFilterViewRequest}
\usage{
gsv4_DuplicateFilterViewRequest(filterId = NULL)
}
\arguments{
\item{filterId}{integer. The ID of the filter being duplicated.}
}
\value{
DuplicateFilterViewRequest
}
\description{
Duplicates a particular filter view.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DuplicateFilterViewRequest}{Google's Documentation for DuplicateFilterViewRequest}
}

