% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_DuplicateSheetRequest}
\alias{gsv4_DuplicateSheetRequest}
\title{gsv4_DuplicateSheetRequest}
\usage{
gsv4_DuplicateSheetRequest(insertSheetIndex=NULL, newSheetId=NULL, newSheetName=NULL, sourceSheetId=NULL)
}
\arguments{
\item{insertSheetIndex}{integer. The zero-based index where the new sheet should be inserted.
The index of all sheets after this are incremented.}

\item{newSheetId}{integer. If set, the ID of the new sheet. If not set, an ID is chosen.
If set, the ID must not conflict with any existing sheet ID.
If set, it must be non-negative.}

\item{newSheetName}{string. The name of the new sheet.  If empty, a new name is chosen for you.}

\item{sourceSheetId}{integer. The sheet to duplicate.}
}
\value{
DuplicateSheetRequest
}
\description{
Duplicates the contents of a sheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#DuplicateSheetRequest}{Google's Documentation for DuplicateSheetRequest}
}

