% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_Editors}
\alias{gsv4_Editors}
\title{gsv4_Editors}
\usage{
gsv4_Editors(domainUsersCanEdit = NULL, groups = NULL, users = NULL)
}
\arguments{
\item{domainUsersCanEdit}{logical. TRUE if anyone in the document's domain has edit access to the protected
range.  Domain protection is only supported on documents within a domain.}

\item{groups}{list. The email addresses of groups with edit access to the protected range.}

\item{users}{list. The email addresses of users with edit access to the protected range.}
}
\value{
Editors
}
\description{
The editors of a protected range.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#Editors}{Google's Documentation for Editors}
}

