% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_EmbeddedChart}
\alias{gsv4_EmbeddedChart}
\title{gsv4_EmbeddedChart}
\usage{
gsv4_EmbeddedChart(chartId=NULL, position=NULL, spec=NULL)
}
\arguments{
\item{chartId}{integer. The ID of the chart.}

\item{position}{\code{\link{gsv4_EmbeddedObjectPosition}} object. The position of an embedded object such as a chart.}

\item{spec}{\code{\link{gsv4_ChartSpec}} object. The specifications of a chart.}
}
\value{
EmbeddedChart
}
\description{
A chart embedded in a sheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#EmbeddedChart}{Google's Documentation for EmbeddedChart}
}

