% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_EmbeddedObjectPosition}
\alias{gsv4_EmbeddedObjectPosition}
\title{gsv4_EmbeddedObjectPosition}
\usage{
gsv4_EmbeddedObjectPosition(sheetId=NULL, newSheet=NULL, overlayPosition=NULL)
}
\arguments{
\item{sheetId}{integer. The sheet this is on. Set only if the embedded object
is on its own sheet. Must be non-negative.}

\item{newSheet}{logical. If TRUE, the embedded object will be put on a new sheet whose ID
is chosen for you. Used only when writing.}

\item{overlayPosition}{\code{\link{gsv4_OverlayPosition}} object. The location an object is overlaid on top of a grid.}
}
\value{
EmbeddedObjectPosition
}
\description{
The position of an embedded object such as a chart.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#EmbeddedObjectPosition}{Google's Documentation for EmbeddedObjectPosition}
}

