% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_ErrorValue}
\alias{gsv4_ErrorValue}
\title{gsv4_ErrorValue}
\usage{
gsv4_ErrorValue(message = NULL, type = NULL)
}
\arguments{
\item{message}{string. A message with more information about the error
(in the spreadsheet's locale).}

\item{type}{string. The type of error. type must take one of the following values: ERROR_TYPE_UNSPECIFIED, ERROR, NULL_VALUE, DIVIDE_BY_ZERO, VALUE, REF, NAME, NUM, N_A, LOADING
See the details section for the definition of each of these values.}
}
\value{
ErrorValue
}
\description{
An error in a cell.
}
\details{
type takes one of the following values:
\itemize{
 \item{ERROR_TYPE_UNSPECIFIED - The default error type, do not use this.}
 \item{ERROR - Corresponds to the `#ERROR!` error.}
 \item{NULL_VALUE - Corresponds to the `#NULL!` error.}
 \item{DIVIDE_BY_ZERO - Corresponds to the `#DIV/0` error.}
 \item{VALUE - Corresponds to the `#VALUE!` error.}
 \item{REF - Corresponds to the `#REF!` error.}
 \item{NAME - Corresponds to the `#NAME?` error.}
 \item{NUM - Corresponds to the `#NUM!` error.}
 \item{N_A - Corresponds to the `#N/A` error.}
 \item{LOADING - Corresponds to the `Loading...` state.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#ErrorValue}{Google's Documentation for ErrorValue}
}

