% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_ExtendedValue}
\alias{gsv4_ExtendedValue}
\title{gsv4_ExtendedValue}
\usage{
gsv4_ExtendedValue(boolValue=NULL, errorValue=NULL, formulaValue=NULL, numberValue=NULL, stringValue=NULL)
}
\arguments{
\item{boolValue}{logical. Represents a boolean value.}

\item{errorValue}{\code{\link{gsv4_ErrorValue}} object. An error in a cell.}

\item{formulaValue}{string. Represents a formula.}

\item{numberValue}{numeric. Represents a double value.
Note: Dates, Times and DateTimes are represented as doubles in
"serial number" format.}

\item{stringValue}{string. Represents a string value.
Leading single quotes are not included. For example, if the user typed
`'123` into the UI, this would be represented as a `stringValue` of
`"123"`.}
}
\value{
ExtendedValue
}
\description{
The kinds of value that a cell in a spreadsheet can have.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#ExtendedValue}{Google's Documentation for ExtendedValue}
}

