% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_FilterCriteria}
\alias{gsv4_FilterCriteria}
\title{gsv4_FilterCriteria}
\usage{
gsv4_FilterCriteria(condition=NULL, hiddenValues=NULL)
}
\arguments{
\item{condition}{\code{\link{gsv4_BooleanCondition}} object. A condition that can evaluate to TRUE or FALSE.
BooleanConditions are used by conditional formatting,
data validation, and the criteria in filters.}

\item{hiddenValues}{list. Values that should be hidden.}
}
\value{
FilterCriteria
}
\description{
Criteria for showing/hiding rows in a filter or filter view.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#FilterCriteria}{Google's Documentation for FilterCriteria}
}

