% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_FindReplaceRequest}
\alias{gsv4_FindReplaceRequest}
\title{gsv4_FindReplaceRequest}
\usage{
gsv4_FindReplaceRequest(sheetId = NULL, range = NULL, allSheets = NULL,
  find = NULL, includeFormulas = NULL, matchCase = NULL,
  matchEntireCell = NULL, replacement = NULL, searchByRegex = NULL)
}
\arguments{
\item{sheetId}{integer. The sheet to find/replace over.}

\item{range}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}

\item{allSheets}{logical. TRUE to find/replace over all sheets.}

\item{find}{string. The value to search.}

\item{includeFormulas}{logical. TRUE if the search should include cells with formulas.
FALSE to skip cells with formulas.}

\item{matchCase}{logical. TRUE if the search is case sensitive.}

\item{matchEntireCell}{logical. TRUE if the find value should match the entire cell.}

\item{replacement}{string. The value to use as the replacement.}

\item{searchByRegex}{logical. TRUE if the find value is a regex.
The regular expression and replacement should follow Java regex rules
at https://docs.oracle.com/javase/8/docs/api/java/util/regex/Pattern.html.
The replacement string is allowed to refer to capturing groups.
For example, if one cell has the contents `"Google Sheets"` and another
has `"Google Docs"`, then searching for `"o.* (.*)"` with a replacement of
`"$1 Rocks"` would change the contents of the cells to
`"GSheets Rocks"` and `"GDocs Rocks"` respectively.}
}
\value{
FindReplaceRequest
}
\description{
Finds and replaces data in cells over a range, sheet, or all sheets.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#FindReplaceRequest}{Google's Documentation for FindReplaceRequest}
}

