% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gsv4_form_query_string}
\alias{gsv4_form_query_string}
\title{gsv4_form_query_string}
\usage{
gsv4_form_query_string(standard_params = list(), ...)
}
\arguments{
\item{standard_params}{a list of parameters for controlling the HTTP request and its response.
Refer to \code{\link{gsv4_standard_parameters}} for details on its arguments}

\item{...}{arguments to be used to form standard_params argument if it is not supplied directly.}
}
\value{
character; a string that represents query parameters of a URL
}
\description{
A convenience function for constructing a query string to tack onto URL
based on a named list of arguments. The function assumes the standard
query parameters as defined by the Sheets API v4.
}
\examples{
\dontrun{
gsv4_form_query_string(standard_params=list(fields='sheets.properties'), other='o')
gsv4_form_query_string(a=1, b=TRUE, x='x')
gsv4_form_query_string()
}
}
\seealso{
\href{https://developers.google.com/sheets/api/query-parameters}{Google's Documentation of Standard Query Parameters}
}

