% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheets.R
\name{gsv4_get}
\alias{gsv4_get}
\title{gsv4_get}
\usage{
gsv4_get(spreadsheetId, includeGridData = NULL, ranges = NULL,
  standard_params = list(...), ...)
}
\arguments{
\item{spreadsheetId}{string (required). The spreadsheet to request.}

\item{includeGridData}{logical. TRUE if grid data should be returned.
This parameter is ignored if a field mask was set in the request.}

\item{ranges}{string (repeated). The ranges to retrieve from the spreadsheet.}

\item{standard_params}{a list of parameters for controlling the HTTP request and its response.
Refer to \code{\link{gsv4_standard_parameters}} for details on its arguments}

\item{...}{arguments to be used to form standard_params argument if it is not supplied directly.}
}
\value{
Spreadsheet
}
\description{
Returns the spreadsheet at the given ID.
The caller must specify the spreadsheet ID.
}
\details{
By default, data within grids will not be returned.
You can include grid data one of two ways:

* Specify a field mask listing your desired fields using the `fields` URL
parameter in HTTP

* Set the includeGridData
URL parameter to TRUE.  If a field mask is set, the `includeGridData`
parameter is ignored

For large spreadsheets, it is recommended to retrieve only the specific
fields of the spreadsheet that you want.

To retrieve only subsets of the spreadsheet, use the
ranges URL parameter.
Multiple ranges can be specified.  Limiting the range will
return only the portions of the spreadsheet that intersect the requested
ranges. Ranges are specified using A1 notation.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets/get}{Google's Documentation for get}
}

