% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_GradientRule}
\alias{gsv4_GradientRule}
\title{gsv4_GradientRule}
\usage{
gsv4_GradientRule(maxpoint=NULL, midpoint=NULL, minpoint=NULL)
}
\arguments{
\item{maxpoint}{\code{\link{gsv4_InterpolationPoint}} object. A single interpolation point on a gradient conditional format.
These pin the gradient color scale according to the color,
type and value chosen.}

\item{midpoint}{\code{\link{gsv4_InterpolationPoint}} object. A single interpolation point on a gradient conditional format.
These pin the gradient color scale according to the color,
type and value chosen.}

\item{minpoint}{\code{\link{gsv4_InterpolationPoint}} object. A single interpolation point on a gradient conditional format.
These pin the gradient color scale according to the color,
type and value chosen.}
}
\value{
GradientRule
}
\description{
A rule that applies a gradient color scale format, based on
the interpolation points listed. The format of a cell will vary
based on its contents as compared to the values of the interpolation
points.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#GradientRule}{Google's Documentation for GradientRule}
}

