% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_GridCoordinate}
\alias{gsv4_GridCoordinate}
\title{gsv4_GridCoordinate}
\usage{
gsv4_GridCoordinate(sheetId = NULL, columnIndex = NULL, rowIndex = NULL)
}
\arguments{
\item{sheetId}{integer. The sheet this coordinate is on.}

\item{columnIndex}{integer. The column index of the coordinate.}

\item{rowIndex}{integer. The row index of the coordinate.}
}
\value{
GridCoordinate
}
\description{
A coordinate in a sheet.
All indexes are zero-based.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#GridCoordinate}{Google's Documentation for GridCoordinate}
}

