% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_GridData}
\alias{gsv4_GridData}
\title{gsv4_GridData}
\usage{
gsv4_GridData(columnMetadata = NULL, rowData = NULL, rowMetadata = NULL,
  startColumn = NULL, startRow = NULL)
}
\arguments{
\item{columnMetadata}{list of \code{\link{gsv4_DimensionProperties}} objects. Metadata about the requested columns in the grid, starting with the column
in start_column.}

\item{rowData}{list of \code{\link{gsv4_RowData}} objects. The data in the grid, one entry per row,
starting with the row in startRow.
The values in RowData will correspond to columns starting
at start_column.}

\item{rowMetadata}{list of \code{\link{gsv4_DimensionProperties}} objects. Metadata about the requested rows in the grid, starting with the row
in start_row.}

\item{startColumn}{integer. The first column this GridData refers to, zero-based.}

\item{startRow}{integer. The first row this GridData refers to, zero-based.}
}
\value{
GridData
}
\description{
Data in the grid, as well as metadata about the dimensions.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#GridData}{Google's Documentation for GridData}
}

