% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_GridProperties}
\alias{gsv4_GridProperties}
\title{gsv4_GridProperties}
\usage{
gsv4_GridProperties(columnCount = NULL, frozenColumnCount = NULL,
  frozenRowCount = NULL, hideGridlines = NULL, rowCount = NULL)
}
\arguments{
\item{columnCount}{integer. The number of columns in the grid.}

\item{frozenColumnCount}{integer. The number of columns that are frozen in the grid.}

\item{frozenRowCount}{integer. The number of rows that are frozen in the grid.}

\item{hideGridlines}{logical. TRUE if the grid isn't showing gridlines in the UI.}

\item{rowCount}{integer. The number of rows in the grid.}
}
\value{
GridProperties
}
\description{
Properties of a grid.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#GridProperties}{Google's Documentation for GridProperties}
}

