% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_GridRange}
\alias{gsv4_GridRange}
\title{gsv4_GridRange}
\usage{
gsv4_GridRange(sheetId = NULL, endColumnIndex = NULL, endRowIndex = NULL,
  startColumnIndex = NULL, startRowIndex = NULL)
}
\arguments{
\item{sheetId}{integer. The sheet this range is on.}

\item{endColumnIndex}{integer. The end column (exclusive) of the range, or not set if unbounded.}

\item{endRowIndex}{integer. The end row (exclusive) of the range, or not set if unbounded.}

\item{startColumnIndex}{integer. The start column (inclusive) of the range, or not set if unbounded.}

\item{startRowIndex}{integer. The start row (inclusive) of the range, or not set if unbounded.}
}
\value{
GridRange
}
\description{
A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.
}
\details{
For example, if `"Sheet1"` is sheet ID 0, then:

`Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

`Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

`Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

`Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

`Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#GridRange}{Google's Documentation for GridRange}
}

