% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_InsertDimensionRequest}
\alias{gsv4_InsertDimensionRequest}
\title{gsv4_InsertDimensionRequest}
\usage{
gsv4_InsertDimensionRequest(range = NULL, inheritFromBefore = NULL)
}
\arguments{
\item{range}{\code{\link{gsv4_DimensionRange}} object. A range along a single dimension on a sheet.
All indexes are zero-based.
Indexes are half open: the start index is inclusive
and the end index is exclusive.
Missing indexes indicate the range is unbounded on that side.}

\item{inheritFromBefore}{logical. Whether dimension properties should be extended from the dimensions
before or after the newly inserted dimensions.
TRUE to inherit from the dimensions before (in which case the start
index must be greater than 0), and FALSE to inherit from the dimensions
after.

For example, if row index 0 has red background and row index 1
has a green background, then inserting 2 rows at index 1 can inherit
either the green or red background.  If `inheritFromBefore` is TRUE,
the two new rows will be red (because the row before the insertion point
was red), whereas if `inheritFromBefore` is FALSE, the two new rows will
be green (because the row after the insertion point was green).}
}
\value{
InsertDimensionRequest
}
\description{
Inserts rows or columns in a sheet at a particular index.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#InsertDimensionRequest}{Google's Documentation for InsertDimensionRequest}
}

