% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_IterativeCalculationSettings}
\alias{gsv4_IterativeCalculationSettings}
\title{gsv4_IterativeCalculationSettings}
\usage{
gsv4_IterativeCalculationSettings(convergenceThreshold = NULL,
  maxIterations = NULL)
}
\arguments{
\item{convergenceThreshold}{numeric. When iterative calculation is enabled and successive results differ by
less than this threshold value, the calculation rounds stop.}

\item{maxIterations}{integer. When iterative calculation is enabled, the maximum number of calculation
rounds to perform.}
}
\value{
IterativeCalculationSettings
}
\description{
Settings to control how circular dependencies are resolved with iterative
calculation.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#IterativeCalculationSettings}{Google's Documentation for IterativeCalculationSettings}
}

