% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gsv4_limits_to_grid_range}
\alias{gsv4_limits_to_grid_range}
\title{gsv4_limits_to_grid_range}
\usage{
gsv4_limits_to_grid_range(lim, ss = NULL)
}
\arguments{
\item{lim}{\code{cell_limits}; a list of three components representing upper-left, 
lower-right limits of the range, plus the sheet component}

\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}
}
\description{
Converting a \code{cell_limits} object to a gridRange object as specified in the Sheets V4 API.
}
\details{
The sheet component of the limits must be an integer that refers to 
a specific sheetId that exists in the spreadsheet.
}
\examples{
\dontrun{
gsv4_limits_to_grid_range(lim=cell_limits(c(1, 3), c(1, 5)))
}

}

