% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_MergeCellsRequest}
\alias{gsv4_MergeCellsRequest}
\title{gsv4_MergeCellsRequest}
\usage{
gsv4_MergeCellsRequest(range=NULL, mergeType=NULL)
}
\arguments{
\item{range}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}

\item{mergeType}{string. How the cells should be merged. mergeType must take one of the following values: MERGE_ALL, MERGE_COLUMNS, MERGE_ROWS
See the details section for the definition of each of these values.}
}
\value{
MergeCellsRequest
}
\description{
Merges all cells in the range.
}
\details{
mergeType takes one of the following values:
\itemize{
 \item{MERGE_ALL - Create a single merge from the range}
 \item{MERGE_COLUMNS - Create a merge for each column in the range}
 \item{MERGE_ROWS - Create a merge for each row in the range}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#MergeCellsRequest}{Google's Documentation for MergeCellsRequest}
}

