% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_NumberFormat}
\alias{gsv4_NumberFormat}
\title{gsv4_NumberFormat}
\usage{
gsv4_NumberFormat(pattern = NULL, type = NULL)
}
\arguments{
\item{pattern}{string. Pattern string used for formatting.  If not set, a default pattern based on
the user's locale will be used if necessary for the given type.
See the \href{https://developers.google.com/sheets/api/guides/formats}{Date and Number Formats guide} for more
information about the supported patterns.}

\item{type}{string. The type of the number format.
When writing, this field must be set. type must take one of the following values: NUMBER_FORMAT_TYPE_UNSPECIFIED, TEXT, NUMBER, PERCENT, CURRENCY, DATE, TIME, DATE_TIME, SCIENTIFIC
See the details section for the definition of each of these values.}
}
\value{
NumberFormat
}
\description{
The number format of a cell.
}
\details{
type takes one of the following values:
\itemize{
 \item{NUMBER_FORMAT_TYPE_UNSPECIFIED - The number format is not specified
and is based on the contents of the cell.
Do not explicitly use this.}
 \item{TEXT - Text formatting, e.g `1000.12`}
 \item{NUMBER - Number formatting, e.g, `1,000.12`}
 \item{PERCENT - Percent formatting, e.g `10.12\%`}
 \item{CURRENCY - Currency formatting, e.g `$1,000.12`}
 \item{DATE - Date formatting, e.g `9/26/2008`}
 \item{TIME - Time formatting, e.g `3:59:00 PM`}
 \item{DATE_TIME - Date+Time formatting, e.g `9/26/08 15:59:00`}
 \item{SCIENTIFIC - Scientific number formatting, e.g `1.01E+03`}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#NumberFormat}{Google's Documentation for NumberFormat}
}

