% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_OverlayPosition}
\alias{gsv4_OverlayPosition}
\title{gsv4_OverlayPosition}
\usage{
gsv4_OverlayPosition(anchorCell=NULL, heightPixels=NULL, offsetXPixels=NULL, offsetYPixels=NULL, widthPixels=NULL)
}
\arguments{
\item{anchorCell}{\code{\link{gsv4_GridCoordinate}} object. A coordinate in a sheet.
All indexes are zero-based.}

\item{heightPixels}{integer. The height of the object, in pixels. Defaults to 371.}

\item{offsetXPixels}{integer. The horizontal offset, in pixels, that the object is offset
from the anchor cell.}

\item{offsetYPixels}{integer. The vertical offset, in pixels, that the object is offset
from the anchor cell.}

\item{widthPixels}{integer. The width of the object, in pixels. Defaults to 600.}
}
\value{
OverlayPosition
}
\description{
The location an object is overlaid on top of a grid.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#OverlayPosition}{Google's Documentation for OverlayPosition}
}

