% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_Padding}
\alias{gsv4_Padding}
\title{gsv4_Padding}
\usage{
gsv4_Padding(bottom = NULL, left = NULL, right = NULL, top = NULL)
}
\arguments{
\item{bottom}{integer. The bottom padding of the cell.}

\item{left}{integer. The left padding of the cell.}

\item{right}{integer. The right padding of the cell.}

\item{top}{integer. The top padding of the cell.}
}
\value{
Padding
}
\description{
The amount of padding around the cell, in pixels.
When updating padding, every field must be specified.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#Padding}{Google's Documentation for Padding}
}

