% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gsv4_parse_values}
\alias{gsv4_parse_values}
\title{gs_parse_values}
\usage{
gsv4_parse_values(values, col_names=TRUE)
}
\arguments{
\item{values}{\code{list}; a list parsed from the Sheets V4 API that represents an array}

\item{col_names}{logical; indicates whether column names of input should be parsed as part of the returne values matrix}
}
\value{
\code{data.frame} parsed from a values matrix returned by the Sheets V4 API
}
\description{
Parsing the values portion of a reply into a data.frame
}
\examples{
\dontrun{
reply <- gsv4_values_get(spreadsheetId = this_spreadsheetId, range="Africa!A3:C5")
my_dat <- gsv4_parse_values(reply$values, col_names=FALSE)
head(my_dat)
}

}

