% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_PasteDataRequest}
\alias{gsv4_PasteDataRequest}
\title{gsv4_PasteDataRequest}
\usage{
gsv4_PasteDataRequest(coordinate=NULL, data=NULL, delimiter=NULL, html=NULL, type=NULL)
}
\arguments{
\item{coordinate}{\code{\link{gsv4_GridCoordinate}} object. A coordinate in a sheet.
All indexes are zero-based.}

\item{data}{string. The data to insert.}

\item{delimiter}{string. The delimiter in the data.}

\item{html}{logical. TRUE if the data is HTML.}

\item{type}{string. How the data should be pasted. type must take one of the following values: PASTE_NORMAL, PASTE_VALUES, PASTE_FORMAT, PASTE_NO_BORDERS, PASTE_FORMULA, PASTE_DATA_VALIDATION, PASTE_CONDITIONAL_FORMATTING
See the details section for the definition of each of these values.}
}
\value{
PasteDataRequest
}
\description{
Inserts data into the spreadsheet starting at the specified coordinate.
}
\details{
type takes one of the following values:
\itemize{
 \item{PASTE_NORMAL - Paste values, formulas, formats, and merges.}
 \item{PASTE_VALUES - Paste the values ONLY without formats, formulas, or merges.}
 \item{PASTE_FORMAT - Paste the format and data validation only.}
 \item{PASTE_NO_BORDERS - Like PASTE_NORMAL but without borders.}
 \item{PASTE_FORMULA - Paste the formulas only.}
 \item{PASTE_DATA_VALIDATION - Paste the data validation only.}
 \item{PASTE_CONDITIONAL_FORMATTING - Paste the conditional formatting rules only.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#PasteDataRequest}{Google's Documentation for PasteDataRequest}
}

