% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_PieChartSpec}
\alias{gsv4_PieChartSpec}
\title{gsv4_PieChartSpec}
\usage{
gsv4_PieChartSpec(domain = NULL, legendPosition = NULL, pieHole = NULL,
  series = NULL, threeDimensional = NULL)
}
\arguments{
\item{domain}{\code{\link{gsv4_ChartData}} object. The data included in a domain or series.}

\item{legendPosition}{string. Where the legend of the pie chart should be drawn. legendPosition must take one of the following values: PIE_CHART_LEGEND_POSITION_UNSPECIFIED, BOTTOM_LEGEND, LEFT_LEGEND, RIGHT_LEGEND, TOP_LEGEND, NO_LEGEND, LABELED_LEGEND
See the details section for the definition of each of these values.}

\item{pieHole}{numeric. The size of the hole in the pie chart.}

\item{series}{\code{\link{gsv4_ChartData}} object. The data included in a domain or series.}

\item{threeDimensional}{logical. TRUE if the pie is three dimensional.}
}
\value{
PieChartSpec
}
\description{
A <a href="/chart/interactive/docs/gallery/piechart">pie chart</a>.
}
\details{
legendPosition takes one of the following values:
\itemize{
 \item{PIE_CHART_LEGEND_POSITION_UNSPECIFIED - Default value, do not use.}
 \item{BOTTOM_LEGEND - The legend is rendered on the bottom of the chart.}
 \item{LEFT_LEGEND - The legend is rendered on the left of the chart.}
 \item{RIGHT_LEGEND - The legend is rendered on the right of the chart.}
 \item{TOP_LEGEND - The legend is rendered on the top of the chart.}
 \item{NO_LEGEND - No legend is rendered.}
 \item{LABELED_LEGEND - Each pie slice has a label attached to it.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#PieChartSpec}{Google's Documentation for PieChartSpec}
}

