% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_PivotFilterCriteria}
\alias{gsv4_PivotFilterCriteria}
\title{gsv4_PivotFilterCriteria}
\usage{
gsv4_PivotFilterCriteria(visibleValues = NULL)
}
\arguments{
\item{visibleValues}{list. Values that should be included.  Values not listed here are excluded.}
}
\value{
PivotFilterCriteria
}
\description{
Criteria for showing/hiding rows in a pivot table.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#PivotFilterCriteria}{Google's Documentation for PivotFilterCriteria}
}

