% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_PivotGroup}
\alias{gsv4_PivotGroup}
\title{gsv4_PivotGroup}
\usage{
gsv4_PivotGroup(showTotals = NULL, sortOrder = NULL,
  sourceColumnOffset = NULL, valueBucket = NULL, valueMetadata = NULL)
}
\arguments{
\item{showTotals}{logical. TRUE if the pivot table should include the totals for this grouping.}

\item{sortOrder}{string. The order the values in this group should be sorted. sortOrder must take one of the following values: SORT_ORDER_UNSPECIFIED, ASCENDING, DESCENDING
See the details section for the definition of each of these values.}

\item{sourceColumnOffset}{integer. The column offset of the source range that this grouping is based on.

For example, if the source was `C10:E15`, a `sourceColumnOffset` of `0`
means this group refers to column `C`, whereas the offset `1` would refer
to column `D`.}

\item{valueBucket}{\code{\link{gsv4_PivotGroupSortValueBucket}} object. Information about which values in a pivot group should be used for sorting.}

\item{valueMetadata}{list of \code{\link{gsv4_PivotGroupValueMetadata}} objects. Metadata about values in the grouping.}
}
\value{
PivotGroup
}
\description{
A single grouping (either row or column) in a pivot table.
}
\details{
sortOrder takes one of the following values:
\itemize{
 \item{SORT_ORDER_UNSPECIFIED - Default value, do not use this.}
 \item{ASCENDING - Sort ascending.}
 \item{DESCENDING - Sort descending.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#PivotGroup}{Google's Documentation for PivotGroup}
}

