% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_PivotGroupSortValueBucket}
\alias{gsv4_PivotGroupSortValueBucket}
\title{gsv4_PivotGroupSortValueBucket}
\usage{
gsv4_PivotGroupSortValueBucket(buckets=NULL, valuesIndex=NULL)
}
\arguments{
\item{buckets}{list of \code{\link{gsv4_ExtendedValue}} objects. Determines the bucket from which values are chosen to sort.

For example, in a pivot table with one row group & two column groups,
the row group can list up to two values. The first value corresponds
to a value within the first column group, and the second value
corresponds to a value in the second column group.  If no values
are listed, this would indicate that the row should be sorted according
to the "Grand Total" over the column groups. If a single value is listed,
this would correspond to using the "Total" of that bucket.}

\item{valuesIndex}{integer. The offset in the PivotTable.values list which the values in this
grouping should be sorted by.}
}
\value{
PivotGroupSortValueBucket
}
\description{
Information about which values in a pivot group should be used for sorting.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#PivotGroupSortValueBucket}{Google's Documentation for PivotGroupSortValueBucket}
}

