% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_PivotGroupValueMetadata}
\alias{gsv4_PivotGroupValueMetadata}
\title{gsv4_PivotGroupValueMetadata}
\usage{
gsv4_PivotGroupValueMetadata(collapsed = NULL, value = NULL)
}
\arguments{
\item{collapsed}{logical. TRUE if the data corresponding to the value is collapsed.}

\item{value}{\code{\link{gsv4_ExtendedValue}} object. The kinds of value that a cell in a spreadsheet can have.}
}
\value{
PivotGroupValueMetadata
}
\description{
Metadata about a value in a pivot grouping.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#PivotGroupValueMetadata}{Google's Documentation for PivotGroupValueMetadata}
}

