% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_PivotValue}
\alias{gsv4_PivotValue}
\title{gsv4_PivotValue}
\usage{
gsv4_PivotValue(formula = NULL, name = NULL, sourceColumnOffset = NULL,
  summarizeFunction = NULL)
}
\arguments{
\item{formula}{string. A custom formula to calculate the value.  The formula must start
with an `=` character.}

\item{name}{string. A name to use for the value. This is only used if formula was set.
Otherwise, the column name is used.}

\item{sourceColumnOffset}{integer. The column offset of the source range that this value reads from.

For example, if the source was `C10:E15`, a `sourceColumnOffset` of `0`
means this value refers to column `C`, whereas the offset `1` would
refer to column `D`.}

\item{summarizeFunction}{string. A function to summarize the value.
If formula is set, the only supported values are
SUM and
CUSTOM.
If sourceColumnOffset is set, then `CUSTOM`
is not supported. summarizeFunction must take one of the following values: PIVOT_STANDARD_VALUE_FUNCTION_UNSPECIFIED, SUM, COUNTA, COUNT, COUNTUNIQUE, AVERAGE, MAX, MIN, MEDIAN, PRODUCT, STDEV, STDEVP, VAR, VARP, CUSTOM
See the details section for the definition of each of these values.}
}
\value{
PivotValue
}
\description{
The definition of how a value in a pivot table should be calculated.
}
\details{
summarizeFunction takes one of the following values:
\itemize{
 \item{PIVOT_STANDARD_VALUE_FUNCTION_UNSPECIFIED - The default, do not use.}
 \item{SUM - Corresponds to the `SUM` function.}
 \item{COUNTA - Corresponds to the `COUNTA` function.}
 \item{COUNT - Corresponds to the `COUNT` function.}
 \item{COUNTUNIQUE - Corresponds to the `COUNTUNIQUE` function.}
 \item{AVERAGE - Corresponds to the `AVERAGE` function.}
 \item{MAX - Corresponds to the `MAX` function.}
 \item{MIN - Corresponds to the `MIN` function.}
 \item{MEDIAN - Corresponds to the `MEDIAN` function.}
 \item{PRODUCT - Corresponds to the `PRODUCT` function.}
 \item{STDEV - Corresponds to the `STDEV` function.}
 \item{STDEVP - Corresponds to the `STDEVP` function.}
 \item{VAR - Corresponds to the `VAR` function.}
 \item{VARP - Corresponds to the `VARP` function.}
 \item{CUSTOM - Indicates the formula should be used as-is.
Only valid if PivotValue.formula was set.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#PivotValue}{Google's Documentation for PivotValue}
}

