% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gsv4_prep_values}
\alias{gsv4_prep_values}
\title{gs_prep_values}
\usage{
gsv4_prep_values(values, col_names=TRUE)
}
\arguments{
\item{values}{\code{data.frame}; A data.frame to be passed as a values matrix typically
used with values operations write, update, or append.}

\item{col_names}{logical; indicates whether column names of input should be included in the edit, i.e. prepended to the input
authorization, respectively}
}
\value{
\code{data.frame} parsed from a values matrix returned by the Sheets V4 API
}
\description{
Preparing a data.frame to be passed as the values argument of a function call
}
\examples{
\dontrun{
my_values <- gsv4_prep_values(iris[5,], col_names=FALSE)
gsv4_values_update(spreadsheetId = this_spreadsheetId,
                   valueInputOption = 'RAW', 
                   range = "iris!A6", 
                   input = gsv4_ValueRange(values=my_values, 
                                           majorDimension = 'ROWS', 
                                           range="iris!A6"))
}

}

