% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_ProtectedRange}
\alias{gsv4_ProtectedRange}
\title{gsv4_ProtectedRange}
\usage{
gsv4_ProtectedRange(range=NULL, description=NULL, editors=NULL, namedRangeId=NULL, protectedRangeId=NULL, requestingUserCanEdit=NULL, unprotectedRanges=NULL, warningOnly=NULL)
}
\arguments{
\item{range}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}

\item{description}{string. The description of this protected range.}

\item{editors}{\code{\link{gsv4_Editors}} object. The editors of a protected range.}

\item{namedRangeId}{string. The named range this protected range is backed by, if any.

When writing, only one of range or named_range_id
may be set.}

\item{protectedRangeId}{integer. The ID of the protected range.
This field is read-only.}

\item{requestingUserCanEdit}{logical. TRUE if the user who requested this protected range can edit the
protected area.
This field is read-only.}

\item{unprotectedRanges}{list of \code{\link{gsv4_GridRange}} objects. The list of unprotected ranges within a protected sheet.
Unprotected ranges are only supported on protected sheets.}

\item{warningOnly}{logical. TRUE if this protected range will show a warning when editing.
Warning-based protection means that every user can edit data in the
protected range, except editing will prompt a warning asking the user
to confirm the edit.

When writing: if this field is TRUE, then editors is ignored.
Additionally, if this field is changed from TRUE to FALSE and the
`editors` field is not set (nor included in the field mask), then
the editors will be set to all the editors in the document.}
}
\value{
ProtectedRange
}
\description{
A protected range.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#ProtectedRange}{Google's Documentation for ProtectedRange}
}

