% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_RepeatCellRequest}
\alias{gsv4_RepeatCellRequest}
\title{gsv4_RepeatCellRequest}
\usage{
gsv4_RepeatCellRequest(range=NULL, fields=NULL, cell=NULL)
}
\arguments{
\item{range}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}

\item{fields}{string. The fields that should be updated.  At least one field must be specified.
The root `cell` is implied and should not be specified.
A single `"*"` can be used as short-hand for listing every field.}

\item{cell}{\code{\link{gsv4_CellData}} object. Data about a specific cell.}
}
\value{
RepeatCellRequest
}
\description{
Updates all cells in the range to the values in the given Cell object.
Only the fields listed in the fields field are updated; others are
unchanged.
}
\details{
If writing a cell with a formula, the formula's ranges will automatically
increment for each field in the range.
For example, if writing a cell with formula `=A1` into range B2:C4,
B2 would be `=A1`, B3 would be `=A2`, B4 would be `=A3`,
C2 would be `=B1`, C3 would be `=B2`, C4 would be `=B3`.

To keep the formula's ranges static, use the `$` indicator.
For example, use the formula `=$A$1` to prevent both the row and the
column from incrementing.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#RepeatCellRequest}{Google's Documentation for RepeatCellRequest}
}

