% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_Request}
\alias{gsv4_Request}
\title{gsv4_Request}
\usage{
gsv4_Request(addBanding = NULL, addChart = NULL,
  addConditionalFormatRule = NULL, addFilterView = NULL,
  addNamedRange = NULL, addProtectedRange = NULL, addSheet = NULL,
  appendCells = NULL, appendDimension = NULL, autoFill = NULL,
  autoResizeDimensions = NULL, clearBasicFilter = NULL, copyPaste = NULL,
  cutPaste = NULL, deleteBanding = NULL,
  deleteConditionalFormatRule = NULL, deleteDimension = NULL,
  deleteEmbeddedObject = NULL, deleteFilterView = NULL,
  deleteNamedRange = NULL, deleteProtectedRange = NULL,
  deleteRange = NULL, deleteSheet = NULL, duplicateFilterView = NULL,
  duplicateSheet = NULL, findReplace = NULL, insertDimension = NULL,
  insertRange = NULL, mergeCells = NULL, moveDimension = NULL,
  pasteData = NULL, repeatCell = NULL, setBasicFilter = NULL,
  setDataValidation = NULL, sortRange = NULL, textToColumns = NULL,
  unmergeCells = NULL, updateBanding = NULL, updateBorders = NULL,
  updateCells = NULL, updateChartSpec = NULL,
  updateConditionalFormatRule = NULL, updateDimensionProperties = NULL,
  updateEmbeddedObjectPosition = NULL, updateFilterView = NULL,
  updateNamedRange = NULL, updateProtectedRange = NULL,
  updateSheetProperties = NULL, updateSpreadsheetProperties = NULL)
}
\arguments{
\item{addBanding}{\code{\link{gsv4_AddBandingRequest}} object. Adds a new banded range to the spreadsheet.}

\item{addChart}{\code{\link{gsv4_AddChartRequest}} object. Adds a chart to a sheet in the spreadsheet.}

\item{addConditionalFormatRule}{\code{\link{gsv4_AddConditionalFormatRuleRequest}} object. Adds a new conditional format rule at the given index.
All subsequent rules' indexes are incremented.}

\item{addFilterView}{\code{\link{gsv4_AddFilterViewRequest}} object. Adds a filter view.}

\item{addNamedRange}{\code{\link{gsv4_AddNamedRangeRequest}} object. Adds a named range to the spreadsheet.}

\item{addProtectedRange}{\code{\link{gsv4_AddProtectedRangeRequest}} object. Adds a new protected range.}

\item{addSheet}{\code{\link{gsv4_AddSheetRequest}} object. Adds a new sheet.
When a sheet is added at a given index,
all subsequent sheets' indexes are incremented.
To add an object sheet, use AddChartRequest instead and specify
EmbeddedObjectPosition.sheetId or
EmbeddedObjectPosition.newSheet.}

\item{appendCells}{\code{\link{gsv4_AppendCellsRequest}} object. Adds new cells after the last row with data in a sheet,
inserting new rows into the sheet if necessary.}

\item{appendDimension}{\code{\link{gsv4_AppendDimensionRequest}} object. Appends rows or columns to the end of a sheet.}

\item{autoFill}{\code{\link{gsv4_AutoFillRequest}} object. Fills in more data based on existing data.}

\item{autoResizeDimensions}{\code{\link{gsv4_AutoResizeDimensionsRequest}} object. Automatically resizes one or more dimensions based on the contents
of the cells in that dimension.}

\item{clearBasicFilter}{\code{\link{gsv4_ClearBasicFilterRequest}} object. Clears the basic filter, if any exists on the sheet.}

\item{copyPaste}{\code{\link{gsv4_CopyPasteRequest}} object. Copies data from the source to the destination.}

\item{cutPaste}{\code{\link{gsv4_CutPasteRequest}} object. Moves data from the source to the destination.}

\item{deleteBanding}{\code{\link{gsv4_DeleteBandingRequest}} object. Removes the banded range with the given ID from the spreadsheet.}

\item{deleteConditionalFormatRule}{\code{\link{gsv4_DeleteConditionalFormatRuleRequest}} object. Deletes a conditional format rule at the given index.
All subsequent rules' indexes are decremented.}

\item{deleteDimension}{\code{\link{gsv4_DeleteDimensionRequest}} object. Deletes the dimensions from the sheet.}

\item{deleteEmbeddedObject}{\code{\link{gsv4_DeleteEmbeddedObjectRequest}} object. Deletes the embedded object with the given ID.}

\item{deleteFilterView}{\code{\link{gsv4_DeleteFilterViewRequest}} object. Deletes a particular filter view.}

\item{deleteNamedRange}{\code{\link{gsv4_DeleteNamedRangeRequest}} object. Removes the named range with the given ID from the spreadsheet.}

\item{deleteProtectedRange}{\code{\link{gsv4_DeleteProtectedRangeRequest}} object. Deletes the protected range with the given ID.}

\item{deleteRange}{\code{\link{gsv4_DeleteRangeRequest}} object. Deletes a range of cells, shifting other cells into the deleted area.}

\item{deleteSheet}{\code{\link{gsv4_DeleteSheetRequest}} object. Deletes the requested sheet.}

\item{duplicateFilterView}{\code{\link{gsv4_DuplicateFilterViewRequest}} object. Duplicates a particular filter view.}

\item{duplicateSheet}{\code{\link{gsv4_DuplicateSheetRequest}} object. Duplicates the contents of a sheet.}

\item{findReplace}{\code{\link{gsv4_FindReplaceRequest}} object. Finds and replaces data in cells over a range, sheet, or all sheets.}

\item{insertDimension}{\code{\link{gsv4_InsertDimensionRequest}} object. Inserts rows or columns in a sheet at a particular index.}

\item{insertRange}{\code{\link{gsv4_InsertRangeRequest}} object. Inserts cells into a range, shifting the existing cells over or down.}

\item{mergeCells}{\code{\link{gsv4_MergeCellsRequest}} object. Merges all cells in the range.}

\item{moveDimension}{\code{\link{gsv4_MoveDimensionRequest}} object. Moves one or more rows or columns.}

\item{pasteData}{\code{\link{gsv4_PasteDataRequest}} object. Inserts data into the spreadsheet starting at the specified coordinate.}

\item{repeatCell}{\code{\link{gsv4_RepeatCellRequest}} object. Updates all cells in the range to the values in the given Cell object.
Only the fields listed in the fields field are updated; others are
unchanged.

If writing a cell with a formula, the formula's ranges will automatically
increment for each field in the range.
For example, if writing a cell with formula `=A1` into range B2:C4,
B2 would be `=A1`, B3 would be `=A2`, B4 would be `=A3`,
C2 would be `=B1`, C3 would be `=B2`, C4 would be `=B3`.

To keep the formula's ranges static, use the `$` indicator.
For example, use the formula `=$A$1` to prevent both the row and the
column from incrementing.}

\item{setBasicFilter}{\code{\link{gsv4_SetBasicFilterRequest}} object. Sets the basic filter associated with a sheet.}

\item{setDataValidation}{\code{\link{gsv4_SetDataValidationRequest}} object. Sets a data validation rule to every cell in the range.
To clear validation in a range, call this with no rule specified.}

\item{sortRange}{\code{\link{gsv4_SortRangeRequest}} object. Sorts data in rows based on a sort order per column.}

\item{textToColumns}{\code{\link{gsv4_TextToColumnsRequest}} object. Splits a column of text into multiple columns,
based on a delimiter in each cell.}

\item{unmergeCells}{\code{\link{gsv4_UnmergeCellsRequest}} object. Unmerges cells in the given range.}

\item{updateBanding}{\code{\link{gsv4_UpdateBandingRequest}} object. Updates properties of the supplied banded range.}

\item{updateBorders}{\code{\link{gsv4_UpdateBordersRequest}} object. Updates the borders of a range.
If a field is not set in the request, that means the border remains as-is.
For example, with two subsequent UpdateBordersRequest:

 1. range: A1:A5 `{ top: RED, bottom: WHITE }`
 2. range: A1:A5 `{ left: BLUE }`

That would result in A1:A5 having a borders of
`{ top: RED, bottom: WHITE, left: BLUE }`.
If you want to clear a border, explicitly set the style to
NONE.}

\item{updateCells}{\code{\link{gsv4_UpdateCellsRequest}} object. Updates all cells in a range with new data.}

\item{updateChartSpec}{\code{\link{gsv4_UpdateChartSpecRequest}} object. Updates a chart's specifications.
(This does not move or resize a chart. To move or resize a chart, use
 UpdateEmbeddedObjectPositionRequest.)}

\item{updateConditionalFormatRule}{\code{\link{gsv4_UpdateConditionalFormatRuleRequest}} object. Updates a conditional format rule at the given index,
or moves a conditional format rule to another index.}

\item{updateDimensionProperties}{\code{\link{gsv4_UpdateDimensionPropertiesRequest}} object. Updates properties of dimensions within the specified range.}

\item{updateEmbeddedObjectPosition}{\code{\link{gsv4_UpdateEmbeddedObjectPositionRequest}} object. Update an embedded object's position (such as a moving or resizing a
chart or image).}

\item{updateFilterView}{\code{\link{gsv4_UpdateFilterViewRequest}} object. Updates properties of the filter view.}

\item{updateNamedRange}{\code{\link{gsv4_UpdateNamedRangeRequest}} object. Updates properties of the named range with the specified
namedRangeId.}

\item{updateProtectedRange}{\code{\link{gsv4_UpdateProtectedRangeRequest}} object. Updates an existing protected range with the specified
protectedRangeId.}

\item{updateSheetProperties}{\code{\link{gsv4_UpdateSheetPropertiesRequest}} object. Updates properties of the sheet with the specified
sheetId.}

\item{updateSpreadsheetProperties}{\code{\link{gsv4_UpdateSpreadsheetPropertiesRequest}} object. Updates properties of a spreadsheet.}
}
\value{
Request
}
\description{
A single kind of update to apply to a spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#Request}{Google's Documentation for Request}
}

