% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_RowData}
\alias{gsv4_RowData}
\title{gsv4_RowData}
\usage{
gsv4_RowData(values=NULL)
}
\arguments{
\item{values}{list of \code{\link{gsv4_CellData}} objects. The values in the row, one per column.}
}
\value{
RowData
}
\description{
Data about each cell in a row.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#RowData}{Google's Documentation for RowData}
}

