% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_SetBasicFilterRequest}
\alias{gsv4_SetBasicFilterRequest}
\title{gsv4_SetBasicFilterRequest}
\usage{
gsv4_SetBasicFilterRequest(filter=NULL)
}
\arguments{
\item{filter}{\code{\link{gsv4_BasicFilter}} object. The default filter associated with a sheet.}
}
\value{
SetBasicFilterRequest
}
\description{
Sets the basic filter associated with a sheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#SetBasicFilterRequest}{Google's Documentation for SetBasicFilterRequest}
}

