% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_Sheet}
\alias{gsv4_Sheet}
\title{gsv4_Sheet}
\usage{
gsv4_Sheet(bandedRanges = NULL, basicFilter = NULL, charts = NULL,
  conditionalFormats = NULL, data = NULL, filterViews = NULL,
  merges = NULL, properties = NULL, protectedRanges = NULL)
}
\arguments{
\item{bandedRanges}{list of \code{\link{gsv4_BandedRange}} objects. The banded (i.e. alternating colors) ranges on this sheet.}

\item{basicFilter}{\code{\link{gsv4_BasicFilter}} object. The default filter associated with a sheet.}

\item{charts}{list of \code{\link{gsv4_EmbeddedChart}} objects. The specifications of every chart on this sheet.}

\item{conditionalFormats}{list of \code{\link{gsv4_ConditionalFormatRule}} objects. The conditional format rules in this sheet.}

\item{data}{list of \code{\link{gsv4_GridData}} objects. Data in the grid, if this is a grid sheet.
The number of GridData objects returned is dependent on the number of
ranges requested on this sheet. For example, if this is representing
`Sheet1`, and the spreadsheet was requested with ranges
`Sheet1!A1:C10` and `Sheet1!D15:E20`, then the first GridData will have a
startRow/startColumn of `0`,
while the second one will have `startRow 14` (zero-based row 15),
and `startColumn 3` (zero-based column D).}

\item{filterViews}{list of \code{\link{gsv4_FilterView}} objects. The filter views in this sheet.}

\item{merges}{list of \code{\link{gsv4_GridRange}} objects. The ranges that are merged together.}

\item{properties}{\code{\link{gsv4_SheetProperties}} object. Properties of a sheet.}

\item{protectedRanges}{list of \code{\link{gsv4_ProtectedRange}} objects. The protected ranges in this sheet.}
}
\value{
Sheet
}
\description{
A sheet in a spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#Sheet}{Google's Documentation for Sheet}
}

