% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheets.sheets.R
\name{gsv4_sheets_copyTo}
\alias{gsv4_sheets_copyTo}
\title{gsv4_sheets_copyTo}
\usage{
gsv4_sheets_copyTo(spreadsheetId, sheetId, input, standard_params = list(...),
  ...)
}
\arguments{
\item{spreadsheetId}{string (required). The ID of the spreadsheet containing the sheet to copy.}

\item{sheetId}{integer (required). The ID of the sheet to copy.}

\item{input}{\code{\link{gsv4_CopySheetToAnotherSpreadsheetRequest}}. The request to copy a sheet across spreadsheets.}

\item{standard_params}{a list of parameters for controlling the HTTP request and its response.
Refer to \code{\link{gsv4_standard_parameters}} for details on its arguments}

\item{...}{arguments to be used to form standard_params argument if it is not supplied directly.}
}
\value{
SheetProperties
}
\description{
Copies a single sheet from a spreadsheet to another spreadsheet.
Returns the properties of the newly created sheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets.sheets/copyTo}{Google's Documentation for copyTo}
}

