% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_SortSpec}
\alias{gsv4_SortSpec}
\title{gsv4_SortSpec}
\usage{
gsv4_SortSpec(dimensionIndex=NULL, sortOrder=NULL)
}
\arguments{
\item{dimensionIndex}{integer. The dimension the sort should be applied to.}

\item{sortOrder}{string. The order data should be sorted. sortOrder must take one of the following values: SORT_ORDER_UNSPECIFIED, ASCENDING, DESCENDING
See the details section for the definition of each of these values.}
}
\value{
SortSpec
}
\description{
A sort order associated with a specific column or row.
}
\details{
sortOrder takes one of the following values:
\itemize{
 \item{SORT_ORDER_UNSPECIFIED - Default value, do not use this.}
 \item{ASCENDING - Sort ascending.}
 \item{DESCENDING - Sort descending.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#SortSpec}{Google's Documentation for SortSpec}
}

