% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_SourceAndDestination}
\alias{gsv4_SourceAndDestination}
\title{gsv4_SourceAndDestination}
\usage{
gsv4_SourceAndDestination(dimension = NULL, fillLength = NULL,
  source = NULL)
}
\arguments{
\item{dimension}{string. The dimension that data should be filled into. dimension must take one of the following values: DIMENSION_UNSPECIFIED, ROWS, COLUMNS
See the details section for the definition of each of these values.}

\item{fillLength}{integer. The number of rows or columns that data should be filled into.
Positive numbers expand beyond the last row or last column
of the source.  Negative numbers expand before the first row
or first column of the source.}

\item{source}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}
}
\value{
SourceAndDestination
}
\description{
A combination of a source range and how to extend that source.
}
\details{
dimension takes one of the following values:
\itemize{
 \item{DIMENSION_UNSPECIFIED - The default value, do not use.}
 \item{ROWS - Operates on the rows of a sheet.}
 \item{COLUMNS - Operates on the columns of a sheet.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#SourceAndDestination}{Google's Documentation for SourceAndDestination}
}

