% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_Spreadsheet}
\alias{gsv4_Spreadsheet}
\title{gsv4_Spreadsheet}
\usage{
gsv4_Spreadsheet(spreadsheetId=NULL, namedRanges=NULL, properties=NULL, sheets=NULL, spreadsheetUrl=NULL)
}
\arguments{
\item{spreadsheetId}{string. The ID of the spreadsheet.
This field is read-only.}

\item{namedRanges}{list of \code{\link{gsv4_NamedRange}} objects. The named ranges defined in a spreadsheet.}

\item{properties}{\code{\link{gsv4_SpreadsheetProperties}} object. Properties of a spreadsheet.}

\item{sheets}{list of \code{\link{gsv4_Sheet}} objects. The sheets that are part of a spreadsheet.}

\item{spreadsheetUrl}{string. The url of the spreadsheet.
This field is read-only.}
}
\value{
Spreadsheet
}
\description{
Resource that represents a spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#Spreadsheet}{Google's Documentation for Spreadsheet}
}

