% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_SpreadsheetProperties}
\alias{gsv4_SpreadsheetProperties}
\title{gsv4_SpreadsheetProperties}
\usage{
gsv4_SpreadsheetProperties(autoRecalc=NULL, defaultFormat=NULL, iterativeCalculationSettings=NULL, locale=NULL, timeZone=NULL, title=NULL)
}
\arguments{
\item{autoRecalc}{string. The amount of time to wait before volatile functions are recalculated. autoRecalc must take one of the following values: RECALCULATION_INTERVAL_UNSPECIFIED, ON_CHANGE, MINUTE, HOUR
See the details section for the definition of each of these values.}

\item{defaultFormat}{\code{\link{gsv4_CellFormat}} object. The format of a cell.}

\item{iterativeCalculationSettings}{\code{\link{gsv4_IterativeCalculationSettings}} object. Settings to control how circular dependencies are resolved with iterative
calculation.}

\item{locale}{string. The locale of the spreadsheet in one of the following formats:

* an ISO 639-1 language code such as `en`

* an ISO 639-2 language code such as `fil`, if no 639-1 code exists

* a combination of the ISO language code and country code, such as `en_US`

Note: when updating this field, not all locales/languages are supported.}

\item{timeZone}{string. The time zone of the spreadsheet, in CLDR format such as
`America/New_York`. If the time zone isn't recognized, this may
be a custom time zone such as `GMT-07:00`.}

\item{title}{string. The title of the spreadsheet.}
}
\value{
SpreadsheetProperties
}
\description{
Properties of a spreadsheet.
}
\details{
autoRecalc takes one of the following values:
\itemize{
 \item{RECALCULATION_INTERVAL_UNSPECIFIED - Default value. This value must not be used.}
 \item{ON_CHANGE - Volatile functions are updated on every change.}
 \item{MINUTE - Volatile functions are updated on every change and every minute.}
 \item{HOUR - Volatile functions are updated on every change and hourly.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#SpreadsheetProperties}{Google's Documentation for SpreadsheetProperties}
}

