% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_standard_parameters}
\alias{gsv4_standard_parameters}
\title{gsv4_standard_parameters}
\usage{
gsv4_standard_parameters(fields=NULL, .xgafv=NULL, alt=NULL, callback=NULL, pp=NULL, prettyPrint=NULL, quotaUser=NULL, upload_protocol=NULL, uploadType=NULL)
}
\arguments{
\item{fields}{string. Selector specifying which fields to include in a partial response.}

\item{.xgafv}{string. V1 error format. .xgafv must take one of the following values: 1, 2
See the details section for the definition of each of these values.}

\item{alt}{string. Data format for response. alt must take one of the following values: json, media, proto
See the details section for the definition of each of these values.}

\item{callback}{string. JSONP}

\item{pp}{logical. Pretty-print response.}

\item{prettyPrint}{logical. Returns response with indentations and line breaks.}

\item{quotaUser}{string. Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{upload_protocol}{string. Upload protocol for media (e.g. "raw", "multipart").}

\item{uploadType}{string. Legacy upload protocol for media (e.g. "media", "multipart").}
}
\value{
A \code{list} with components named as the arguments
.xgafv takes one of the following values:
\itemize{
 \item{1 - v1 error format}
 \item{2 - v2 error format}
}

alt takes one of the following values:
\itemize{
 \item{json - Responses with Content-Type of application/json}
 \item{media - Media download with context-dependent Content-Type}
 \item{proto - Responses with Content-Type of application/x-protobuf}
}
}
\description{
Auxiliary function for modifying HTTP methods in the Sheets API v4.
}
\seealso{
\href{https://developers.google.com/sheets/api/query-parameters}{Google's Documentation of Standard Query Parameters}
}

