% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_TextFormatRun}
\alias{gsv4_TextFormatRun}
\title{gsv4_TextFormatRun}
\usage{
gsv4_TextFormatRun(format = NULL, startIndex = NULL)
}
\arguments{
\item{format}{\code{\link{gsv4_TextFormat}} object. The format of a run of text in a cell.
Absent values indicate that the field isn't specified.}

\item{startIndex}{integer. The character index where this run starts.}
}
\value{
TextFormatRun
}
\description{
A run of a text format. The format of this run continues until the start
index of the next run.
When updating, all fields must be set.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#TextFormatRun}{Google's Documentation for TextFormatRun}
}

