% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_TextRotation}
\alias{gsv4_TextRotation}
\title{gsv4_TextRotation}
\usage{
gsv4_TextRotation(angle=NULL, vertical=NULL)
}
\arguments{
\item{angle}{integer. The angle between the standard orientation and the desired orientation.
Measured in degrees. Valid values are between -90 and 90. Positive
angles are angled upwards, negative are angled downwards.

Note: For LTR text direction positive angles are in the counterclockwise
direction, whereas for RTL they are in the clockwise direction}

\item{vertical}{logical. If TRUE, text reads top to bottom, but the orientation of individual
characters is unchanged.
For example:

    | V |
    | e |
    | r |
    | t |
    | i |
    | c |
    | a |
    | l |}
}
\value{
TextRotation
}
\description{
The rotation applied to text in a cell.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#TextRotation}{Google's Documentation for TextRotation}
}

