% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_TextToColumnsRequest}
\alias{gsv4_TextToColumnsRequest}
\title{gsv4_TextToColumnsRequest}
\usage{
gsv4_TextToColumnsRequest(delimiter = NULL, delimiterType = NULL,
  source = NULL)
}
\arguments{
\item{delimiter}{string. The delimiter to use. Used only if delimiterType is
CUSTOM.}

\item{delimiterType}{string. The delimiter type to use. delimiterType must take one of the following values: DELIMITER_TYPE_UNSPECIFIED, COMMA, SEMICOLON, PERIOD, SPACE, CUSTOM
See the details section for the definition of each of these values.}

\item{source}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}
}
\value{
TextToColumnsRequest
}
\description{
Splits a column of text into multiple columns,
based on a delimiter in each cell.
}
\details{
delimiterType takes one of the following values:
\itemize{
 \item{DELIMITER_TYPE_UNSPECIFIED - Default value. This value must not be used.}
 \item{COMMA - ","}
 \item{SEMICOLON - ";"}
 \item{PERIOD - "."}
 \item{SPACE - " "}
 \item{CUSTOM - A custom value as defined in delimiter.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#TextToColumnsRequest}{Google's Documentation for TextToColumnsRequest}
}

