% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_UpdateBandingRequest}
\alias{gsv4_UpdateBandingRequest}
\title{gsv4_UpdateBandingRequest}
\usage{
gsv4_UpdateBandingRequest(fields = NULL, bandedRange = NULL)
}
\arguments{
\item{fields}{string. The fields that should be updated.  At least one field must be specified.
The root `bandedRange` is implied and should not be specified.
A single `"*"` can be used as short-hand for listing every field.}

\item{bandedRange}{\code{\link{gsv4_BandedRange}} object. A banded (alternating colors) range in a sheet.}
}
\value{
UpdateBandingRequest
}
\description{
Updates properties of the supplied banded range.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#UpdateBandingRequest}{Google's Documentation for UpdateBandingRequest}
}

