% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_UpdateBordersRequest}
\alias{gsv4_UpdateBordersRequest}
\title{gsv4_UpdateBordersRequest}
\usage{
gsv4_UpdateBordersRequest(range=NULL, bottom=NULL, innerHorizontal=NULL, innerVertical=NULL, left=NULL, right=NULL, top=NULL)
}
\arguments{
\item{range}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}

\item{bottom}{\code{\link{gsv4_Border}} object. A border along a cell.}

\item{innerHorizontal}{\code{\link{gsv4_Border}} object. A border along a cell.}

\item{innerVertical}{\code{\link{gsv4_Border}} object. A border along a cell.}

\item{left}{\code{\link{gsv4_Border}} object. A border along a cell.}

\item{right}{\code{\link{gsv4_Border}} object. A border along a cell.}

\item{top}{\code{\link{gsv4_Border}} object. A border along a cell.}
}
\value{
UpdateBordersRequest
}
\description{
Updates the borders of a range.
If a field is not set in the request, that means the border remains as-is.
For example, with two subsequent UpdateBordersRequest:
}
\details{
1. range: A1:A5 `{ top: RED, bottom: WHITE }`
 2. range: A1:A5 `{ left: BLUE }`

That would result in A1:A5 having a borders of
`{ top: RED, bottom: WHITE, left: BLUE }`.
If you want to clear a border, explicitly set the style to
NONE.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#UpdateBordersRequest}{Google's Documentation for UpdateBordersRequest}
}

