% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_UpdateCellsRequest}
\alias{gsv4_UpdateCellsRequest}
\title{gsv4_UpdateCellsRequest}
\usage{
gsv4_UpdateCellsRequest(range=NULL, rows=NULL, fields=NULL, start=NULL)
}
\arguments{
\item{range}{\code{\link{gsv4_GridRange}} object. A range on a sheet.
All indexes are zero-based.
Indexes are half open, e.g the start index is inclusive
and the end index is exclusive -- [start_index, end_index).
Missing indexes indicate the range is unbounded on that side.

For example, if `"Sheet1"` is sheet ID 0, then:

  `Sheet1!A1:A1 == sheet_id: 0,
                  start_row_index: 0, end_row_index: 1,
                  start_column_index: 0, end_column_index: 1`

  `Sheet1!A3:B4 == sheet_id: 0,
                  start_row_index: 2, end_row_index: 4,
                  start_column_index: 0, end_column_index: 2`

  `Sheet1!A:B == sheet_id: 0,
                start_column_index: 0, end_column_index: 2`

  `Sheet1!A5:B == sheet_id: 0,
                 start_row_index: 4,
                 start_column_index: 0, end_column_index: 2`

  `Sheet1 == sheet_id:0`

The start index must always be less than or equal to the end index.
If the start index equals the end index, then the range is empty.
Empty ranges are typically not meaningful and are usually rendered in the
UI as `#REF!`.}

\item{rows}{list of \code{\link{gsv4_RowData}} objects. The data to write.}

\item{fields}{string. The fields of CellData that should be updated.
At least one field must be specified.
The root is the CellData; 'row.values.' should not be specified.
A single `"*"` can be used as short-hand for listing every field.}

\item{start}{\code{\link{gsv4_GridCoordinate}} object. A coordinate in a sheet.
All indexes are zero-based.}
}
\value{
UpdateCellsRequest
}
\description{
Updates all cells in a range with new data.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#UpdateCellsRequest}{Google's Documentation for UpdateCellsRequest}
}

