% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_UpdateChartSpecRequest}
\alias{gsv4_UpdateChartSpecRequest}
\title{gsv4_UpdateChartSpecRequest}
\usage{
gsv4_UpdateChartSpecRequest(chartId=NULL, spec=NULL)
}
\arguments{
\item{chartId}{integer. The ID of the chart to update.}

\item{spec}{\code{\link{gsv4_ChartSpec}} object. The specifications of a chart.}
}
\value{
UpdateChartSpecRequest
}
\description{
Updates a chart's specifications.
(This does not move or resize a chart. To move or resize a chart, use
 UpdateEmbeddedObjectPositionRequest.)
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#UpdateChartSpecRequest}{Google's Documentation for UpdateChartSpecRequest}
}

