% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_UpdateConditionalFormatRuleRequest}
\alias{gsv4_UpdateConditionalFormatRuleRequest}
\title{gsv4_UpdateConditionalFormatRuleRequest}
\usage{
gsv4_UpdateConditionalFormatRuleRequest(sheetId=NULL, index=NULL, newIndex=NULL, rule=NULL)
}
\arguments{
\item{sheetId}{integer. The sheet of the rule to move.  Required if new_index is set,
unused otherwise.}

\item{index}{integer. The zero-based index of the rule that should be replaced or moved.}

\item{newIndex}{integer. The zero-based new index the rule should end up at.}

\item{rule}{\code{\link{gsv4_ConditionalFormatRule}} object. A rule describing a conditional format.}
}
\value{
UpdateConditionalFormatRuleRequest
}
\description{
Updates a conditional format rule at the given index,
or moves a conditional format rule to another index.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#UpdateConditionalFormatRuleRequest}{Google's Documentation for UpdateConditionalFormatRuleRequest}
}

