% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_UpdateDimensionPropertiesRequest}
\alias{gsv4_UpdateDimensionPropertiesRequest}
\title{gsv4_UpdateDimensionPropertiesRequest}
\usage{
gsv4_UpdateDimensionPropertiesRequest(range=NULL, fields=NULL, properties=NULL)
}
\arguments{
\item{range}{\code{\link{gsv4_DimensionRange}} object. A range along a single dimension on a sheet.
All indexes are zero-based.
Indexes are half open: the start index is inclusive
and the end index is exclusive.
Missing indexes indicate the range is unbounded on that side.}

\item{fields}{string. The fields that should be updated.  At least one field must be specified.
The root `properties` is implied and should not be specified.
A single `"*"` can be used as short-hand for listing every field.}

\item{properties}{\code{\link{gsv4_DimensionProperties}} object. Properties about a dimension.}
}
\value{
UpdateDimensionPropertiesRequest
}
\description{
Updates properties of dimensions within the specified range.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#UpdateDimensionPropertiesRequest}{Google's Documentation for UpdateDimensionPropertiesRequest}
}

