% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_UpdateEmbeddedObjectPositionRequest}
\alias{gsv4_UpdateEmbeddedObjectPositionRequest}
\title{gsv4_UpdateEmbeddedObjectPositionRequest}
\usage{
gsv4_UpdateEmbeddedObjectPositionRequest(fields = NULL, newPosition = NULL,
  objectId = NULL)
}
\arguments{
\item{fields}{string. The fields of OverlayPosition
that should be updated when setting a new position. Used only if
newPosition.overlayPosition
is set, in which case at least one field must
be specified.  The root `newPosition.overlayPosition` is implied and
should not be specified.
A single `"*"` can be used as short-hand for listing every field.}

\item{newPosition}{\code{\link{gsv4_EmbeddedObjectPosition}} object. The position of an embedded object such as a chart.}

\item{objectId}{integer. The ID of the object to moved.}
}
\value{
UpdateEmbeddedObjectPositionRequest
}
\description{
Update an embedded object's position (such as a moving or resizing a
chart or image).
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#UpdateEmbeddedObjectPositionRequest}{Google's Documentation for UpdateEmbeddedObjectPositionRequest}
}

