% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_UpdateFilterViewRequest}
\alias{gsv4_UpdateFilterViewRequest}
\title{gsv4_UpdateFilterViewRequest}
\usage{
gsv4_UpdateFilterViewRequest(fields=NULL, filter=NULL)
}
\arguments{
\item{fields}{string. The fields that should be updated.  At least one field must be specified.
The root `filter` is implied and should not be specified.
A single `"*"` can be used as short-hand for listing every field.}

\item{filter}{\code{\link{gsv4_FilterView}} object. A filter view.}
}
\value{
UpdateFilterViewRequest
}
\description{
Updates properties of the filter view.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#UpdateFilterViewRequest}{Google's Documentation for UpdateFilterViewRequest}
}

