% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_UpdateNamedRangeRequest}
\alias{gsv4_UpdateNamedRangeRequest}
\title{gsv4_UpdateNamedRangeRequest}
\usage{
gsv4_UpdateNamedRangeRequest(fields = NULL, namedRange = NULL)
}
\arguments{
\item{fields}{string. The fields that should be updated.  At least one field must be specified.
The root `namedRange` is implied and should not be specified.
A single `"*"` can be used as short-hand for listing every field.}

\item{namedRange}{\code{\link{gsv4_NamedRange}} object. A named range.}
}
\value{
UpdateNamedRangeRequest
}
\description{
Updates properties of the named range with the specified
namedRangeId.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#UpdateNamedRangeRequest}{Google's Documentation for UpdateNamedRangeRequest}
}

