% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_UpdateProtectedRangeRequest}
\alias{gsv4_UpdateProtectedRangeRequest}
\title{gsv4_UpdateProtectedRangeRequest}
\usage{
gsv4_UpdateProtectedRangeRequest(fields=NULL, protectedRange=NULL)
}
\arguments{
\item{fields}{string. The fields that should be updated.  At least one field must be specified.
The root `protectedRange` is implied and should not be specified.
A single `"*"` can be used as short-hand for listing every field.}

\item{protectedRange}{\code{\link{gsv4_ProtectedRange}} object. A protected range.}
}
\value{
UpdateProtectedRangeRequest
}
\description{
Updates an existing protected range with the specified
protectedRangeId.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#UpdateProtectedRangeRequest}{Google's Documentation for UpdateProtectedRangeRequest}
}

