% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_UpdateSheetPropertiesRequest}
\alias{gsv4_UpdateSheetPropertiesRequest}
\title{gsv4_UpdateSheetPropertiesRequest}
\usage{
gsv4_UpdateSheetPropertiesRequest(fields=NULL, properties=NULL)
}
\arguments{
\item{fields}{string. The fields that should be updated.  At least one field must be specified.
The root `properties` is implied and should not be specified.
A single `"*"` can be used as short-hand for listing every field.}

\item{properties}{\code{\link{gsv4_SheetProperties}} object. Properties of a sheet.}
}
\value{
UpdateSheetPropertiesRequest
}
\description{
Updates properties of the sheet with the specified
sheetId.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#UpdateSheetPropertiesRequest}{Google's Documentation for UpdateSheetPropertiesRequest}
}

