% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_UpdateSpreadsheetPropertiesRequest}
\alias{gsv4_UpdateSpreadsheetPropertiesRequest}
\title{gsv4_UpdateSpreadsheetPropertiesRequest}
\usage{
gsv4_UpdateSpreadsheetPropertiesRequest(fields=NULL, properties=NULL)
}
\arguments{
\item{fields}{string. The fields that should be updated.  At least one field must be specified.
The root 'properties' is implied and should not be specified.
A single `"*"` can be used as short-hand for listing every field.}

\item{properties}{\code{\link{gsv4_SpreadsheetProperties}} object. Properties of a spreadsheet.}
}
\value{
UpdateSpreadsheetPropertiesRequest
}
\description{
Updates properties of a spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets#UpdateSpreadsheetPropertiesRequest}{Google's Documentation for UpdateSpreadsheetPropertiesRequest}
}

