% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsv4_object.R
\name{gsv4_ValueRange}
\alias{gsv4_ValueRange}
\title{gsv4_ValueRange}
\usage{
gsv4_ValueRange(range=NULL, majorDimension=NULL, values=NULL)
}
\arguments{
\item{range}{string. The range the values cover, in A1 notation.
For output, this range indicates the entire requested range,
even though the values will exclude trailing rows and columns.
When appending values, this field represents the range to search for a
table, after which values will be appended.}

\item{majorDimension}{string. The major dimension of the values.

For output, if the spreadsheet data is: `A1=1,B1=2,A2=3,B2=4`,
then requesting `range=A1:B2,majorDimension=ROWS` will return
`[[1,2],[3,4]]`,
whereas requesting `range=A1:B2,majorDimension=COLUMNS` will return
`[[1,3],[2,4]]`.

For input, with `range=A1:B2,majorDimension=ROWS` then `[[1,2],[3,4]]`
will set `A1=1,B1=2,A2=3,B2=4`. With `range=A1:B2,majorDimension=COLUMNS`
then `[[1,2],[3,4]]` will set `A1=1,B1=3,A2=2,B2=4`.

When writing, if this field is not set, it defaults to ROWS. majorDimension must take one of the following values: DIMENSION_UNSPECIFIED, ROWS, COLUMNS
See the details section for the definition of each of these values.}

\item{values}{list. The data that was read or to be written.  This is an array of arrays,
the outer array representing all the data and each inner array
representing a major dimension. Each item in the inner array
corresponds with one cell.

For output, empty trailing rows and columns will not be included.

For input, supported value types are: bool, string, and double.
Null values will be skipped.
To set a cell to an empty value, set the string value to an empty string.}
}
\value{
ValueRange
}
\description{
Data within a range of the spreadsheet.
}
\details{
majorDimension takes one of the following values:
\itemize{
 \item{DIMENSION_UNSPECIFIED - The default value, do not use.}
 \item{ROWS - Operates on the rows of a sheet.}
 \item{COLUMNS - Operates on the columns of a sheet.}
}
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets.values#ValueRange}{Google's Documentation for ValueRange}
}

