% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheets.values.R
\name{gsv4_values_append}
\alias{gsv4_values_append}
\title{gsv4_values_append}
\usage{
gsv4_values_append(spreadsheetId, range, valueInputOption,
  includeValuesInResponse = NULL, insertDataOption = NULL,
  responseDateTimeRenderOption = NULL, responseValueRenderOption = NULL,
  input, standard_params = list(...), ...)
}
\arguments{
\item{spreadsheetId}{string (required). The ID of the spreadsheet to update.}

\item{range}{string (required). The A1 notation of a range to search for a logical table of data.
Values will be appended after the last row of the table.}

\item{valueInputOption}{string (required). How the input data should be interpreted.}

\item{includeValuesInResponse}{logical. Determines if the update response should include the values
of the cells that were appended. By default, responses
do not include the updated values.}

\item{insertDataOption}{string. How the input data should be inserted.}

\item{responseDateTimeRenderOption}{string. Determines how dates, times, and durations in the response should be
rendered. This is ignored if response_value_render_option is
FORMATTED_VALUE.
The default dateTime render option is [DateTimeRenderOption.SERIAL_NUMBER].}

\item{responseValueRenderOption}{string. Determines how values in the response should be rendered.
The default render option is ValueRenderOption.FORMATTED_VALUE.}

\item{input}{\code{\link{gsv4_ValueRange}}. Data within a range of the spreadsheet.}

\item{standard_params}{a list of parameters for controlling the HTTP request and its response.
Refer to \code{\link{gsv4_standard_parameters}} for details on its arguments}

\item{...}{arguments to be used to form standard_params argument if it is not supplied directly.}
}
\value{
AppendValuesResponse
}
\description{
Appends values to a spreadsheet. The input range is used to search for
existing data and find a "table" within that range. Values will be
appended to the next row of the table, starting with the first column of
the table. See the
\href{https://developers.google.com/sheets/api/guides/values#appending_values}{guide}
and
\href{https://developers.google.com/sheets/api/samples/writing#append_values}{sample code}
for specific details of how tables are detected and data is appended.
}
\details{
The caller must specify the spreadsheet ID, range, and
a valueInputOption.  The `valueInputOption` only
controls how the input data will be added to the sheet (column-wise or
row-wise), it does not influence what cell the data starts being written
to.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets.values/append}{Google's Documentation for append}
}

