% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheets.values.R
\name{gsv4_values_batchClear}
\alias{gsv4_values_batchClear}
\title{gsv4_values_batchClear}
\usage{
gsv4_values_batchClear(spreadsheetId, input, standard_params = list(...), ...)
}
\arguments{
\item{spreadsheetId}{string (required). The ID of the spreadsheet to update.}

\item{input}{\code{\link{gsv4_BatchClearValuesRequest}}. The request for clearing more than one range of values in a spreadsheet.}

\item{standard_params}{a list of parameters for controlling the HTTP request and its response.
Refer to \code{\link{gsv4_standard_parameters}} for details on its arguments}

\item{...}{arguments to be used to form standard_params argument if it is not supplied directly.}
}
\value{
BatchClearValuesResponse
}
\description{
Clears one or more ranges of values from a spreadsheet.
The caller must specify the spreadsheet ID and one or more ranges.
Only values are cleared -- all other properties of the cell (such as
formatting, data validation, etc..) are kept.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets.values/batchClear}{Google's Documentation for batchClear}
}

