% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheets.values.R
\name{gsv4_values_batchGet}
\alias{gsv4_values_batchGet}
\title{gsv4_values_batchGet}
\usage{
gsv4_values_batchGet(spreadsheetId, dateTimeRenderOption=NULL, majorDimension=NULL, ranges=NULL, valueRenderOption=NULL, standard_params = list(...), ...)
}
\arguments{
\item{spreadsheetId}{string (required). The ID of the spreadsheet to retrieve data from.}

\item{dateTimeRenderOption}{string. How dates, times, and durations should be represented in the output.
This is ignored if value_render_option is
FORMATTED_VALUE.
The default dateTime render option is [DateTimeRenderOption.SERIAL_NUMBER].}

\item{majorDimension}{string. The major dimension that results should use.

For example, if the spreadsheet data is: `A1=1,B1=2,A2=3,B2=4`,
then requesting `range=A1:B2,majorDimension=ROWS` will return
`[[1,2],[3,4]]`,
whereas requesting `range=A1:B2,majorDimension=COLUMNS` will return
`[[1,3],[2,4]]`.}

\item{ranges}{string (repeated). The A1 notation of the values to retrieve.}

\item{valueRenderOption}{string. How values should be represented in the output.
The default render option is ValueRenderOption.FORMATTED_VALUE.}

\item{standard_params}{a list of parameters for controlling the HTTP request and its response.
Refer to \code{\link{gsv4_standard_parameters}} for details on its arguments}

\item{...}{arguments to be used to form standard_params argument if it is not supplied directly.}
}
\value{
BatchGetValuesResponse
}
\description{
Returns one or more ranges of values from a spreadsheet.
The caller must specify the spreadsheet ID and one or more ranges.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets.values/batchGet}{Google's Documentation for batchGet}
}

