% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheets.values.R
\name{gsv4_values_batchUpdate}
\alias{gsv4_values_batchUpdate}
\title{gsv4_values_batchUpdate}
\usage{
gsv4_values_batchUpdate(spreadsheetId, input, standard_params = list(...), ...)
}
\arguments{
\item{spreadsheetId}{string (required). The ID of the spreadsheet to update.}

\item{input}{\code{\link{gsv4_BatchUpdateValuesRequest}}. The request for updating more than one range of values in a spreadsheet.}

\item{standard_params}{a list of parameters for controlling the HTTP request and its response.
Refer to \code{\link{gsv4_standard_parameters}} for details on its arguments}

\item{...}{arguments to be used to form standard_params argument if it is not supplied directly.}
}
\value{
BatchUpdateValuesResponse
}
\description{
Sets values in one or more ranges of a spreadsheet.
The caller must specify the spreadsheet ID,
a valueInputOption, and one or more
ValueRanges.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets.values/batchUpdate}{Google's Documentation for batchUpdate}
}

