% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheets.values.R
\name{gsv4_values_clear}
\alias{gsv4_values_clear}
\title{gsv4_values_clear}
\usage{
gsv4_values_clear(spreadsheetId, range, input, standard_params = list(...), ...)
}
\arguments{
\item{spreadsheetId}{string (required). The ID of the spreadsheet to update.}

\item{range}{string (required). The A1 notation of the values to clear.}

\item{input}{\code{\link{gsv4_ClearValuesRequest}}. The request for clearing a range of values in a spreadsheet.}

\item{standard_params}{a list of parameters for controlling the HTTP request and its response.
Refer to \code{\link{gsv4_standard_parameters}} for details on its arguments}

\item{...}{arguments to be used to form standard_params argument if it is not supplied directly.}
}
\value{
ClearValuesResponse
}
\description{
Clears values from a spreadsheet.
The caller must specify the spreadsheet ID and range.
Only values are cleared -- all other properties of the cell (such as
formatting, data validation, etc..) are kept.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets.values/clear}{Google's Documentation for clear}
}

