% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheets.values.R
\name{gsv4_values_update}
\alias{gsv4_values_update}
\title{gsv4_values_update}
\usage{
gsv4_values_update(spreadsheetId, range, valueInputOption, includeValuesInResponse=NULL, responseDateTimeRenderOption=NULL, responseValueRenderOption=NULL, input, standard_params = list(...), ...)
}
\arguments{
\item{spreadsheetId}{string (required). The ID of the spreadsheet to update.}

\item{range}{string (required). The A1 notation of the values to update.}

\item{valueInputOption}{string (required). How the input data should be interpreted.}

\item{includeValuesInResponse}{logical. Determines if the update response should include the values
of the cells that were updated. By default, responses
do not include the updated values.
If the range to write was larger than than the range actually written,
the response will include all values in the requested range (excluding
trailing empty rows and columns).}

\item{responseDateTimeRenderOption}{string. Determines how dates, times, and durations in the response should be
rendered. This is ignored if response_value_render_option is
FORMATTED_VALUE.
The default dateTime render option is [DateTimeRenderOption.SERIAL_NUMBER].}

\item{responseValueRenderOption}{string. Determines how values in the response should be rendered.
The default render option is ValueRenderOption.FORMATTED_VALUE.}

\item{input}{\code{\link{gsv4_ValueRange}}. Data within a range of the spreadsheet.}

\item{standard_params}{a list of parameters for controlling the HTTP request and its response.
Refer to \code{\link{gsv4_standard_parameters}} for details on its arguments}

\item{...}{arguments to be used to form standard_params argument if it is not supplied directly.}
}
\value{
UpdateValuesResponse
}
\description{
Sets values in a range of a spreadsheet.
The caller must specify the spreadsheet ID, range, and
a valueInputOption.
}
\seealso{
\href{https://developers.google.com/sheets/reference/rest/v4/spreadsheets.values/update}{Google's Documentation for update}
}

