% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/register.R
\name{identify_sheet}
\alias{identify_sheet}
\title{Retrieve the identifiers for a spreadsheet}
\usage{
identify_sheet(x, verbose = TRUE)
}
\arguments{
\item{x}{character vector of length one, with sheet-identifying
information; it will be considered as URL, sheet title or key, until one of
those hopefully make sense and uniquely identifies sheet}

\item{verbose}{logical}
}
\value{
a list with information about the sheet
}
\description{
Retrieve a list with all the identifying information for a specific
spreadsheet. It calls \code{\link{list_sheets}} and attempts to return the
row uniquely specified by input \code{x}. The listing provided by
\code{\link{list_sheets}} is only available to an authorized user, so
authorization will be required.
}

