% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{is_legit_token}
\alias{is_legit_token}
\title{Check that token appears to be legitimate}
\usage{
is_legit_token(x, verbose = FALSE)
}
\description{
This unexported function exists to catch tokens that are technically valid,
i.e. `inherits(token, "Token2.0")` is TRUE, but that have dysfunctional
credentials.
}
\keyword{internal}

