% Generated by roxygen2 (4.0.2): do not edit by hand
\name{label_to_coord}
\alias{label_to_coord}
\title{Convert label (A1) notation to coordinate (R1C1) notation}
\usage{
label_to_coord(x)
}
\arguments{
\item{x}{label notation for position of cell}
}
\description{
A1 and R1C1 are equivalent addresses for position of cells.
}

