% Generated by roxygen2 (4.0.2): do not edit by hand
\name{letter_to_col}
\alias{letter_to_col}
\title{Convert column letter to number}
\usage{
letter_to_col(x)
}
\arguments{
\item{x}{column letter}
}
\description{
Take first row of data frame and make it the header. Rownames are also reset.
}
\examples{
letter_to_col("A")
letter_to_col("AB")
}

