% Generated by roxygen2 (4.0.2): do not edit by hand
\name{letter_to_num}
\alias{letter_to_num}
\title{Convert column letter to number}
\usage{
letter_to_num(x)
}
\arguments{
\item{x}{column letter (case insensitive)}
}
\description{
Convert column letter to number
}
\examples{
letter_to_num("A")
letter_to_num("AB")
letter_to_num("ab")
}

