% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{lfilt}
\alias{lfilt}
\title{Filter a list by name}
\usage{
lfilt(x, name, ...)
}
\arguments{
\item{x}{a list}

\item{name}{a regular expression}

\item{...}{other parameters you might want to pass to grep}
}
\description{
Filter a list by name
}
\keyword{internal}

