% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permissions.R
\name{list_perm}
\alias{list_perm}
\title{Get a listing of permissions for a spreadsheet}
\usage{
list_perm(ss)
}
\arguments{
\item{ss}{a registered Google spreadsheet}
}
\value{
a tbl_df, one row per permission
}
\description{
This function returns the information available from the
\href{https://developers.google.com/drive/v2/reference/permissions}{permissions
feed} of the Google Drive API.
}
\details{
This listing gives the user all the permissions for a spreadsheet. A
simplified viewing is available in the sharing dialog of a Google Sheet.

The first row will always be the permission for the owner of the spreadsheet.
Any additional permissions follow.

A permission for a sheet includes the following information: the name for
this permission, the email of the user or group the permission refers to, the
primary role for the user, any additional roles they have, the type of user
they are, the (unique) ID of the user this permission refers to, a link
back to this permission, and the ETag of the permission.
}
\examples{
\dontrun{
foo <- new_ss("foo")
list_perm(foo)
}
}

