% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/register.R
\name{list_spreadsheets}
\alias{list_spreadsheets}
\title{Get a listing of available spreadsheets}
\usage{
list_spreadsheets(include_key = TRUE)
}
\arguments{
\item{include_key}{\code{logical} to control inclusion of spreadsheet keys}
}
\description{
For all spreadsheets available to the authenticated user, get spreadsheet
title, owner, user's permission, date-time of last update and, optionally,
the unique key. Should return the spreadsheets that are visible in Google
Drive. Return as a data.frame, one row per spreadsheet.
}

