% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_worksheet_objs}
\alias{list_worksheet_objs}
\title{Get a list of worksheet objects}
\usage{
list_worksheet_objs(ss)
}
\arguments{
\item{ss}{a spreadsheet object returned by \code{\link{open_spreadsheet}}}
}
\description{
The returned list of worksheet objects enables \code{plyr} functions to
operate on mutiple worksheets at once.
}

