% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_worksheets}
\alias{list_worksheets}
\title{Get the titles of the worksheets contained in spreadsheet.}
\usage{
list_worksheets(x)
}
\arguments{
\item{x}{A spreadsheet object returned by \code{\link{open_spreadsheet}}}
}
\value{
The titles of worksheets contained in spreadsheet.

This is a mini wrapper for x$ws_names.
}
\description{
Retrieve list of worksheet titles contained in spreadsheet
}

