% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{llpluck}
\alias{llpluck}
\title{Pluck out elements from list components by name}
\usage{
llpluck(x, xpath)
}
\arguments{
\item{x}{a list}

\item{xpath}{a string giving the name of the component you want, XPath style}
}
\description{
Pluck out elements from list components by name
}
\keyword{internal}

