% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{modify_ws}
\alias{modify_ws}
\title{Modify a worksheet's title or size}
\usage{
modify_ws(ss, ws_title, new_title = NULL, new_dim = NULL)
}
\arguments{
\item{ss}{a registered Google sheet}

\item{ws_title}{character string for title of worksheet}

\item{new_title}{character string for worksheet's new title}

\item{new_dim}{list of length 2 specifying the row and column extent of the worksheet}
}
\description{
Modify a worksheet's title or size
}

