% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ncells}
\alias{ncells}
\title{Determine the number of cells in the range}
\usage{
ncells(range)
}
\arguments{
\item{range}{character string for range value}
}
\value{
numeric value for count of cells in range.
}
\description{
Determine the number of cells in the range
}

