% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{new_sheet}
\alias{new_sheet}
\title{Create a new spreadsheet}
\usage{
new_sheet(title = "my_sheet", verbose = TRUE)
}
\arguments{
\item{title}{the title for the new sheet}

\item{verbose}{logical; do you want informative message?}
}
\description{
Create a new (empty) spreadsheet in your Google Drive. The new sheet will
contain 1 default worksheet titled "Sheet1".
}

