% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{new_ss}
\alias{new_ss}
\title{Create a new spreadsheet}
\usage{
new_ss(title)
}
\arguments{
\item{title}{the title for the new spreadsheet}
}
\description{
Create a new (empty) spreadsheet in your Google Drive. The new spreadsheet
will contain 1 default worksheet titled "Sheet1".
}

