% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{new_ws}
\alias{new_ws}
\title{Add a new (empty) worksheet to spreadsheet}
\usage{
new_ws(ss, ws_title, nrow = 1000, ncol = 26)
}
\arguments{
\item{ss}{a registered Google spreadsheet}

\item{ws_title}{character string for title of new worksheet}

\item{nrow}{number of rows (default is 1000)}

\item{ncol}{number of columns (default is 26)}
}
\description{
Add a new (empty) worksheet to a spreadsheet, specify the title, worksheet
extent (number of rows and columns). The title of the new worksheet can not
be the same as any of the existing worksheets in the spreadsheet.
}

