% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{open_at_once}
\alias{open_at_once}
\title{Open a worksheet from a spreadsheet in one go}
\usage{
open_at_once(ss_title, ws_value)
}
\arguments{
\item{ss_title}{spreadsheet title}

\item{ws_value}{title or numeric index of worksheet}
}
\description{
Open a worksheet from a spreadsheet in one go
}

