% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{open_by_key}
\alias{open_by_key}
\title{Open spreadsheet by key}
\usage{
open_by_key(key, visibility = "private")
}
\arguments{
\item{key}{the key of a spreadsheet as it appears in browser URL.}

\item{visibility}{either "public" for public spreadsheets or "private"
for private spreadsheets}
}
\value{
Object of class spreadsheet.

This function only works for keys of public spreadsheets.
}
\description{
Use key found in browser URL and return an object of class spreadsheet.
}

