% Generated by roxygen2 (4.0.2): do not edit by hand
\name{open_by_url}
\alias{open_by_url}
\title{Open spreadsheet by url}
\usage{
open_by_url(url)
}
\arguments{
\item{url}{URL of spreadsheet as it appears in browser}
}
\value{
Object of class spreadsheet.

This function currently only works for public spreadsheets.
This function extracts the key from the url and calls on open_by_key().
}
\description{
Use url of spreadsheet and return an object of class spreadsheet.
}

