% Generated by roxygen2 (4.0.2): do not edit by hand
\name{open_sheet}
\alias{open_sheet}
\title{Open spreadsheet by title}
\usage{
open_sheet(client, title)
}
\arguments{
\item{client}{Client object returned by login(email, passwd)}

\item{title}{Name of spreadsheet}
}
\value{
Object of class spreadsheet.
}
\description{
Use key found in browser URL and return an object of class spreadsheet.
}

