% Generated by roxygen2 (4.0.2): do not edit by hand
\name{open_spreadsheet}
\alias{open_spreadsheet}
\title{Open spreadsheet by title}
\usage{
open_spreadsheet(title)
}
\arguments{
\item{title}{the title of a spreadsheet.}
}
\description{
Use the spreadsheet title (as it appears in Google Drive) to get a
spreadsheet object, containing the spreadsheet id, title, time of last
update, the titles and number of worksheets contained, and a list of
worksheet objects.
}

