% Generated by roxygen2 (4.0.2): do not edit by hand
\name{open_worksheet}
\alias{open_worksheet}
\title{Open worksheet given worksheet title or index}
\usage{
open_worksheet(ss, x, vis = "private")
}
\arguments{
\item{ss}{Spreadsheet object containing worksheet}

\item{x}{chr string for title of worksheet or numeric for index of worksheet
in spreadsheet}
}
\value{
An object of class worksheet and number of rows and cols attribute.
}
\description{
Use title or index of worksheet to retrieve object of class worksheet.
}

