% Generated by roxygen2 (4.0.2): do not edit by hand
\name{open_worksheets}
\alias{open_worksheets}
\title{Get a list of worksheet objects}
\usage{
open_worksheets(ss)
}
\arguments{
\item{ss}{a spreadsheet object returned by \code{\link{open_spreadsheet}}}
}
\description{
Return a list of all the worksheets in a spreadsheet as worksheet objects so
that \code{plyr} functions can be used to perform worksheet operations on
multiple worksheets at once.
}

