% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheet-print.R
\name{print.googlesheet}
\alias{print.googlesheet}
\title{Print info about a \code{googlesheet} object}
\usage{
\method{print}{googlesheet}(x, ...)
}
\arguments{
\item{x}{\code{\link{googlesheet}} object returned by functions such as \code{\link{gs_title}}, \code{\link{gs_key}}, and friends}

\item{...}{potential further arguments (required for Method/Generic reasons)}
}
\description{
Display information about a Google spreadsheet that has been registered with
\code{googlesheets}: the title of the spreadsheet, date-time of registration,
date-time of last update (at time of registration), visibility, permissions,
version, the number of worksheets contained, worksheet titles and extent, and
sheet key.
}
\examples{
\dontrun{
foo <- gs_new("foo")
foo
print(foo)
}

}

