% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/print.R
\name{print.gspreadsheet}
\alias{print.gspreadsheet}
\title{Print information about a Google spreadsheet registered with gspreadr}
\usage{
\method{print}{gspreadsheet}(x, ...)
}
\arguments{
\item{x}{gspreadsheet object returned by \code{register_ss} and other
\code{gspreadr} functions}

\item{...}{potential further arguments (required for Method/Generic reasons)}
}
\description{
Display information about a Google spreadsheet that has been registered with
\code{gspreadr}: the title of the spreadsheet, date-time of registration,
date-time of last update (at time of registration), the number of worksheets
contained, worksheet titles and extent, and sheet key.
}
\examples{
\dontrun{
foo <- new_ss("foo")
foo
print(foo)
}
}

