% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/inspect.R
\name{print.spreadsheet}
\alias{print.spreadsheet}
\title{Print method for a spreadsheet object}
\usage{
\method{print}{spreadsheet}(x, ...)
}
\arguments{
\item{x}{spreadsheet object}

\item{...}{potential further arguments}
}
\description{
Print method for a spreadsheet object
}

