% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/inspect.R
\name{print.worksheet}
\alias{print.worksheet}
\title{Print method for a worksheet object}
\usage{
\method{print}{worksheet}(x, ...)
}
\arguments{
\item{x}{worksheet object}

\item{...}{potential further arguments}
}
\description{
Print method for a worksheet object
}

