% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cell-specification.R
\name{rc}
\alias{rc}
\title{Convert a cell range, special case wrapper for R1C1 positioning notation}
\usage{
rc(range)
}
\arguments{
\item{range}{character vector, length one, such as "R1C1:R7C4" or "R3C3"}
}
\value{
list with 4 positive integer elements named \code{min-row},
  \code{max-row}, \code{min-col}, \code{max-col}
}
\description{
Convert a cell range, special case wrapper for R1C1 positioning notation
}
\examples{
\dontrun{
rc("R1C1:R7C4")
rc("R3C3")
}
}
\keyword{internal}

