% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{read_all}
\alias{read_all}
\title{Get all values in a worksheet.}
\usage{
read_all(ws, header = TRUE)
}
\arguments{
\item{ws}{worksheet object}

\item{header}{\code{logical} to indicate if the first row should be taken as the
header}
}
\value{
A dataframe.

This function calls on \code{\link{get_cols}} with \code{to} set as the
number of columns of the worksheet.
}
\description{
Extract the entire worksheet and turn it into a data frame. This function
uses the rightmost cell with a value as the maximum number of columns and
bottom-most cell as the maximum row.
}
\seealso{
\code{\link{read_region}}, \code{\link{read_range}},
\code{\link{get_row}}, \code{\link{get_rows}}, \code{\link{get_col}},
\code{\link{get_cols}}
}

