% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_range}
\alias{read_range}
\title{Get a region of a worksheet by range}
\usage{
read_range(ws, x, header = TRUE, vis = "private")
}
\arguments{
\item{ws}{Worksheet object}

\item{x}{character string for range separated by ":"}

\item{header}{\code{logical} for whether or not first row should be taken as
header}

\item{vis}{either "private" (default) or "public" for public spreadsheets}
}
\description{
Get a region of a worksheet by range
}
\examples{
read_range("A1:B10")
read_range("C10:D20")
}

