% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/register.R
\name{register_ss}
\alias{register_ss}
\title{Register a Google spreadsheet}
\usage{
register_ss(x, key = NULL, ws_feed = NULL, visibility = "private",
  verbose = TRUE)
}
\arguments{
\item{x}{character vector of length one, with sheet-identifying information;
valid inputs are title, key, URL, worksheets feed}

\item{key}{character vector of length one that is guaranteed to be unique key
for sheet; supercedes argument \code{x}}

\item{ws_feed}{character vector of length one that is guaranteed to be
worksheets feed for sheet; supercedes arguments \code{x} and \code{key}}

\item{visibility}{either "public" or "private"; used to specify visibility
when sheet identified via \code{key}}

\item{verbose}{logical; do you want informative message?}
}
\value{
Object of class googlesheet.
}
\description{
Specify a Google spreadsheet via its URL, unique key, title, or worksheets
feed and register it for further use. This function returns an object of
class \code{googlesheet}, which contains all the information other
\code{googlesheets} functions will need to consume data from the sheet or to
edit the sheet. This object also contains sheet information that may be of
interest to the user, such as the time of last update, the number of
worksheets contained, and their titles.
}
\note{
Re: the reported extent of the worksheets. Contain your excitement,
  because it may not be what you think or hope it is. It does not report how
  many rows or columns are actually nonempty. This cannot be determined via
  the Google sheets API without consuming the data and noting which cells are
  populated. Therefore, these numbers often reflect the default extent of a
  new worksheet, e.g., 1000 rows and 26 columns at the time or writing, and
  provide an upper bound on the true number of rows and columns.

The visibility can only be "public" if the sheet is "Published to the
  web". Gotcha: this is different from setting the sheet to "Public on the
  web" in the visibility options in the sharing dialog of a Google Sheets
  file.
}
\examples{
\dontrun{
gap_key <- "1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA"
gap_ss <- register_ss(gap_key)
gap_ss
get_row(gap_ss, "Africa", row = 1)
}
}

