% Generated by roxygen2 (4.0.2): do not edit by hand
\name{report_ss}
\alias{report_ss}
\title{Get report for a spreadsheet}
\usage{
report_ss(ss)
}
\arguments{
\item{ss}{spreadsheet object}
}
\value{
A list containing the name of the spreadsheet and the number of
worksheets it contains and worksheet dimensions.
}
\description{
Generate a report for a spreadsheet.
}

