% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/consume-data.R
\name{reshape_cf}
\alias{reshape_cf}
\title{Reshape cell-level data and convert to data.frame}
\usage{
reshape_cf(x, header = TRUE)
}
\arguments{
\item{x}{a data.frame returned by \code{get_via_cf()}}

\item{header}{logical indicating whether first row should be taken as
  variable names}
}
\description{
This will not be exported long-term. Will write wrappers. Temporary export.
}

