% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{resize_worksheet}
\alias{resize_worksheet}
\title{Resize a worksheet}
\usage{
resize_worksheet(ws, nrow = NULL, ncol = NULL)
}
\arguments{
\item{ws}{worksheet object}

\item{nrow}{\code{numeric} for new row dimension}

\item{ncol}{\code{numeric} for new column dimension}
}
\description{
Resize a worksheet
}
\note{
Setting rows and columns to less than the current worksheet dimensions
will delete contents without warning.
}

