% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{resize_ws}
\alias{resize_ws}
\title{Resize a worksheet}
\usage{
resize_ws(ss, ws_title, row_extent = NULL, col_extent = NULL,
  verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google sheet}

\item{ws_title}{character string for title of worksheet}

\item{row_extent}{integer for new row extent}

\item{col_extent}{integer for new column extent}

\item{verbose}{logical; do you want informative message?}
}
\description{
Set the number of rows and columns of a worksheet. This function is useful
when you need to send a batch update request and the data would exceed
the current worksheet extent.
}
\details{
This function will probably only be called internally... exporting for now.
}
\note{
Setting rows and columns to less than the current worksheet dimensions
will delete contents without warning.
}

