% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_header}
\alias{set_header}
\title{Set header for data frame}
\usage{
set_header(x)
}
\arguments{
\item{x}{a data frame.}
}
\description{
Take first row of data frame and make it the header. Rownames are also reset.
}

