% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny-with-gs.R
\name{shiny_get_url}
\alias{shiny_get_url}
\title{Form the URL to use to authenticate a user}
\usage{
shiny_get_url(client_id, client_secret, redirect_uri)
}
\arguments{
\item{client_id}{application client id}

\item{redirect_uri}{redirect uri (shiny app url or localhost for local testing)}

\item{client}{secret application client secret}
}
\description{
Once a user authenticates, the response is sent to the \code{redirect_uri}
and contains an authorization code which \code{shiny_get_token} uses to
exchange for an access token.
}

