% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{shiny_token_exists}
\alias{shiny_token_exists}
\title{Check if authorization was done through shiny app}
\usage{
shiny_token_exists()
}
\value{
logical
}
\description{
Check if authorization was done through shiny app
}
\keyword{internal}

