% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume-data.R
\name{simplify_cf}
\alias{simplify_cf}
\title{Simplify data from the cell feed}
\usage{
simplify_cf(x, convert = TRUE, as.is = TRUE, notation = c("A1", "R1C1"),
  header = NULL)
}
\arguments{
\item{x}{a data.frame returned by \code{get_via_cf()}}

\item{convert}{logical, indicating whether to attempt to convert the result
vector from character to something more appropriate, such as logical,
integer, or numeric; if TRUE, result is passed through \code{type.convert};
if FALSE, result will be character}

\item{as.is}{logical, passed through to the \code{as.is} argument of
\code{type.convert}}

\item{notation}{character; the result vector will have names that reflect
  which cell the data came from; this argument selects the positioning
  notation, i.e. "A1" vs. "R1C1"}

\item{header}{logical indicating whether first row should be taken as
  variable names}
}
\value{
a named vector
}
\description{
In some cases, you might not want to convert the data retrieved from the cell
feed into a data.frame via \code{\link{reshape_cf}}. You might prefer it as
an atomic vector. That's what this function does. Note that, unlike
\code{\link{reshape_cf}}, empty cells will NOT necessarily appear in this
result. By default, the API does not transmit data for these cells;
\code{googlesheets} inserts these cells in \code{\link{reshape_cf}} because
it is necessary to give the data rectangular shape. In contrast, empty cells
will only appear in the output of \code{simplify_cf} if they were already
present in the data from the cell feed, i.e. if the original call to
\code{\link{get_via_cf}} had argument \code{return_empty} set to \code{TRUE}.
}
\examples{
\dontrun{
gap_key <- "1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA"
gap_ss <- register_ss(gap_key)
get_row(gap_ss, row = 1)
simplify_cf(get_row(gap_ss, row = 1))
simplify_cf(get_row(gap_ss, row = 1), notation = "R1C1")
}
}
\seealso{
Other data.consumption.functions: \code{\link{get_cells}};
  \code{\link{get_col}}; \code{\link{get_row}};
  \code{\link{get_via_cf}}; \code{\link{get_via_csv}};
  \code{\link{get_via_lf}}; \code{\link{reshape_cf}}
}

