% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/helpers.R
\name{slaste}
\alias{slaste}
\title{paste with separator set to slash}
\usage{
slaste(...)
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors}
}
\description{
paste with separator set to slash, for use in building URLs
}

