% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/models.R
\name{spreadsheet}
\alias{spreadsheet}
\title{The spreadsheet object}
\usage{
spreadsheet()
}
\description{
The spreadsheet object stores information about a spreadsheet. It includes the fields:
}
\details{
\itemize{
  \item \code{sheet_id} the key of the spreadsheet
  \item \code{sheet_title} the title of the spreadsheet
  \item \code{updated} the time of last update
  \item \code{nsheets} the number of worksheets contained in the spreadsheet
  \item \code{ws_names} the names of the worksheets contained in the
  spreadsheet
  \item \code{visibility} visibility of spreadsheet, determines whether
  authorization is required for request
  \item \code{worksheets} a list of worksheet objects for every worksheet
  contained in the spreadsheet
}
}

