% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spreadsheets_GET}
\alias{spreadsheets_GET}
\title{Create GET request}
\usage{
spreadsheets_GET(feed_type, client = NULL, key = NULL, ws_id = NULL,
  visibility = "private", projection = "full")
}
\arguments{
\item{feed_type}{One of the following: spreadsheets, worksheets, list, cells}

\item{client}{Client object or NULL if accessing public worksheets feed}

\item{key}{Spreadsheet key}

\item{visibility}{Either private or public}

\item{projection}{Either full or basic}

\item{worksheet_id}{id of worksheet contained in spreadsheet}
}
\description{
Construct GET requests to Google Sheets API. This function handles the
construction of urls and making GET requests so urls arent lying around.
}
\examples{
spreadsheets_GET("spreadsheets", NULL,
"1WNUDoBbGsPccRkXlLqeUK9JUQNnqq2yvc9r7-cmEaZU", visibility = "public")
}

