% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spreadsheets_info}
\alias{spreadsheets_info}
\title{Get list of spreadsheets: name, last updated, and key}
\usage{
spreadsheets_info(client)
}
\arguments{
\item{client}{Client object returned by \code{\link{login}} or
\code{\link{authorize}}}
}
\value{
A dataframe containing name of spreadsheets, last updated, and uris for worksheets feed.
}
\description{
Retrieve dataframe of spreadsheets information
}

