% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/str.R
\name{str.spreadsheet}
\alias{str.spreadsheet}
\title{Overview of a Google spreadsheet registered with gspreadr}
\usage{
\method{str}{spreadsheet}(object, ...)
}
\arguments{
\item{object}{spreadsheet object returned by \code{register}}

\item{...}{potential further arguments (required for Method/Generic reasons)}
}
\description{
Display the structure of a spreadsheet that has been registered with
\code{gspreadr}: the title of the spreadsheet, the number of worksheets
contained and the corresponding worksheet dimensions.
}

