% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str.worksheet}
\alias{str.worksheet}
\title{Display the structure of a worksheet}
\usage{
\method{str}{worksheet}(object, ...)
}
\arguments{
\item{object}{worksheet object}

\item{...}{potential further arguments (required for Method/Generic reasons)}
}
\value{
Does not return anything, prints to the console.
}
\description{
Show the structure of a worksheet. Display the title, number of rows and
columns, and summary of what patterns each column contains (ie. the number
of missing cells before cells with values).
}

