% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{token_available}
\alias{token_available}
\title{Check token availability}
\usage{
token_available(verbose = TRUE)
}
\value{
logical
}
\description{
Check if a token is available in \code{\link{googlesheets}}' internal
\code{.state} environment.
}
\keyword{internal}
