% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{token_exists}
\alias{token_exists}
\title{Check if authorization currently in force}
\usage{
token_exists(verbose = TRUE)
}
\value{
logical
}
\description{
Check if authorization currently in force
}
\keyword{internal}

