% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-data.R
\name{update_cells}
\alias{update_cells}
\title{Update a cell or range of cells}
\usage{
update_cells(ss, ws, range, input, verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google sheet}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet to consume}

\item{range}{single character string specifying which cell or range of cells
to update; positioning notation can be either "A1" or "R1C1"; a single cell
can be updated, e.g. "B4" or "R4C2" or a rectangular range can be updated,
e.g. "B2:D4" or "R2C2:R4C4"}

\item{input}{either a vector of a data frame of values to update cells with}

\item{verbose}{logical; do you want informative message?}
}
\description{
Modify the contents of a cell or a range of cells by specifying a range or a
reference cell. If a reference cell is given, the range occupied by the input
will be calculated and updated correspondingly. Input can be a character
vector or a data frame.
}
\details{
Cells are updated by rows, hence specifying a reference cell and a vector of
values for input will result in the specified row being updated.
}
\note{
Data frame row names are ignored.
}

