% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{upload_ss}
\alias{upload_ss}
\title{Upload a file and convert it to a Google Sheet}
\usage{
upload_ss(file, sheet_title = NULL, verbose = TRUE)
}
\arguments{
\item{file}{the file to upload, if it does not contain the absolute path,
then the file is relative to the current working directory}

\item{sheet_title}{the title of the spreadsheet; optional,
if not specified then the name of the file will be used}

\item{verbose}{logical; do you want informative message?}
}
\description{
Google supports the following file types to be converted to a Google
spreadsheet: .xls, .xlsx, .csv, .tsv, .txt, .tab, .xlsm, .xlt, .xltx, .xltm,
.ods. The newly uploaded file will appear in the top level of your Google
Sheets home screen.
}

