% Generated by roxygen2 (4.0.2): do not edit by hand
\name{view_all}
\alias{view_all}
\title{View all your worksheets}
\usage{
view_all(ss)
}
\arguments{
\item{ss}{spreadsheet object}
}
\value{
a list of 2 ggplot objects
}
\description{
Get a view of all the worksheets contained in the spreadsheet. This function
returns a list of two ggplot objects, first is a gallery of all the
worksheets and the second is an overlay of all the worksheets.
}

