% Generated by roxygen2 (4.0.2): do not edit by hand
\name{worksheet}
\alias{worksheet}
\title{Worksheets}
\usage{
worksheet()
}
\value{
Object of class worksheet.

This function currently only works for public spreadsheets (visibility = TRUE and projection = FULL).
}
\description{
This function creates worksheet objects
}

