% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permissions.R
\name{add_perm}
\alias{add_perm}
\title{Add a permission to a spreadsheet}
\usage{
add_perm(ss, value = NULL, type = NULL, role = NULL, with_link = TRUE,
  send_email = TRUE, verbose = TRUE)
}
\arguments{
\item{value}{The email address or domain name for the entity.}

\item{type}{The account type. Allowed values "user", "group", "domain", or
"anyone".}

\item{role}{The primary role for this user. Allowed values are: "owner",
"writer", or "reader".}

\item{with_link}{boolean, whether the link is required for this permission}

\item{send_email}{logical, do you want to send notification emails when
sharing to users or groups?}

\item{verbose}{logical, do you want informative messages?}
}
\value{
Information about newly added permission.
}
\description{
An email will be sent automatically to the entity to notify them of the
permission.
}
\examples{
\dontrun{
foo <- new_ss("foo")

# Add anyone as a writer/reader:
add_perm(foo, value = NULL, type = "anyone", role = "writer")
add_perm(foo, value = NULL, type = "anyone", role = "reader")

}
}

