% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_spreadsheet}
\alias{add_spreadsheet}
\title{Create a new spreadsheet}
\usage{
add_spreadsheet(title)
}
\arguments{
\item{title}{the title for the new spreadsheet}
}
\description{
A new (empty) spreadsheet with 1 worksheet titled "Sheet1" (default) will be
created in your Google Drive.
}

