% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_spreadsheet}
\alias{add_spreadsheet}
\title{Add a new spreadsheet}
\usage{
add_spreadsheet(title)
}
\arguments{
\item{title}{the title for the new spreadsheet}
}
\description{
Create a new (empty) spreadsheet with 1 worksheet titled "Sheet1" (default)
in your Google Drive.
}

