% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_spreadsheet}
\alias{add_spreadsheet}
\title{Create a new spreadsheet}
\usage{
add_spreadsheet(title)
}
\arguments{
\item{title}{the title for the new spreadsheet}
}
\description{
Create a new (empty) spreadsheet in your Google Drive. The new spreadsheet
will contain 1 default worksheet titled "Sheet1".
}

