% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_worksheet}
\alias{add_worksheet}
\title{Add a new (empty) worksheet to spreadsheet}
\usage{
add_worksheet(ss, title, nrow, ncol)
}
\arguments{
\item{ss}{spreadsheet object}

\item{title}{character string for title of new worksheet}

\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{token}{Google token obtained from \code{\link{login}} or
\code{\link{authorize}}}
}
\description{
Add a new (empty) worksheet to spreadsheet, specify title, number of rows
and columns.
}

