% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_worksheet}
\alias{add_worksheet}
\title{Add new worksheet to spreadsheet}
\usage{
add_worksheet(client, sheet, name, n_row, n_col)
}
\arguments{
\item{client}{Client object}

\item{sheet}{Spreadsheet object}

\item{name}{character string for name of new worksheet}

\item{n_row}{Number of rows}

\item{n_col}{Number of columns}
}
\description{
Add a new worksheet to spreadsheet, specify name, number of rows and cols.
}

