% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_worksheet}
\alias{add_worksheet}
\title{Add new worksheet to spreadsheet}
\usage{
add_worksheet(client, ss, title, rows, cols)
}
\arguments{
\item{client}{Client object}

\item{ss}{Spreadsheet object}

\item{rows}{Number of rows}

\item{cols}{Number of columns}

\item{name}{character string for name of new worksheet}
}
\description{
Add a new worksheet to spreadsheet, specify name, number of rows and cols.
}

