% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit.R
\name{add_worksheet}
\alias{add_worksheet}
\title{Add a new (empty) worksheet to spreadsheet}
\usage{
add_worksheet(ss, title, nrow = 1000, ncol = 26)
}
\arguments{
\item{ss}{spreadsheet object}

\item{title}{character string for title of new worksheet}

\item{nrow}{number of rows (default is 1000)}

\item{ncol}{number of columns (default is 26)}
}
\description{
Add a new (empty) worksheet to spreadsheet, specify title, number of rows
and columns. The title of the new worksheet should not be the same as any of
the existing worksheets
}

