% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{add_ws}
\alias{add_ws}
\title{Add a new (empty) worksheet to spreadsheet}
\usage{
add_ws(ss, ws_title = "Sheet1", nrow = 1000, ncol = 26, verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google sheet}

\item{ws_title}{character string for title of new worksheet}

\item{nrow}{number of rows (default is 1000)}

\item{ncol}{number of columns (default is 26)}

\item{verbose}{logical; do you want informative message?}
}
\value{
a googlesheet object, resulting from re-registering the host
  spreadsheet after adding the new worksheet
}
\description{
Add a new (empty) worksheet to spreadsheet: specify title and worksheet
extent (number of rows and columns). The title of the new worksheet can not
be the same as any existing worksheets in the sheet.
}
\examples{
\dontrun{
# get a copy of the Gapminder spreadsheet
gap_key <- "1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA"
gap_ss <- copy_ss(key = gap_key, to = "Gapminder_copy")
gap_ss <- add_ws(gap_ss, ws_title = "Atlantis")
gap_ss
}
}

