% Generated by roxygen2 (4.0.2): do not edit by hand
\name{authorize}
\alias{authorize}
\title{Authorize user using Oauth2.0 Credentials}
\usage{
authorize(new_user = FALSE)
}
\arguments{
\item{new_user}{set to \code{TRUE} if you want to authenticate a different
google account}
}
\description{
User will be directed to web browser and asked to sign into their Google
account and grant gspreadr access permission to user data for Google
Spreadsheets and Google Drive. User credentials will be cached in .httr-oauth
in the current working directory.
}
\details{
This method follows the demo described at:
\url{https://github.com/hadley/httr/blob/master/demo/oauth2-google.r}
}

