% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_query}
\alias{build_query}
\title{Build query string for GET URL}
\usage{
build_query(min_row, max_row, min_col, max_col)
}
\arguments{
\item{min_row,max_row,min_col,max_col}{query parameters}
}
\description{
Form the query string to append to GET URL.
}

