% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_range}
\alias{build_range}
\title{Determine the range occupied by a data frame}
\usage{
build_range(dat, anchor, header)
}
\arguments{
\item{dat}{a data frame}

\item{anchor}{position of cell used as an "anchor"}

\item{header}{\code{logical} indicating whether the header of the data should
be included in the range}
}
\value{
character string for the range taken up by a data frame
}
\description{
Determine the range occupied by a data frame
}

