% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_req_url}
\alias{build_req_url}
\title{Build URL for GET requests}
\usage{
build_req_url(feed_type, key = NULL, ws_id = NULL, visibility = "private",
  projection = "full", min_row = NULL, max_row = NULL, min_col = NULL,
  max_col = NULL)
}
\arguments{
\item{feed_type}{one of the following: spreadsheets, worksheets, list, cells}

\item{key}{spreadsheet key}

\item{ws_id}{id of worksheet contained in spreadsheet}

\item{visibility}{either private or public}

\item{projection}{either full or basic}
}
\value{
URL
}
\description{
Construct URL for talking to Google Sheets API.
}

