% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_token}
\alias{check_token}
\title{Check Google token for validity}
\usage{
check_token(token)
}
\arguments{
\item{token}{Google authorization token}
}
\description{
Make sure Google token is good to use upon retrieval so error is found
before making requests.
}

