% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cell-specification.R
\name{convert_limit_list_to_range}
\alias{convert_limit_list_to_range}
\title{Convert a limits list to a cell range}
\usage{
convert_limit_list_to_range(limits, rc = FALSE)
}
\arguments{
\item{limits}{limits list}

\item{rc}{logical, indicating "R1C1" positioning notation}
}
\value{
length one character vector holding a cell range, in either A1 or
  R1C1 positioning notation
}
\description{
Convert a limits list to a cell range
}
\examples{
rgLL <- list(`min-row` = 1, `max-row` = 4, `min-col` = 1, `max-col` = 3)
convert_limit_list_to_range(rgLL)
convert_limit_list_to_range(rgLL, rc = TRUE)
}
\keyword{internal}

