% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cell-specification.R
\name{convert_range_to_limit_list}
\alias{convert_range_to_limit_list}
\title{Convert a cell range into a limits list}
\usage{
convert_range_to_limit_list(x, rc = FALSE)
}
\arguments{
\item{x}{character vector of length one, representing a cell range, such as
"A1:D7" or "C3"}

\item{rc}{logical, indicating "R1C1" positioning notation}
}
\value{
list with 4 positive integer elements named \code{min-row},
  \code{max-row}, \code{min-col}, \code{max-col}
}
\description{
Convert a cell range into a limits list
}
\examples{
convert_range_to_limit_list("A1")
convert_range_to_limit_list("Q24")
convert_range_to_limit_list("A1:D8")
convert_range_to_limit_list("R5C11", rc = TRUE)
convert_range_to_limit_list("R2C3:R6C9", rc = TRUE)
}
\keyword{internal}

