% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coord_to_label}
\alias{coord_to_label}
\title{Convert coordinate (R1C1) notation to label (A1) notation}
\usage{
coord_to_label(x)
}
\arguments{
\item{x}{coord notation for position of cell}
}
\description{
A1 and R1C1 are equivalent addresses for position of cells.
}
\examples{
coord_to_label("R1C1")
coord_to_label("R10C10")
}

