% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{copy_ss}
\alias{copy_ss}
\title{Make a copy of an existing spreadsheet}
\usage{
copy_ss(from, key = NULL, to = NULL, verbose = TRUE)
}
\arguments{
\item{from}{sheet-identifying information, either a googlesheet object or a
character vector of length one, giving a URL, sheet title, key or
worksheets feed}

\item{key}{character string guaranteed to provide unique key of the sheet;
overrides \code{from}}

\item{to}{character string giving the new title of the sheet; if \code{NULL},
then the copy will be titled "Copy of ..."}

\item{verbose}{logical; do you want informative message?}
}
\description{
You can copy a spreadsheet that you own or a sheet owned by a third party
that has been made accessible via the sharing dialog options. If the sheet
you want to copy is visible in the listing provided by
\code{\link{list_sheets}}, you can specify it by title (or any of the other
spreadsheet-identifying methods). Otherwise, you'll have to explicitly
specify it by key.
}
\note{
if two sheets with the same name exist in your Google drive then sheet
  with the most recent "last updated" timestamp will be copied.
}
\examples{
\dontrun{
gap_key <- "1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA"
gap_ss <- copy_ss(key = gap_key, to = "Gapminder_copy")
gap_ss
}
}
\seealso{
\code{\link{identify_ss}}, \code{\link{extract_key_from_url}}
}

