% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_lookup_tbl}
\alias{create_lookup_tbl}
\title{Create lookup table for cell indices and its values}
\usage{
create_lookup_tbl(cellsfeed)
}
\arguments{
\item{cellsfeed}{Cell based feed parsed with  \code{\link{gsheets_parse}}}
}
\value{
A data frame with cell row and col number and corresponding input value.
}
\description{
Convert cell row and col attributes stored in entry nodes, along with
input values into a single data frame and serve as a lookup table for
further processing.
}

