% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{del_spreadsheet}
\alias{del_spreadsheet}
\title{Move spreadsheet to trash}
\usage{
del_spreadsheet(title)
}
\arguments{
\item{title}{the title of the spreadsheet}
}
\description{
Move a spreadsheet to trash in Google Drive.
}
\note{
Shared spreadsheets can not be removed from your Google Drive with
this function. You must remove it manually in the web browser.
}

