% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{del_worksheet}
\alias{del_worksheet}
\title{Delete a worksheet from spreadsheet}
\usage{
del_worksheet(ss, ws_title)
}
\arguments{
\item{ss}{spreadsheet object}

\item{ws_title}{title of worksheet}
}
\description{
The worksheet and all of its data will be removed from the spreadsheet.
}

