% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permissions.R
\name{delete_perm}
\alias{delete_perm}
\title{Delete a permission from a spreadsheet}
\usage{
delete_perm(ss, email = NULL, perm_id = NULL, verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google Spreadsheet}

\item{email}{The email address or domain name for the entity.}

\item{perm_id}{The ID for the permission.}

\item{verbose}{logical; do you want informative messages?}
}
\description{
Identify the permission to be deleted via email or permission ID.
}
\examples{
\dontrun{
foo <- new_ss("foo")
add_perm(foo, email = NULL, type = "anyone", role = "reader")
delete_perm(foo, email = NA)
}
}

