% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{delete_sheet}
\alias{delete_sheet}
\title{Move a spreadsheet to trash on Google Drive}
\usage{
delete_sheet(x, verbose = TRUE)
}
\arguments{
\item{x}{character string giving identifying information for the sheet:
title, key, URL}

\item{verbose}{logical; do you want informative message?}
}
\description{
You must own the sheet in order to move it to the trash. If you try to delete
a sheet you do not own, a 403 Forbidden HTTP status code will be returned;
such shared spreadsheets can only be moved to the trash manually in the web
browser. If you trash a spreadsheet that is shared with others, it will no
longer appear in any of their Google Drives.
}
\note{
Use the key when there are multiple sheets with the same name, since
  the default will just send the most recent sheet to the trash.
}

