% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{delete_ss}
\alias{delete_ss}
\title{Move a spreadsheet to trash{"__src__":"file","ids":"1PmCRtmcX4r7LOS-XVw6u8f-E3sVnFgamJaKCmVF3Sv0","parentId":"0AFlKxBjVWN80Uk9PVA"}}
\usage{
delete_ss(x)
}
\arguments{
\item{x}{the title or key of a spreadsheet}
}
\description{
Move a spreadsheet to trash in Google Drive.
}
\note{
Using the key is useful when there are spreadsheets with the same name
since the default is to send the most recent spreadsheet to the trash.

Shared spreadsheets can not be removed from your Google Drive with
this function. You must remove it manually in the web browser.
}

