% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{delete_ss}
\alias{delete_ss}
\title{Move a spreadsheet to trash on Google Drive}
\usage{
delete_ss(x, verbose = TRUE)
}
\arguments{
\item{x}{sheet-identifying information, either a spreadsheet object or a
character vector of length one, giving a URL, sheet title, key or
worksheets feed}

\item{verbose}{logical; do you want informative message?}
}
\value{
logical, TRUE if deletion has been explicitly confirmed, FALSE
  otherwise
}
\description{
You must own the sheet in order to move it to the trash. If you try to delete
a sheet you do not own, a 403 Forbidden HTTP status code will be returned;
such shared spreadsheets can only be moved to the trash manually in the web
browser. If you trash a spreadsheet that is shared with others, it will no
longer appear in any of their Google Drives.
}
\note{
Use the key when there are multiple sheets with the same name, since
  the default will just send the most recent sheet to the trash.
}

