% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{delete_ss}
\alias{delete_ss}
\title{Move spreadsheets to trash on Google Drive}
\usage{
delete_ss(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{a regular expression; spreadsheets whose titles match will be moved
to trash}

\item{verbose}{logical; do you want informative message?}

\item{...}{optional arguments to be passed to \code{\link{grepl}}}
}
\value{
tbl_df with one row per matching sheet, a variable holding
  spreadsheet titles, a logical vector indicating deletion success
}
\description{
You must own a sheet in order to move it to the trash. If you try to delete a
sheet you do not own, a 403 Forbidden HTTP status code will be returned; such
shared spreadsheets can only be moved to the trash manually in the web
browser. If you trash a spreadsheet that is shared with others, it will no
longer appear in any of their Google Drives. If you delete something my
mistake, visit the \href{https://drive.google.com/drive/#trash}{trash in
Google Drive}, find the sheet and restore it.
}

