% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{delete_ss}
\alias{delete_ss}
\title{Move spreadsheets to trash on Google Drive}
\usage{
delete_ss(x = NULL, regex = NULL, verbose = TRUE, ...)
}
\arguments{
\item{x}{sheet-identifying information, either a googlesheet object or a
character vector of length one, giving a URL, sheet title, key or
worksheets feed; if \code{x} is specified, the \code{regex} argument will
be ignored}

\item{regex}{character; a regular expression; sheets whose titles match will
be deleted}

\item{verbose}{logical; do you want informative message?}

\item{...}{optional arguments to be passed to \code{\link{grepl}} when
matching \code{regex} to sheet titles}
}
\value{
tbl_df with one row per specified or matching sheet, a variable
  holding spreadsheet titles, a logical vector indicating deletion success
}
\description{
You must own a sheet in order to move it to the trash. If you try to delete a
sheet you do not own, a 403 Forbidden HTTP status code will be returned; such
shared spreadsheets can only be moved to the trash manually in the web
browser. If you trash a spreadsheet that is shared with others, it will no
longer appear in any of their Google Drives. If you delete something by
mistake, remain calm, and visit the
\href{https://drive.google.com/drive/#trash}{trash in Google Drive}, find the
sheet, and restore it.
}
\note{
If there are multiple sheets with the same name and you don't want to
  delete them all, identify the sheet to be deleted via key.
}
\examples{
\dontrun{
foo <- new_ss("foo")
foo <- edit_cells(foo, input = head(iris))
delete_ss("foo")
}
}

