% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{delete_ws}
\alias{delete_ws}
\title{Delete a worksheet from a spreadsheet}
\usage{
delete_ws(ss, ws = 1, verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google spreadsheet}

\item{ws}{positive integer or character string specifying index or title,
  respectively, of the worksheet to consume}

\item{verbose}{logical; do you want informative message?}
}
\description{
The worksheet and all of its contents will be removed from the spreadsheet.
}
\examples{
\dontrun{
gap_key <- "1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA"
gap_ss <- copy_ss(key = gap_key, to = "gap_copy")
list_ws(gap_ss)
gap_ss <- add_ws(gap_ss, "new_stuff")
gap_ss <- edit_cells(gap_ss, "new_stuff", input = head(iris), header = TRUE,
                     trim = TRUE)
gap_ss
gap_ss <- delete_ws(gap_ss, "new_stuff")
list_ws(gap_ss)
gap_ss <- delete_ws(gap_ss, ws = 3)
list_ws(gap_ss)
delete_ss(gap_ss)
}
}

