% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download-spreadsheets.R
\name{download_ss}
\alias{download_ss}
\title{Download a Google spreadsheet}
\usage{
download_ss(from, key = NULL, ws = NULL, to = "my_sheet.xlsx",
  overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{from}{sheet-identifying information, either a googlesheet object or a
character vector of length one, giving a URL, sheet title, key or
worksheets feed}

\item{key}{character string guaranteed to provide unique key of the sheet;
overrides \code{from}}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet to export; if \code{NULL} then the entire
spreadsheet will be exported}

\item{to}{path to write file, if it does not contain the absolute path, then
the file is relative to the current working directory; file extension must
be one of .csv, .pdf, or .xlsx}

\item{overwrite}{logical indicating whether to overwrite an existing local
file}

\item{verbose}{logical; do you want informative message?}
}
\description{
Export a Google sheet as a .csv, .pdf, or .xlsx file. You can download a
sheet that you own or a sheet owned by a third party that has been made
accessible via the sharing dialog options. You can download an entire
spreadsheet or a single worksheet from a spreadsheet if you provide worksheet
identifying information. If the chosen format is csv, the first worksheet
will be exported, unless another worksheet is specified. If pdf format is
chosen, all sheets will be catenated into one PDF document.
}
\examples{
\dontrun{
gap_key <- "1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA"
download_ss(gap_key, to = "gapminder.xlsx")
file.remove("gapminder.xlsx")
}
}

