% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permissions.R
\name{edit_perm}
\alias{edit_perm}
\title{Update an existing permission}
\usage{
edit_perm(ss, perm_id = NULL, role = "reader", verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google Spreadsheet}

\item{perm_id}{The ID for the permission.}

\item{role}{The primary role for this user. Allowed values are "owner",
"reader", and "writer".}

\item{verbose}{logical; do you want informative messages?}
}
\description{
Assign a new role to an existing user permission. This function is used when
you want to change "writer" to a "reader" and vice versa.
}

