% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_key_from_url}
\alias{extract_key_from_url}
\title{Extract sheet key from a URL}
\usage{
extract_key_from_url(url)
}
\arguments{
\item{url}{character; a URL associated with a Google Sheet}
}
\description{
Extract a sheet's unique key from a wide variety of URLs, i.e. a browser URL
for both old and new Sheets, the "worksheets feed", and other links returned
by the Sheets API.
}
\examples{
\dontrun{
gap_url <- "https://docs.google.com/spreadsheets/d/1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA/"
gap_key <- extract_key_from_url(gap_url)
gap_ss <- gs_key(gap_key)
gap_ss
}
}

