% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_key_from_url}
\alias{extract_key_from_url}
\title{Extract sheet key from a URL}
\usage{
extract_key_from_url(url)
}
\arguments{
\item{url}{character; a URL associated with a Google Sheet}
}
\description{
Extract a sheet's unique key from a wide variety of URLs, i.e. a browser URL
for both old and new Sheets, the "worksheets feed", and other links returned
by the Sheets API.
}
\examples{
\dontrun{
GAP_URL <- gs_gap_url()
GAP_KEY <- extract_key_from_url(GAP_URL)
gap_ss <- gs_key(GAP_KEY)
gap_ss
}
}

