% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_key_from_url}
\alias{extract_key_from_url}
\title{Extract sheet key from its browser URL}
\usage{
extract_key_from_url(url)
}
\arguments{
\item{url}{URL seen in the browser when visiting the sheet}
}
\description{
Extract sheet key from its browser URL
}
\examples{
\dontrun{
gap_url <- "https://docs.google.com/spreadsheets/d/1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA/"
gap_key <- extract_key_from_url(gap_url)
gap_ss <- register_ss(gap_key)
gap_ss
}
}

