% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fetch_ws}
\alias{fetch_ws}
\title{Create worksheet objects from worksheets feed}
\usage{
fetch_ws(node, sheet_id)
}
\arguments{
\item{node}{Entry node for worksheet}

\item{sheet_id}{Spreadsheet id housing worksheet}
}
\description{
Store worksheet info (spreadsheet id, worksheet id, worksheet title) that is
embedded as entry nodes in worksheets feed.
}

