% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fetch_ws}
\alias{fetch_ws}
\title{Retrieve worksheet object from worksheets feed
store info from <entry> ... </entry>}
\usage{
fetch_ws(node, sheet_id)
}
\description{
Retrieve worksheet object from worksheets feed
store info from <entry> ... </entry>
}

