% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{fill_missing_col}
\alias{fill_missing_col}
\title{Fill in missing columns in a row}
\usage{
fill_missing_col(x, col_min)
}
\arguments{
\item{x}{data frame returned by \code{\link{get_lookup_tbl}}}

\item{col_min}{leftmost column that row begins at}
}
\description{
The lookup table returned by \code{\link{get_lookup_tbl}} may contain missing tuples
for empty cells. This function fills in the table so that there is a tuple
for every column up to the right-most column of the row.
}
\note{
This function operates on the lookup table grouped by row.
}

