% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{fill_missing_row}
\alias{fill_missing_row}
\title{Fill in missing rows in column}
\usage{
fill_missing_row(x, row_min)
}
\arguments{
\item{x}{data frame returned by \code{\link{get_lookup_tbl}}}

\item{row_min}{top-most row that column begins at}
}
\description{
The lookup table returned by \code{\link{get_lookup_tbl}} may contain missing tuples
for empty cells. This function fills in the table so that there is a tuple
for every row down to the bottom-most row of the column.
}
\note{
This function operates on the lookup table grouped by column.
}

