% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find_cell}
\alias{find_cell}
\title{Find a cell with string value}
\usage{
find_cell(ws, x)
}
\arguments{
\item{ws}{worksheet object}

\item{x}{a character string}
}
\description{
Get the cell location of the first occurence of a cell value.
}

