% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http-requests.R
\name{gdrive_GET}
\alias{gdrive_GET}
\title{Make GET request to Google Drive API}
\usage{
gdrive_GET(url, ...)
}
\arguments{
\item{url}{URL for GET request}

\item{...}{optional; further named parameters, such as \code{query},
  \code{path}, etc, passed on to \code{\link[httr]{modify_url}}. Unnamed
  parameters will be combined with \code{\link[httr]{config}}.}
}
\description{
Used in download_ss()
}
\keyword{internal}

