% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http-requests.R
\name{gdrive_PUT}
\alias{gdrive_PUT}
\title{Make PUT request to Google Drive API}
\usage{
gdrive_PUT(url, the_body)
}
\arguments{
\item{url}{URL for PUT request}

\item{the_body}{body of PUT request}
}
\description{
Used in upload_ss()
}
\keyword{internal}

