% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_all}
\alias{get_all}
\title{Get all values in a worksheet.}
\usage{
get_all(ws, header = TRUE, vis = "private")
}
\arguments{
\item{ws}{worksheet object}

\item{header}{logical for if first row should be taken as header}

\item{vis}{either "private" or "public"}
}
\value{
A dataframe.

This function calls on \code{\link{get_cols}} with to set as number of
columns of the worksheet.
}
\description{
Extract the entire worksheet and turn it into a data frame. This function
uses the rightmost cell with value as the max columns and bottom most cell as max row.
}
\seealso{
\code{\link{get_region}}, \code{\link{get_row}},
\code{\link{get_rows}}, \code{\link{get_col}}, \code{\link{get_cols}}
}

