% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_cell}
\alias{get_cell}
\title{Get the value of a cell}
\usage{
get_cell(ws, cell)
}
\arguments{
\item{ws}{worksheet object}

\item{cell}{character string specifying cell position,
either in label or coordinate notation}
}
\value{
The value of the cell as a character string.
}
\description{
Get the value of a cell using label (A1) notation or coordinate (R1C1)
notation.
}
\examples{
get_cell(ws, "A1")
get_cell(ws, "R1C1")
get_cell(ws, "AB1")
get_cell(ws, "R1C27")
}

