% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_col}
\alias{get_col}
\title{Get all values in a column.}
\usage{
get_col(ws, col, vis = "private")
}
\arguments{
\item{ws}{worksheet object}

\item{col}{column}

\item{vis}{either "private" or "public"}
}
\value{
A data frame.
}
\description{
Get all values in a column.
}
\seealso{
\code{\link{get_cols}}, \code{\link{get_row}},
\code{\link{get_rows}}, \code{\link{get_all}}, \code{\link{get_region}}
}

