% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_col}
\alias{get_col}
\title{Get all values from a column.}
\usage{
get_col(client, ws, col)
}
\arguments{
\item{client}{a client object returned by \code{\link{login}}
or \code{\link{authorize}}.}

\item{ws}{worksheet object returned by \code{\link{get_worksheet}}}

\item{col}{column}
}
\value{
Vector of all values in column
}
\description{
Get all values from a column.
}

