% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/consume-data.R
\name{get_col}
\alias{get_col}
\title{Get data from a column or range of columns}
\usage{
get_col(ss, ws = 1, col)
}
\arguments{
\item{ss}{a registered Google spreadsheet}

\item{ws}{positive integer or character string specifying index or title,
  respectively, of the worksheet to consume}

\item{col}{vector of positive integers, possibly of length one, specifying
  which columns to retrieve; only contiguous ranges of columns are supported,
  i.e. if \code{col = c(2, 8)}, you will get columns 2 through 8}
}
\description{
Get data via the cell feed for one column or for a range of columns.
}
\seealso{
\code{\link{reshape_cf}} to reshape the retrieved data into a more
  usable data.frame

Other data.consumption.functions: \code{\link{get_cells}};
  \code{\link{get_row}}; \code{\link{get_via_cf}};
  \code{\link{get_via_lf}}; \code{\link{reshape_cf}};
  \code{\link{simplify_cf}}
}

