% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume-data.R
\name{get_col}
\alias{get_col}
\title{Get data from a column or range of columns}
\usage{
get_col(ss, ws = 1, col, verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google spreadsheet}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet to consume}

\item{col}{vector of positive integers, possibly of length one, specifying
  which columns to retrieve; only contiguous ranges of columns are supported,
  i.e. if \code{col = c(2, 8)}, you will get columns 2 through 8}

\item{verbose}{logical; do you want informative messages?}
}
\description{
Get data via the cell feed for one column or for a range of columns.
}
\examples{
\dontrun{
gap_key <- "1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA"
gap_ss <- gs_copy(gs_key(gap_key), to = "gap_copy")
get_col(gap_ss, "Oceania", col = 1:2)
reshape_cf(get_col(gap_ss, "Oceania", col = 1:2))
}
}
\seealso{
\code{\link{reshape_cf}} to reshape the retrieved data into a more
  usable data.frame

Other data.consumption.functions: \code{\link{get_cells}};
  \code{\link{get_row}}; \code{\link{get_via_cf}};
  \code{\link{get_via_csv}}; \code{\link{get_via_lf}};
  \code{\link{reshape_cf}}; \code{\link{simplify_cf}}
}

