% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_cols}
\alias{get_cols}
\title{Get columns of worksheet}
\usage{
get_cols(client = NULL, ws, from, to)
}
\arguments{
\item{client}{a client object returned by \code{\link{login}}
or \code{\link{authorize}}.}

\item{ws}{worksheet object returned by \code{\link{get_worksheet}}}

\item{from,to}{start and end column indexes}
}
\value{
Dataframe of requested columns.
}
\description{
Specify range of columns to get from worksheet.
}
\examples{
get_cols(ws, 1, 2)
get_cols(ws, 30, 40)
}

