% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_cols}
\alias{get_cols}
\title{Get all values in a range of columns.}
\usage{
get_cols(ws, from, to, header = TRUE)
}
\arguments{
\item{ws}{worksheet object}

\item{from,to}{start and end column indexes either integer or letter}

\item{header}{\code{logical} to indicate if the first row should be taken as
the header}
}
\value{
A data frame.
}
\description{
Specify range of columns to get from worksheet.
}
\seealso{
\code{\link{get_col}}, \code{\link{get_row}},
\code{\link{get_rows}}, \code{\link{read_all}}, \code{\link{read_region}},
\code{\link{read_range}}
}

