% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_cols}
\alias{get_cols}
\title{Get all values in a range of columns.}
\usage{
get_cols(ws, from, to, header = TRUE, vis = "private")
}
\arguments{
\item{ws}{worksheet object}

\item{from,to}{start and end column indexes either integer or letter}
}
\value{
A data frame.
}
\description{
Specify range of columns to get from worksheet.
}
\examples{
get_cols(ws, 1, 2)
get_cols(ws, 30, 40)

get_cols(ws, "A", "B")
get_cols(ws, "c", "f")
}
\seealso{
\code{\link{get_col}}, \code{\link{get_row}},
\code{\link{get_rows}}, \code{\link{get_all}}, \code{\link{read_region}},
\code{\link{read_range}}
}

