% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_dataframe}
\alias{get_dataframe}
\title{Get all values of worksheet as a dataframe}
\usage{
get_dataframe(ws, vis = "private")
}
\arguments{
\item{ws}{Worksheet object}
}
\value{
A dataframe.

This function calls on \code{\link{get_cols}} with to set as number of
columns of the worksheet.
}
\description{
Use worksheet object and turn it into a dataframe.
}

