% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_row}
\alias{get_row}
\title{Get all values from a row.}
\usage{
get_row(client, ws, row)
}
\arguments{
\item{client}{a client object returned by \code{\link{login}}
or \code{\link{authorize}}.}

\item{ws}{worksheet object returned by \code{\link{get_worksheet}}}

\item{row}{row}
}
\value{
Vector of all values in row.
}
\description{
Get all values from a row.
}

