% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_row}
\alias{get_row}
\title{Get all values in a row}
\usage{
get_row(ws, row, vis = "private")
}
\arguments{
\item{ws}{worksheet object}

\item{row}{row number}

\item{vis}{either "private" or "public", set to "public" for public for
worksheets}
}
\value{
A data frame.
}
\description{
Specify row of values to get from worksheet.
}
\seealso{
\code{\link{get_rows}}, \code{\link{get_col}},
\code{\link{get_cols}}, \code{\link{get_all}}, \code{\link{read_region}},
\code{\link{read_range}}
}

