% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume-data.R
\name{get_row}
\alias{get_row}
\title{Get data from a row or range of rows}
\usage{
get_row(ss, ws = 1, row, verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google spreadsheet}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet to consume}

\item{row}{vector of positive integers, possibly of length one, specifying
  which rows to retrieve; only contiguous ranges of rows are supported, i.e.
  if \code{row = c(2, 8)}, you will get rows 2 through 8}

\item{verbose}{logical; do you want informative messages?}
}
\description{
Get data via the cell feed for one row or for a range of rows.
}
\examples{
\dontrun{
gap_key <- "1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA"
gap_ss <- copy_ss(key = gap_key, to = "gap_copy")
get_row(gap_ss, "Europe", row = 1)
simplify_cf(get_row(gap_ss, "Europe", row = 1))
}
}
\seealso{
\code{\link{reshape_cf}} to reshape the retrieved data into a more
  usable data.frame

Other data.consumption.functions: \code{\link{get_cells}};
  \code{\link{get_col}}; \code{\link{get_via_cf}};
  \code{\link{get_via_csv}}; \code{\link{get_via_lf}};
  \code{\link{reshape_cf}}; \code{\link{simplify_cf}}
}

