% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/consume-data.R
\name{get_row}
\alias{get_row}
\title{Get all values in a row}
\usage{
get_row(ws, row)
}
\arguments{
\item{ws}{worksheet object}

\item{row}{row number}
}
\value{
A data frame.
}
\description{
Specify row of values to get from worksheet.
}
\seealso{
\code{\link{get_rows}}, \code{\link{get_col}},
\code{\link{get_cols}}, \code{\link{read_all}}, \code{\link{read_region}},
\code{\link{read_range}}
}

