% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_rows}
\alias{get_rows}
\title{Get all values in range of rows.}
\usage{
get_rows(ws, from, to, header = FALSE, vis = "private")
}
\arguments{
\item{ws}{worksheet object}

\item{from,to}{start and end row indexes}
}
\value{
A data frame.
}
\description{
Specify range of rows to get from worksheet.
}
\seealso{
\code{\link{get_row}}, \code{\link{get_col}},
\code{\link{get_cols}}, \code{\link{get_all}}, \code{\link{read_region}}
}

