% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{get_rows}
\alias{get_rows}
\title{Get all values in range of rows}
\usage{
get_rows(ws, from, to, header = FALSE)
}
\arguments{
\item{ws}{worksheet object}

\item{from,to}{start and end row indexes}

\item{header}{\code{logical} to indicate if the first row should be taken as the
header}
}
\value{
A data frame.
}
\description{
Specify range of rows to get from worksheet.
}
\seealso{
\code{\link{get_row}}, \code{\link{get_col}},
\code{\link{get_cols}}, \code{\link{read_all}}, \code{\link{read_region}},
\code{\link{read_range}}
}

