% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_rows}
\alias{get_rows}
\title{Get rows of worksheet}
\usage{
get_rows(client, ws, from, to)
}
\arguments{
\item{client}{a client object returned by \code{\link{login}}
or \code{\link{authorize}}.}

\item{ws}{worksheet object returned by \code{\link{get_worksheet}}}

\item{from,to}{start and end row indexes}
}
\value{
Dataframe of requested rows.
}
\description{
Specify range of rows to get from worksheet.
}

