% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/consume-data.R
\name{get_via_csv}
\alias{get_via_csv}
\title{Get all data from a worksheet as a tbl_df or data.frame}
\usage{
get_via_csv(ss, ws = 1, ...)
}
\arguments{
\item{ss}{a registered Google spreadsheet}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet to consume}

\item{...}{further arguments to be passed to read.csv, read.table}
}
\description{
This function consumes data using the exportcsv links found in the worksheets
feed. Unlike using the list feed, this method does not assume populated cells form
a neat rectangle, all cells within the data rectangle (max row and
column with data) are returned. Also, capitals and special characters
(non alphanumeric) are preserved in the header. Empty cells will be assigned
NA. If you want all of your data, this is the fastest way to get it.
This is faster than methods based on the listfeed and cellfeed.
}

