% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume-data.R
\name{get_via_lf}
\alias{get_via_lf}
\title{Get data from a rectangular worksheet as a tbl_df or data.frame}
\usage{
get_via_lf(ss, ws = 1, verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet}

\item{verbose}{logical; do you want informative messages?}
}
\value{
a tbl_df
}
\description{
Gets data via the list feed, which assumes populated cells form a neat
rectangle. The list feed consumes data row by row. First row regarded as
header row of variable or column names. The related function,
\code{get_via_csv}, also returns data from a neat rectangle of cells, so you
probably want to use that (unless you are dealing with an "old" Google Sheet,
which \code{get_via_csv} does not support).
}
\note{
When you use the listfeed, the Sheets API transforms the variable or
  column names like so: 'The column names are the header values of the
  worksheet lowercased and with all non-alpha-numeric characters removed. For
  example, if the cell A1 contains the value "Time 2 Eat!" the column name
  would be "time2eat".' If this is intolerable to you, consume the data via
  the cell feed or csv download. Or, at least, consume the first row via the
  cell feed and manually restore the variable names post hoc.
}
\examples{
\dontrun{
gap_ss <- gs_gap() # register the Gapminder example sheet
oceania_lf <- get_via_lf(gap_ss, ws = "Oceania")
str(oceania_lf)
oceania_lf
}
}
\seealso{
Other data.consumption.functions: \code{\link{get_cells}};
  \code{\link{get_col}}; \code{\link{get_row}};
  \code{\link{get_via_cf}}; \code{\link{get_via_csv}};
  \code{\link{reshape_cf}}; \code{\link{simplify_cf}}
}

