% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_worksheet}
\alias{get_worksheet}
\title{Open worksheet given worksheet title}
\usage{
get_worksheet(spreadsheet, title)
}
\arguments{
\item{spreadsheet}{Spreadsheet object housing the desired worksheet}

\item{title}{title of worksheet to retrieve}
}
\value{
An object of class worksheet.
}
\description{
Use title of worksheet to retrieve object of class spreadsheet.
}

