% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ws}
\alias{get_ws}
\title{Retrieve a worksheet-describing list from a googlesheet}
\usage{
get_ws(ss, ws, verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google spreadsheet}

\item{ws}{positive integer or character string specifying index or title,
  respectively, of the worksheet to consume}

\item{verbose}{logical, indicating whether to give a message re: title of the
  worksheet being accessed}
}
\description{
From a googlesheet, retrieve a list (actually a row of a data.frame) giving
everything we know about a specific worksheet.
}
\keyword{internal}

