% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ws}
\alias{get_ws}
\title{Retrieve a worksheet-describing list from a spreadsheet}
\usage{
get_ws(ss, ws, verbose = TRUE)
}
\arguments{
\item{ss}{a registered spreadsheet}

\item{ws}{a positive integer or character string specifying which worksheet}
}
\description{
From a registered spreadsheet, retrieve a list (actually a row of a
data.frame) giving everything we know about a specific worksheet.
}

