% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/googlesheet.R
\name{googlesheet}
\alias{googlesheet}
\title{The googlesheet object}
\usage{
googlesheet()
}
\description{
The googlesheet object stores information that \code{googlesheets} requires in
order to communicate with the
\href{https://developers.google.com/google-apps/spreadsheets/}{Google Sheets
API}.
}
\details{
Very little of this is of interest to the user. A googlesheet object
includes the fields:

\itemize{
\item \code{sheet_key} the key of the spreadsheet
\item \code{sheet_title} the title of the spreadsheet
\item \code{n_ws} the number of worksheets contained in the spreadsheet
\item \code{ws_feed} the "worksheets feed" of the spreadsheet
\item \code{sheet_id} the id of the spreadsheet
\item \code{updated} the time of last update (at time of registration)
\item \code{get_date} the time of registration
\item \code{visibility} visibility of spreadsheet (Google's confusing
vocabulary); actually, does not describe a property of spreadsheet itself but
rather whether requests will be made with or without authentication
\item \code{is_public} logical indicating visibility is "public", as opposed to "private"
\item \code{author_name} the name of the owner
\item \code{author_email} the email of the owner
\item \code{links} data.frame of links specific to the spreadsheet
\item \code{ws} a data.frame about the worksheets contained in the
spreadsheet
\item \code{alt_key} alternate key; applies only to "old" sheets
}

TO DO: this documentation is neither here nor there. Either the object is
self-explanatory and this isn't really needed. Or this needs to get beefed
up. Probably the latter.
}

