% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googlesheets.R
\docType{package}
\name{googlesheets}
\alias{googlesheets}
\alias{googlesheets-package}
\title{googlesheets}
\description{
Google spreadsheets R API
}
\details{
See the \href{https://github.com/jennybc/googlesheets}{README} on GitHub
}

