% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gs_new.R
\name{gs_new}
\alias{gs_new}
\title{Create a new spreadsheet}
\usage{
gs_new(title = "my_sheet", verbose = TRUE)
}
\arguments{
\item{title}{the title for the new sheet}

\item{verbose}{logical; do you want informative message?}
}
\value{
a \code{\link{googlesheet}} object
}
\description{
Create a new (empty) spreadsheet in your Google Drive. The new sheet will
contain 1 default worksheet titled "Sheet1". This function
calls the \href{https://developers.google.com/drive/v2/reference/}{Google
Drive API}.
}
\examples{
\dontrun{
foo <- gs_new("foo")
foo
}
}

