% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{gs_user}
\alias{gs_user}
\title{Retrieve information about authorized user}
\usage{
gs_user(verbose = TRUE)
}
\arguments{
\item{verbose}{logical; do you want informative messages?}
}
\value{
a list containing user and session info
}
\description{
Display information about a user that has been authorized via \code{gs_auth}:
the user's display name, email, the date-time of authorization for the
current session, date-time of access token expiry, and the validity of the
current access token. This is a subset of the information available from
\href{https://developers.google.com/drive/v2/reference/about/get}{the "about"
endpoint} of the Drive API.
}
\examples{
\dontrun{
gs_user()
}
}

