% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gs_ws.R
\name{gs_ws_new}
\alias{gs_ws_new}
\title{Add a new worksheet to spreadsheet}
\usage{
gs_ws_new(ss, ws_title = "Sheet1", nrow = 1000, ncol = 26,
  verbose = TRUE)
}
\arguments{
\item{ss}{a \code{\link{googlesheet}} object, i.e. a registered Google
sheet}

\item{ws_title}{character string for title of new worksheet}

\item{nrow}{number of rows (default is 1000)}

\item{ncol}{number of columns (default is 26)}

\item{verbose}{logical; do you want informative message?}
}
\value{
a \code{\link{googlesheet}} object
}
\description{
Add a new (empty) worksheet to spreadsheet: specify title and worksheet
extent (number of rows and columns). The title of the new worksheet can not
be the same as any existing worksheets in the sheet.
}
\examples{
\dontrun{
# get a copy of the Gapminder spreadsheet
gap_key <- "1HT5B8SgkKqHdqHJmn5xiuaC04Ngb7dG9Tv94004vezA"
gap_ss <- gs_copy(gs_key(gap_key), to = "Gapminder_copy")
gap_ss <- gs_ws_new(gap_ss, ws_title = "Atlantis")
gap_ss
}
}

