% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsheets_auth}
\alias{gsheets_auth}
\title{Check if token is obtained from Google login or oauth2.0}
\usage{
gsheets_auth(token)
}
\arguments{
\item{token}{Google token}
}
\description{
Add token as a header or token in configuations in URL request.
}

