% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsheets_check}
\alias{gsheets_check}
\title{Check status of http response}
\usage{
gsheets_check(req)
}
\arguments{
\item{req}{response from \code{\link{gsheets_GET}} request}
}
\description{
Google returns status 200 (success) or 403 (failure), show error msg if 403.
}

