% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http_requests.R
\name{gsheets_DELETE}
\alias{gsheets_DELETE}
\title{Create DELETE request}
\usage{
gsheets_DELETE(url)
}
\arguments{
\item{url}{the url of the page to retrieve}
}
\description{
Make DELETE request to Google Sheets API.
}
\keyword{internal}

