% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http-requests.R
\name{gsheets_DELETE}
\alias{gsheets_DELETE}
\title{Create DELETE request}
\usage{
gsheets_DELETE(url)
}
\arguments{
\item{url}{URL for DELETE request}
}
\description{
Make DELETE request to Google Sheets API.
}
\keyword{internal}

