% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsheets_GET}
\alias{gsheets_GET}
\title{Create GET request}
\usage{
gsheets_GET(url, token = get_google_token())
}
\arguments{
\item{url}{URL for GET request}

\item{token}{Google auth token obtained from \code{\link{login}}
or \code{\link{authorize}}}
}
\description{
Make GET request to Google Sheets API.
}

