% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http_requests.R
\name{gsheets_GET}
\alias{gsheets_GET}
\title{Create GET request}
\usage{
gsheets_GET(url, to_xml = TRUE, use_auth = TRUE, ...)
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{to_xml}{whether to convert response contents to xml_doc() or leave as
character string}

\item{use_auth}{logical; should authorization be used, defaults to
\code{TRUE}, \code{FALSE} when sheet is public}

\item{...}{optional; further named parameters, such as \code{query},
  \code{path}, etc, passed on to \code{\link[httr]{modify_url}}. Unnamed
  parameters will be combined with \code{\link[httr]{config}}.}
}
\description{
Make GET request to Google Sheets API.
}
\keyword{internal}

