% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/http-requests.R
\name{gsheets_GET}
\alias{gsheets_GET}
\title{Create GET request}
\usage{
gsheets_GET(url, token = NULL)
}
\arguments{
\item{url}{URL for GET request}

\item{token}{Google auth token obtained from \code{\link{login}}
or \code{\link{authorize}}}
}
\description{
Make GET request to Google Sheets API.
}

