% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/http-requests.R
\name{gsheets_GET}
\alias{gsheets_GET}
\title{Create GET request}
\usage{
gsheets_GET(url)
}
\arguments{
\item{url}{URL for GET request}
}
\description{
Make GET request to Google Sheets API.
}

