% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http-requests.R
\name{gsheets_GET}
\alias{gsheets_GET}
\title{Create GET request}
\usage{
gsheets_GET(url, to_list = TRUE, ...)
}
\arguments{
\item{url}{URL for GET request}

\item{to_list}{whether to convert response contents to list or not}

\item{...}{optional; further named parameters, such as \code{query},
  \code{path}, etc, passed on to \code{\link[httr]{modify_url}}. Unnamed
  parameters will be combined with \code{\link[httr]{config}}.}
}
\description{
Make GET request to Google Sheets API.
}
\keyword{internal}

