% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/http-requests.R
\name{gsheets_POST}
\alias{gsheets_POST}
\title{Create POST request}
\usage{
gsheets_POST(url, the_body, content_type = "atom+xml",
  token = get_google_token())
}
\arguments{
\item{url}{URL for POST request}

\item{the_body}{body of POST request}

\item{content_type}{the content type, default is "atom+xml" for posting to
Sheets API, "json" for posting to Drive API}

\item{token}{Google auth token obtained from \code{\link{login}}
or \code{\link{authorize}}}
}
\description{
Make POST request to Google Sheets/Drive API.
}

