% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/http-requests.R
\name{gsheets_PUT}
\alias{gsheets_PUT}
\title{Create PUT request}
\usage{
gsheets_PUT(url, the_body, token = get_google_token())
}
\arguments{
\item{url}{URL for PUT request}

\item{the_body}{body of PUT request}

\item{token}{Google auth token obtained from \code{\link{login}}
or \code{\link{authorize}}}
}
\description{
Make PUT request to Google Sheets API.
}

