% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http-requests.R
\name{gsheets_PUT}
\alias{gsheets_PUT}
\title{Create PUT request}
\usage{
gsheets_PUT(url, the_body)
}
\arguments{
\item{url}{URL for PUT request}

\item{the_body}{body of PUT request}
}
\description{
Make PUT request to Google Sheets API.
}
\keyword{internal}

