% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{is_legit_token}
\alias{is_legit_token}
\title{Check that token appears to be legitimate}
\usage{
is_legit_token(x, verbose = FALSE)
}
\description{
Check that token appears to be legitimate
}
\keyword{internal}
