% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{label_to_coord}
\alias{label_to_coord}
\title{Convert A1 positioning notation to R1C1 notation}
\usage{
label_to_coord(x)
}
\arguments{
\item{x}{cell position in A1 notation}
}
\description{
Convert A1 positioning notation to R1C1 notation
}
\keyword{internal}

