% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/helpers.R
\name{letter_to_num}
\alias{letter_to_num}
\title{Convert column letter to number}
\usage{
letter_to_num(x)
}
\arguments{
\item{x}{column letter (case insensitive)}
}
\description{
Convert column letter to number
}

