% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{letter_to_num}
\alias{letter_to_num}
\title{Convert column IDs from letter representation to numeric}
\usage{
letter_to_num(x)
}
\arguments{
\item{x}{character vector of letter-style column IDs (case insensitive)}
}
\description{
Convert column IDs from letter representation to numeric
}
\keyword{internal}

