% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permissions.R
\name{list_perm}
\alias{list_perm}
\title{Get a listing of permissions for a spreadsheet}
\usage{
list_perm(ss)
}
\arguments{
\item{ss}{a registered Google spreadsheet}
}
\value{
a tbl_df, one row per permission
}
\description{
This function returns the information available from the
\href{https://developers.google.com/drive/v2/reference/permissions}{permissions
feed} of the Google Drive API.
}
\details{
This listing gives the user all permissions for a spreadsheet.
The first row will always be the permission for the owner of the spreadsheet.
Any additional permissions follow.
}
\examples{
\dontrun{
foo <- new_ss("foo")
list_perm(foo)
}
}

