% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/register.R
\name{list_sheets}
\alias{list_sheets}
\title{Get a listing of spreadsheets}
\usage{
list_sheets(regex = NULL, ..., verbose = TRUE)
}
\value{
a \code{googlesheet_ls} object, which is a
  \code{\link[dplyr]{tbl_df}} with one row per sheet
}
\description{
Please use \code{\link{gs_ls}} instead. This function is going away.
}
\examples{
\dontrun{
gs_ls()
}
}

