% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_spreadsheets}
\alias{list_spreadsheets}
\title{Get list of spreadsheets for authenticated user}
\usage{
list_spreadsheets(client)
}
\arguments{
\item{x}{Client object returned by login(email, passwd)}
}
\value{
Dataframe of spreadsheet names.

Use auth token stored in http_session object of client object to make GET request.
}
\description{
Retrive list of spreadsheets
}

