% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_spreadsheets}
\alias{list_spreadsheets}
\title{Get list of spreadsheets for authenticated user}
\usage{
list_spreadsheets(client)
}
\arguments{
\item{client}{a client object returned by \code{\link{login}}
or \code{\link{authorize}}}
}
\description{
Retrieve the names of spreadsheets owned by user.
}

