% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_spreadsheets}
\alias{list_spreadsheets}
\title{Get list of spreadsheets for authenticated user}
\usage{
list_spreadsheets(client)
}
\arguments{
\item{client}{Object of class client returned by \code{\link{login}}}
}
\value{
Dataframe of spreadsheet titles.
}
\description{
Retrive list of spreadsheets
}

