% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_worksheets}
\alias{list_worksheets}
\title{Get list of worksheets contained in spreadsheet}
\usage{
list_worksheets(ss)
}
\arguments{
\item{ss}{spreadsheet object returned by \code{\link{open_spreadsheet}}

This is a mini wrapper around spreadsheet$ws_names.}
}
\description{
Retrieve list of worksheet titles (order as it appears in spreadsheet)
contained in spreadsheet.
}

