% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_worksheets}
\alias{list_worksheets}
\title{The worksheets contained in Spreadsheet}
\usage{
list_worksheets(ss)
}
\arguments{
\item{spreadsheet}{A spreadsheet object returned by \code{\link{open_spreadsheet}}}
}
\value{
The titles of worksheets contained in spreadsheet.

This is a mini wrapper for spreadsheet$ws_names.
}
\description{
Get list of worksheet titles contained in spreadsheet.
}

