% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{list_worksheets}
\alias{list_worksheets}
\title{Get list of worksheets contained in spreadsheet}
\usage{
list_worksheets(ss)
}
\arguments{
\item{ss}{a spreadsheet object returned by \code{\link{open_spreadsheet}}

This is a mini wrapper around spreadsheet$ws_names.}
}
\description{
Retrieve list of worksheet titles (order as they appear in spreadsheet).
}

