% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_worksheets}
\alias{list_worksheets}
\title{Get list of worksheets for authenticated user}
\usage{
list_worksheets(x)
}
\arguments{
\item{x}{A spreadsheet object returned by open_sheet(client, title)}
}
\value{
The names of worksheets contained in spreadsheet.

This is a mini wrapper for x$ws_names.
}
\description{
Retrive list of worksheets contained in spreadsheet
}

