% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{list_ws}
\alias{list_ws}
\title{List the worksheets in a gspreadsheet}
\usage{
list_ws(ss)
}
\arguments{
\item{ss}{a registered Google spreadsheet}
}
\description{
Retrieve the titles of all the worksheets in a gpreadsheet.
}

