% Generated by roxygen2 (4.0.2): do not edit by hand
\name{login}
\alias{login}
\title{Authorize client using ClientLogin}
\usage{
login(email, passwd)
}
\arguments{
\item{email}{User's email.}

\item{passwd}{Password for user's email.}
}
\value{
Object of class client which stores the token used to subsequent requests.

This method is using API as described at:
https://developers.google.com/accounts/docs/AuthForInstalledApps

Authorization token will be stored in http_session object which then gets
stord in client object.
}
\description{
Authorize using email and password.
}

