% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auth.R
\name{login}
\alias{login}
\title{Authorize user using ClientLogin}
\usage{
login(email, passwd)
}
\arguments{
\item{email}{User's email.}

\item{passwd}{User's password.

This method is using API as described at:
\url{https://developers.google.com/accounts/docs/AuthForInstalledApps}}
}
\description{
Authorize user using email and password.
}

