% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_ws_obj}
\alias{make_ws_obj}
\title{Create worksheet objects from worksheets feed}
\usage{
make_ws_obj(node, sheet_id)
}
\arguments{
\item{node}{entry node for worksheet}

\item{sheet_id}{spreadsheet id housing worksheet}
}
\description{
Extract worksheet info (spreadsheet id, worksheet id, worksheet title,
row count and col count) from entry nodes in worksheets feed as worksheet
objects.
}

