% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{modify_ws}
\alias{modify_ws}
\title{Modify a worksheet's title or size}
\usage{
modify_ws(ss, from, to = NULL, new_dim = NULL)
}
\arguments{
\item{ss}{a registered Google sheet}

\item{from}{positive integer or character string specifying index or title,
respectively, of the worksheet}

\item{to}{character string for new title of worksheet}

\item{new_dim}{list of length 2 specifying the row and column extent of the
  worksheet}
}
\description{
Modify a worksheet's title or size
}
\keyword{internal}

