% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{new_ss}
\alias{new_ss}
\title{Create a new spreadsheet}
\usage{
new_ss(title = "my_sheet", verbose = TRUE)
}
\arguments{
\item{title}{the title for the new sheet}

\item{verbose}{logical; do you want informative message?}
}
\value{
a googlesheet object
}
\description{
Create a new (empty) spreadsheet in your Google Drive. The new sheet will
contain 1 default worksheet titled "Sheet1".
}
\examples{
\dontrun{
foo <- new_ss("foo")
foo
}
}

