% Generated by roxygen2 (4.0.2): do not edit by hand
\name{num_to_letter}
\alias{num_to_letter}
\title{Convert column number to column letter}
\usage{
num_to_letter(x)
}
\arguments{
\item{x}{column number}
}
\description{
Convert column number to column letter
}
\examples{
num_to_letter(1)
num_to_letter(26)
}

