% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{num_to_letter}
\alias{num_to_letter}
\title{Convert column IDs from numeric to letter representation}
\usage{
num_to_letter(x)
}
\arguments{
\item{x}{vector of numeric column IDs}
}
\description{
Convert column IDs from numeric to letter representation
}
\keyword{internal}

