% Generated by roxygen2 (4.0.2): do not edit by hand
\name{open_at_once}
\alias{open_at_once}
\title{Open a worksheet from a spreadsheet at once}
\usage{
open_at_once(ss_title, ws_index)
}
\arguments{
\item{ws_index}{worksheet name or numeric for worksheet index}

\item{ss_name}{Spreadsheet title}
}
\description{
Open a worksheet from a spreadsheet at once
}

