% Generated by roxygen2 (4.0.2): do not edit by hand
\name{open_by_key}
\alias{open_by_key}
\title{Open spreadsheet by key}
\usage{
open_by_key(spreadsheet_key)
}
\arguments{
\item{spreadsheet_key}{A key of a spreadsheet as it appears in browser URL.}
}
\value{
Object of class spreadsheet.

This function currently only works for public spreadsheets (visibility = TRUE and projection = FULL).
}
\description{
Use key found in browser URL and return an object of class spreadsheet.
}

