% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{open_by_url}
\alias{open_by_url}
\title{Open spreadsheet by url}
\usage{
open_by_url(url, visibility = "private")
}
\arguments{
\item{url}{URL of spreadsheet as it appears in browser}

\item{visibility}{either "public" for public spreadsheets or "private"
for private spreadsheets}
}
\value{
Object of class spreadsheet

This function only works for public spreadsheets.
This function extracts the key from the url and calls on
\code{\link{open_by_key}}.
}
\description{
Use url of spreadsheet and return an object of class spreadsheet.
}
\note{
This function assumes the longest character string separated by "/" in the url is the
key of the spreadsheet.

The visibility should be set to "public" only if the spreadsheet is
"Published to the web". This is different from setting the spreadsheet to "Public on the web"
in the visibility options in the sharing dialog of a Google Sheets file.
}

