% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{open_spreadsheet}
\alias{open_spreadsheet}
\title{Open spreadsheet by title}
\usage{
open_spreadsheet(title)
}
\arguments{
\item{title}{the title of the spreadsheet}
}
\description{
Use the spreadsheet title (as it appears in Google Drive) to get a
spreadsheet object, containing the spreadsheet id, title, time of last
update, the titles of the worksheets contained, the number of worksheets
contained, and a list of worksheet objects for every worksheet.
}
\note{
The list of worksheet objects returned is missing the ncol, nrow and
visibility components. Those are determined when \code{\link{open_worksheet}}
or \code{\link{open_at_once}} is used to open a worksheet. It is time
consuming to make a cellfeed request for every worksheet in the spreadsheet
to determine the number of rows and columns. Use
\code{\link{open_worksheets}} to open all worksheets contained in the
spreadsheet.
}
\seealso{
\code{\link{open_worksheets}}
}

