% Generated by roxygen2 (4.0.2): do not edit by hand
\name{open_spreadsheet}
\alias{open_spreadsheet}
\title{Open spreadsheet by title}
\usage{
open_spreadsheet(client, title)
}
\arguments{
\item{client}{Client object returned by \code{\link{login}}}

\item{title}{title of spreadsheet}
}
\value{
Object of class spreadsheet.
}
\description{
Use title of spreadsheet to retrieve object of class spreadsheet.
}

