% Generated by roxygen2 (4.0.2): do not edit by hand
\name{open_spreadsheet}
\alias{open_spreadsheet}
\title{Open spreadsheet by title}
\usage{
open_spreadsheet(title)
}
\arguments{
\item{title}{the title of a spreadsheet.}
}
\value{
Object of class spreadsheet.
}
\description{
Use title of spreadsheet to get object of class spreadsheet.
}

