% Generated by roxygen2 (4.0.2): do not edit by hand
\name{open_worksheet}
\alias{open_worksheet}
\title{Open worksheet given worksheet title or index}
\usage{
open_worksheet(ss, value, vis = "private")
}
\arguments{
\item{ss}{spreadsheet object containing worksheet}

\item{value}{a character string for title of worksheet or numeric for
index of worksheet}
}
\value{
A worksheet object with number of rows and cols component.
}
\description{
Use title or index of worksheet to retrieve worksheet object.
}
\examples{
ws <- open_worksheet(ss, "Sheet1")
ws <- open_worksheet(ss, 1)
}

