% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print.googlesheet}
\alias{print.googlesheet}
\title{Print information about a Google Sheet registered with \code{googlesheets}}
\usage{
\method{print}{googlesheet}(x, ...)
}
\arguments{
\item{x}{googlesheet object returned by \code{register_ss} and other
\code{googlesheets} functions}

\item{...}{potential further arguments (required for Method/Generic reasons)}
}
\description{
Display information about a Google spreadsheet that has been registered with
\code{googlesheets}: the title of the spreadsheet, date-time of registration,
date-time of last update (at time of registration), the number of worksheets
contained, worksheet titles and extent, and sheet key.
}
\examples{
\dontrun{
foo <- gs_new("foo")
foo
print(foo)
}
}

