% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_range}
\alias{read_range}
\title{Get a region of a worksheet by range}
\usage{
read_range(ws, x, header = TRUE, vis = "private")
}
\arguments{
\item{ws}{worksheet object}

\item{x}{character string for range separated by ":"}

\item{header}{\code{logical} for whether or not first row should be taken as
header}

\item{vis}{either "private" (default) or "public" for public spreadsheets}
}
\description{
Extract cells of a worksheet by specifying the minimum and maximum rows and
columns. If the specified range is beyond the dimensions of the worksheet,
the boundaries of the worksheet will be used instead.
}
\examples{
read_range("A1:B10")
read_range("C10:D20")
}
\seealso{
\code{\link{read_region}}, \code{\link{get_row}},
\code{\link{get_rows}}, \code{\link{get_col}}, \code{\link{get_cols}},
\code{\link{get_all}}
}

