% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{read_range}
\alias{read_range}
\title{Get a region of a worksheet by range}
\usage{
read_range(ws, x, header = TRUE)
}
\arguments{
\item{ws}{worksheet object}

\item{x}{character string for range separated by ":"}

\item{header}{\code{logical} to indicate if the first row should be taken as the
header}
}
\description{
Extract cells of a worksheet by specifying the minimum and maximum rows and
columns. If the specified range is beyond the dimensions of the worksheet,
the boundaries of the worksheet will be used instead.
}
\examples{
\dontrun{
worksheet <- open_at_once("My Spreadsheet", "Sheet1")

read_range(worksheet, "A1:B10")
read_range(worksheet, "C10:D20")
}
}
\seealso{
\code{\link{read_region}}, \code{\link{get_row}},
\code{\link{get_rows}}, \code{\link{get_col}}, \code{\link{get_cols}},
\code{\link{read_all}}
}

