% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_region}
\alias{read_region}
\title{Get a region of a worksheet}
\usage{
read_region(ws, from_row, to_row, from_col, to_col, header = TRUE,
  vis = "private")
}
\arguments{
\item{ws}{Worksheet object}

\item{from_row,to_row}{range of rows to extract}

\item{from_col,to_col}{range of cols to extract}

\item{vis}{Either \code{private} or \code{public}}
}
\value{
A data frame.
}
\description{
Extract cells of a worksheet according to specified range. If range is beyond
the dimensions of the worksheet, the boundary will be used as range.
}
\seealso{
\code{\link{get_all}}, \code{\link{get_row}}, \code{\link{get_rows}},
\code{\link{get_col}}, \code{\link{get_cols}}
}

