% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit.R
\name{rename_worksheet}
\alias{rename_worksheet}
\title{Rename a worksheet}
\usage{
rename_worksheet(ss, old_title, new_title)
}
\arguments{
\item{ss}{spreadsheet object}

\item{old_title}{worksheet's current title}

\item{new_title}{worksheets's new title}
}
\description{
Rename a worksheet
}

