% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-spreadsheets.R
\name{resize_ws}
\alias{resize_ws}
\title{Resize a worksheet}
\usage{
resize_ws(ss, ws = 1, row_extent = NULL, col_extent = NULL,
  verbose = TRUE)
}
\arguments{
\item{ss}{a registered Google spreadsheet}

\item{ws}{positive integer or character string specifying index or title,
  respectively, of the worksheet to consume}

\item{row_extent}{integer for new row extent}

\item{col_extent}{integer for new column extent}

\item{verbose}{logical; do you want informative message?}
}
\description{
Set the number of rows and columns of a worksheet. We use this function
internally during cell updates, if the data would exceed the current
worksheet extent. It is possible a user might want to use this directly?
}
\note{
Setting rows and columns to less than the current worksheet dimensions
  will delete contents without warning!
}
\examples{
\dontrun{
yo <- new_ss("yo")
yo <- edit_cells(yo, input = head(iris), header = TRUE, trim = TRUE)
get_via_csv(yo)
yo <- resize_ws(yo, ws = "Sheet1", row_extent = 5, col_extent = 4)
get_via_csv(yo)
yo <- resize_ws(yo, ws = 1, row_extent = 3, col_extent = 3)
get_via_csv(yo)
yo <- resize_ws(yo, row_extent = 2, col_extent = 2)
get_via_csv(yo)
delete_ss(yo)
}
}
\keyword{internal}

