% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spreadsheet}
\alias{spreadsheet}
\title{Spreadsheets}
\usage{
spreadsheet()
}
\value{
Object of class spreadsheet.

This function currently only works for public spreadsheets (visibility = TRUE and projection = FULL).
}
\description{
The function creates spreadsheet objects.
}

