% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{spreadsheet}
\alias{spreadsheet}
\title{The spreadsheet object}
\usage{
spreadsheet()
}
\description{
The spreadsheet object stores information about a spreadsheet. It includes
the fields:
}
\details{
\itemize{
\item \code{sheet_key} the key of the spreadsheet
\item \code{sheet_id} the id of the spreadsheet
\item \code{sheet_title} the title of the spreadsheet
\item \code{updated} the time of last update
\item \code{n_ws} the number of worksheets contained in the spreadsheet
\item \code{visibility} visibility of spreadsheet, determines whether
authorization is required for request
\item \code{ws} a data.frame about the worksheets contained in the
spreadsheet
}

TO DO: this documentation is neither here nor there. Either the object is
self-explanatory and this isn't really needed. Or this needs to get beefed
up.
}

