% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spreadsheets_info}
\alias{spreadsheets_info}
\title{Get information from spreadsheets feed}
\usage{
spreadsheets_info(client)
}
\arguments{
\item{client}{Client object returned by \code{\link{login}} or
\code{\link{authorize}}}
}
\description{
Get spreadsheets titles, keys, and date/time of last update.
}

