% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ssfeed_to_df}
\alias{ssfeed_to_df}
\title{Put information from spreadsheets feed into data frame}
\usage{
ssfeed_to_df()
}
\description{
Get spreadsheets' titles, date/time of last update and its unique key, and
organize into a data frame for easy post-processing.
}

