% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str.spreadsheet}
\alias{str.spreadsheet}
\title{Get the structure for a spreadsheet}
\usage{
\method{str}{spreadsheet}(object, ...)
}
\arguments{
\item{object}{spreadsheet object}

\item{...}{potential further arguments (required for Method/Generic reasons)}
}
\description{
Display the structure of a spreadsheet: the name of the spreadsheet, the
number of worksheets contained and the corresponding worksheet dimensions.
}

