% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str.spreadsheet}
\alias{str.spreadsheet}
\title{Get report for a spreadsheet}
\usage{
\method{str}{spreadsheet}(ss)
}
\arguments{
\item{ss}{spreadsheet object}
}
\value{
A list containing the name of the spreadsheet and the number of
worksheets it contains and worksheet dimensions.
}
\description{
Generate a report for a spreadsheet.
}

