% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str.worksheet}
\alias{str.worksheet}
\title{Get report for a worksheet}
\usage{
\method{str}{worksheet}(object, ...)
}
\arguments{
\item{object}{worksheet object}

\item{...}{potential further arguments (required for Method/Generic reasons)}
}
\value{
A list of 2.
}
\description{
Generate a report for a worksheet.
}

