% Generated by roxygen2 (4.0.2): do not edit by hand
\name{update_cell}
\alias{update_cell}
\title{Update a cell's value}
\usage{
update_cell(ws, pos, value)
}
\arguments{
\item{ws}{worksheet object}

\item{pos}{cell position in label (ex. "A1") or coordinate (ex. "R1C1")
format}

\item{value}{character string for new value}
}
\description{
Modify the contents of cells (those already with values) in worksheet.
To empty a cell, update it with an empty string.
}

