% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edit-data.R
\name{update_cell}
\alias{update_cell}
\title{Update a cell's value}
\usage{
update_cell(ws, pos, value)
}
\arguments{
\item{ws}{worksheet object}

\item{pos}{cell position in label (ex. "A1") or coordinate (ex. "R1C1")
format}

\item{value}{character string for new value}
}
\description{
Modify the contents of a cell in a worksheet. Formulas can be set on a cell
by starting with an = character. The calculated value will be the cell's
actual value. To empty a cell, update it with an empty string.
}
\note{
If the cell to update is beyond the extent of the worksheet, the
worksheet will be resized.
}
\examples{
\dontrun{
update_cell(worksheet, "A1", "First Cell")
update_cell(worksheet, "C1", "=A1+B1")
}
}

