% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{update_cells}
\alias{update_cells}
\title{Update multiple cells at once}
\usage{
update_cells(ws, range, dat, header = TRUE)
}
\arguments{
\item{ws}{worksheet object}

\item{range}{character string for range of cells (ie. "A1:A2", "A1:B6")}

\item{dat}{character vector or data frame of new values to update cells}

\item{header}{\code{logical} inidicating if data contains header row}
}
\description{
Modify a range of cells by specifiying the range and values to update to.
Values will be updated by row.
}

