% Generated by roxygen2 (4.0.2): do not edit by hand
\name{update_cells}
\alias{update_cells}
\title{Update multiple cells' values}
\usage{
update_cells(ws, range, new_values)
}
\arguments{
\item{ws}{worksheet object}

\item{range}{character string for range of cells (ie. "A1:A2", "A1:B6")}

\item{new_values}{vector of new values to update cells}
}
\description{
Modify a range of cells by specifiying the range and values to update with.
The range should not include empty cells because those cells will not be
captured in the response of the request.
}
\details{
Update by columns
}

