% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/inspect.R
\name{view_all}
\alias{view_all}
\title{View all your worksheets}
\usage{
view_all(ss, show_overlay = FALSE)
}
\arguments{
\item{ss}{spreadsheet object}

\item{show_overlay}{\code{logical} set to \code{TRUE} if want to also
display the overlay of all the worksheets}
}
\description{
Get a view of all the worksheets contained in the spreadsheet. This function
returns a list of two ggplot objects, first is a gallery of all the
worksheets and the second is an overlay of all the worksheets.
}

