% Generated by roxygen2 (4.0.2): do not edit by hand
\name{worksheet_dim}
\alias{worksheet_dim}
\title{Find the dimensions of a worksheet}
\usage{
worksheet_dim(ws, visibility = "private", auth = get_google_token())
}
\arguments{
\item{ws}{a worksheet object}

\item{visibility}{either "private" or "public"}

\item{auth}{Google token}
}
\description{
Get the rows and columns of a worksheet by making a request for cellfeed.
}

