% Generated by roxygen2 (4.0.2): do not edit by hand
\name{worksheet_dim}
\alias{worksheet_dim}
\title{Given client, worksheet object and spreadsheet id, find nrows and ncols of worksheet
need client because making a request for cellfeed}
\usage{
worksheet_dim(client = NULL, ws)
}
\arguments{
\item{client}{Client object}

\item{ws}{Worksheet object}
}
\description{
Given client, worksheet object and spreadsheet id, find nrows and ncols of worksheet
need client because making a request for cellfeed
}

