% Generated by roxygen2 (4.0.2): do not edit by hand
\name{worksheet_dim}
\alias{worksheet_dim}
\title{Find number of rows and columns of worksheet
Get the rows and columns of worksheet by making a request for cellfeed}
\usage{
worksheet_dim(ws, auth = get_google_token(), visibility = "private")
}
\arguments{
\item{ws}{Worksheet object}

\item{auth}{Google token}

\item{visibility}{set to \code{public} for public sheets}
}
\description{
Find number of rows and columns of worksheet
Get the rows and columns of worksheet by making a request for cellfeed
}

