% Generated by roxygen2 (4.0.2): do not edit by hand
\name{worksheet_dim}
\alias{worksheet_dim}
\title{Find number of rows and columns of worksheet}
\usage{
worksheet_dim(client = NULL, ws)
}
\arguments{
\item{client}{Client object}

\item{ws}{Worksheet object}
}
\description{
Get the rows and columns of worksheet by making a request for cellfeed, hence
client object is required if accessing private spreadsheet.
}

